@@ -1,16 +1,15 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
-// 	protoc        v3.12.4
+// 	protoc-gen-go v1.35.2
+// 	protoc        v5.28.2
 // source: topup.proto
 
 package pb
 
 import (
-	empty "github.com/golang/protobuf/ptypes/empty"
-	timestamp "github.com/golang/protobuf/ptypes/timestamp"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
+	emptypb "google.golang.org/protobuf/types/known/emptypb"
 	reflect "reflect"
 	sync "sync"
 )
@@ -22,39 +21,32 @@ const (
 	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
 )
 
-type Topup struct {
+type FindAllTopupRequest struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	TopupId     int32                `protobuf:"varint,1,opt,name=topup_id,json=topupId,proto3" json:"topup_id,omitempty"`
-	UserId      int32                `protobuf:"varint,2,opt,name=user_id,json=userId,proto3" json:"user_id,omitempty"`
-	TopupNo     string               `protobuf:"bytes,3,opt,name=topup_no,json=topupNo,proto3" json:"topup_no,omitempty"`
-	TopupAmount int32                `protobuf:"varint,4,opt,name=topup_amount,json=topupAmount,proto3" json:"topup_amount,omitempty"`
-	TopupMethod string               `protobuf:"bytes,5,opt,name=topup_method,json=topupMethod,proto3" json:"topup_method,omitempty"`
-	TopupTime   *timestamp.Timestamp `protobuf:"bytes,6,opt,name=topup_time,json=topupTime,proto3" json:"topup_time,omitempty"`
-	CreatedAt   *timestamp.Timestamp `protobuf:"bytes,7,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty"`
-	UpdatedAt   *timestamp.Timestamp `protobuf:"bytes,8,opt,name=updated_at,json=updatedAt,proto3" json:"updated_at,omitempty"`
+	Page     int32  `protobuf:"varint,1,opt,name=page,proto3" json:"page,omitempty"`
+	PageSize int32  `protobuf:"varint,2,opt,name=page_size,json=pageSize,proto3" json:"page_size,omitempty"`
+	Search   string `protobuf:"bytes,3,opt,name=search,proto3" json:"search,omitempty"`
 }
 
-func (x *Topup) Reset() {
-	*x = Topup{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_topup_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+func (x *FindAllTopupRequest) Reset() {
+	*x = FindAllTopupRequest{}
+	mi := &file_topup_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
-func (x *Topup) String() string {
+func (x *FindAllTopupRequest) String() string {
 	return protoimpl.X.MessageStringOf(x)
 }
 
-func (*Topup) ProtoMessage() {}
+func (*FindAllTopupRequest) ProtoMessage() {}
 
-func (x *Topup) ProtoReflect() protoreflect.Message {
+func (x *FindAllTopupRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_topup_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -64,138 +56,101 @@ func (x *Topup) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use Topup.ProtoReflect.Descriptor instead.
-func (*Topup) Descriptor() ([]byte, []int) {
+// Deprecated: Use FindAllTopupRequest.ProtoReflect.Descriptor instead.
+func (*FindAllTopupRequest) Descriptor() ([]byte, []int) {
 	return file_topup_proto_rawDescGZIP(), []int{0}
 }
 
-func (x *Topup) GetTopupId() int32 {
+func (x *FindAllTopupRequest) GetPage() int32 {
+	if x != nil {
+		return x.Page
+	}
+	return 0
+}
+
+func (x *FindAllTopupRequest) GetPageSize() int32 {
+	if x != nil {
+		return x.PageSize
+	}
+	return 0
+}
+
+func (x *FindAllTopupRequest) GetSearch() string {
+	if x != nil {
+		return x.Search
+	}
+	return ""
+}
+
+type FindByIdTopupRequest struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	TopupId int32 `protobuf:"varint,1,opt,name=topup_id,json=topupId,proto3" json:"topup_id,omitempty"`
+}
+
+func (x *FindByIdTopupRequest) Reset() {
+	*x = FindByIdTopupRequest{}
+	mi := &file_topup_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
+
+func (x *FindByIdTopupRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*FindByIdTopupRequest) ProtoMessage() {}
+
+func (x *FindByIdTopupRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_topup_proto_msgTypes[1]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use FindByIdTopupRequest.ProtoReflect.Descriptor instead.
+func (*FindByIdTopupRequest) Descriptor() ([]byte, []int) {
+	return file_topup_proto_rawDescGZIP(), []int{1}
+}
+
+func (x *FindByIdTopupRequest) GetTopupId() int32 {
 	if x != nil {
 		return x.TopupId
 	}
 	return 0
 }
 
-func (x *Topup) GetUserId() int32 {
-	if x != nil {
-		return x.UserId
-	}
-	return 0
-}
-
-func (x *Topup) GetTopupNo() string {
-	if x != nil {
-		return x.TopupNo
-	}
-	return ""
-}
-
-func (x *Topup) GetTopupAmount() int32 {
-	if x != nil {
-		return x.TopupAmount
-	}
-	return 0
-}
-
-func (x *Topup) GetTopupMethod() string {
-	if x != nil {
-		return x.TopupMethod
-	}
-	return ""
-}
-
-func (x *Topup) GetTopupTime() *timestamp.Timestamp {
-	if x != nil {
-		return x.TopupTime
-	}
-	return nil
-}
-
-func (x *Topup) GetCreatedAt() *timestamp.Timestamp {
-	if x != nil {
-		return x.CreatedAt
-	}
-	return nil
-}
-
-func (x *Topup) GetUpdatedAt() *timestamp.Timestamp {
-	if x != nil {
-		return x.UpdatedAt
-	}
-	return nil
-}
-
-type TopupRequest struct {
+type FindByCardNumberTopupRequest struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	Id int32 `protobuf:"varint,1,opt,name=id,proto3" json:"id,omitempty"`
+	CardNumber string `protobuf:"bytes,1,opt,name=card_number,json=cardNumber,proto3" json:"card_number,omitempty"`
 }
 
-func (x *TopupRequest) Reset() {
-	*x = TopupRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_topup_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
-}
-
-func (x *TopupRequest) String() string {
-	return protoimpl.X.MessageStringOf(x)
-}
-
-func (*TopupRequest) ProtoMessage() {}
-
-func (x *TopupRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_topup_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		if ms.LoadMessageInfo() == nil {
-			ms.StoreMessageInfo(mi)
-		}
-		return ms
-	}
-	return mi.MessageOf(x)
-}
-
-// Deprecated: Use TopupRequest.ProtoReflect.Descriptor instead.
-func (*TopupRequest) Descriptor() ([]byte, []int) {
-	return file_topup_proto_rawDescGZIP(), []int{1}
-}
-
-func (x *TopupRequest) GetId() int32 {
-	if x != nil {
-		return x.Id
-	}
-	return 0
-}
-
-type TopupsRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-}
-
-func (x *TopupsRequest) Reset() {
-	*x = TopupsRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_topup_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
-}
-
-func (x *TopupsRequest) String() string {
-	return protoimpl.X.MessageStringOf(x)
-}
-
-func (*TopupsRequest) ProtoMessage() {}
-
-func (x *TopupsRequest) ProtoReflect() protoreflect.Message {
+func (x *FindByCardNumberTopupRequest) Reset() {
+	*x = FindByCardNumberTopupRequest{}
 	mi := &file_topup_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
+
+func (x *FindByCardNumberTopupRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*FindByCardNumberTopupRequest) ProtoMessage() {}
+
+func (x *FindByCardNumberTopupRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_topup_proto_msgTypes[2]
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -205,103 +160,16 @@ func (x *TopupsRequest) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use TopupsRequest.ProtoReflect.Descriptor instead.
-func (*TopupsRequest) Descriptor() ([]byte, []int) {
+// Deprecated: Use FindByCardNumberTopupRequest.ProtoReflect.Descriptor instead.
+func (*FindByCardNumberTopupRequest) Descriptor() ([]byte, []int) {
 	return file_topup_proto_rawDescGZIP(), []int{2}
 }
 
-type TopupResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Topup *Topup `protobuf:"bytes,1,opt,name=topup,proto3" json:"topup,omitempty"`
-}
-
-func (x *TopupResponse) Reset() {
-	*x = TopupResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_topup_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
-}
-
-func (x *TopupResponse) String() string {
-	return protoimpl.X.MessageStringOf(x)
-}
-
-func (*TopupResponse) ProtoMessage() {}
-
-func (x *TopupResponse) ProtoReflect() protoreflect.Message {
-	mi := &file_topup_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		if ms.LoadMessageInfo() == nil {
-			ms.StoreMessageInfo(mi)
-		}
-		return ms
-	}
-	return mi.MessageOf(x)
-}
-
-// Deprecated: Use TopupResponse.ProtoReflect.Descriptor instead.
-func (*TopupResponse) Descriptor() ([]byte, []int) {
-	return file_topup_proto_rawDescGZIP(), []int{3}
-}
-
-func (x *TopupResponse) GetTopup() *Topup {
+func (x *FindByCardNumberTopupRequest) GetCardNumber() string {
 	if x != nil {
-		return x.Topup
+		return x.CardNumber
 	}
-	return nil
-}
-
-type TopupsResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Topups []*Topup `protobuf:"bytes,1,rep,name=topups,proto3" json:"topups,omitempty"`
-}
-
-func (x *TopupsResponse) Reset() {
-	*x = TopupsResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_topup_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
-}
-
-func (x *TopupsResponse) String() string {
-	return protoimpl.X.MessageStringOf(x)
-}
-
-func (*TopupsResponse) ProtoMessage() {}
-
-func (x *TopupsResponse) ProtoReflect() protoreflect.Message {
-	mi := &file_topup_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		if ms.LoadMessageInfo() == nil {
-			ms.StoreMessageInfo(mi)
-		}
-		return ms
-	}
-	return mi.MessageOf(x)
-}
-
-// Deprecated: Use TopupsResponse.ProtoReflect.Descriptor instead.
-func (*TopupsResponse) Descriptor() ([]byte, []int) {
-	return file_topup_proto_rawDescGZIP(), []int{4}
-}
-
-func (x *TopupsResponse) GetTopups() []*Topup {
-	if x != nil {
-		return x.Topups
-	}
-	return nil
+	return ""
 }
 
 type CreateTopupRequest struct {
@@ -309,7 +177,7 @@ type CreateTopupRequest struct {
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	UserId      int32  `protobuf:"varint,1,opt,name=user_id,json=userId,proto3" json:"user_id,omitempty"`
+	CardNumber  string `protobuf:"bytes,1,opt,name=card_number,json=cardNumber,proto3" json:"card_number,omitempty"`
 	TopupNo     string `protobuf:"bytes,2,opt,name=topup_no,json=topupNo,proto3" json:"topup_no,omitempty"`
 	TopupAmount int32  `protobuf:"varint,3,opt,name=topup_amount,json=topupAmount,proto3" json:"topup_amount,omitempty"`
 	TopupMethod string `protobuf:"bytes,4,opt,name=topup_method,json=topupMethod,proto3" json:"topup_method,omitempty"`
@@ -317,11 +185,9 @@ type CreateTopupRequest struct {
 
 func (x *CreateTopupRequest) Reset() {
 	*x = CreateTopupRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_topup_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_topup_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CreateTopupRequest) String() string {
@@ -331,8 +197,8 @@ func (x *CreateTopupRequest) String() string {
 func (*CreateTopupRequest) ProtoMessage() {}
 
 func (x *CreateTopupRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_topup_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	mi := &file_topup_proto_msgTypes[3]
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -344,14 +210,14 @@ func (x *CreateTopupRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use CreateTopupRequest.ProtoReflect.Descriptor instead.
 func (*CreateTopupRequest) Descriptor() ([]byte, []int) {
-	return file_topup_proto_rawDescGZIP(), []int{5}
+	return file_topup_proto_rawDescGZIP(), []int{3}
 }
 
-func (x *CreateTopupRequest) GetUserId() int32 {
+func (x *CreateTopupRequest) GetCardNumber() string {
 	if x != nil {
-		return x.UserId
+		return x.CardNumber
 	}
-	return 0
+	return ""
 }
 
 func (x *CreateTopupRequest) GetTopupNo() string {
@@ -380,19 +246,17 @@ type UpdateTopupRequest struct {
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	UserId      int32  `protobuf:"varint,1,opt,name=user_id,json=userId,proto3" json:"user_id,omitempty"`
-	TopupId     int32  `protobuf:"varint,2,opt,name=topup_id,json=topupId,proto3" json:"topup_id,omitempty"`
+	TopupId     int32  `protobuf:"varint,1,opt,name=topup_id,json=topupId,proto3" json:"topup_id,omitempty"`
+	CardNumber  string `protobuf:"bytes,2,opt,name=card_number,json=cardNumber,proto3" json:"card_number,omitempty"`
 	TopupAmount int32  `protobuf:"varint,3,opt,name=topup_amount,json=topupAmount,proto3" json:"topup_amount,omitempty"`
 	TopupMethod string `protobuf:"bytes,4,opt,name=topup_method,json=topupMethod,proto3" json:"topup_method,omitempty"`
 }
 
 func (x *UpdateTopupRequest) Reset() {
 	*x = UpdateTopupRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_topup_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_topup_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *UpdateTopupRequest) String() string {
@@ -402,8 +266,8 @@ func (x *UpdateTopupRequest) String() string {
 func (*UpdateTopupRequest) ProtoMessage() {}
 
 func (x *UpdateTopupRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_topup_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	mi := &file_topup_proto_msgTypes[4]
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -415,14 +279,7 @@ func (x *UpdateTopupRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use UpdateTopupRequest.ProtoReflect.Descriptor instead.
 func (*UpdateTopupRequest) Descriptor() ([]byte, []int) {
-	return file_topup_proto_rawDescGZIP(), []int{6}
-}
-
-func (x *UpdateTopupRequest) GetUserId() int32 {
-	if x != nil {
-		return x.UserId
-	}
-	return 0
+	return file_topup_proto_rawDescGZIP(), []int{4}
 }
 
 func (x *UpdateTopupRequest) GetTopupId() int32 {
@@ -430,6 +287,13 @@ func (x *UpdateTopupRequest) GetTopupId() int32 {
 		return x.TopupId
 	}
 	return 0
+}
+
+func (x *UpdateTopupRequest) GetCardNumber() string {
+	if x != nil {
+		return x.CardNumber
+	}
+	return ""
 }
 
 func (x *UpdateTopupRequest) GetTopupAmount() int32 {
@@ -446,32 +310,37 @@ func (x *UpdateTopupRequest) GetTopupMethod() string {
 	return ""
 }
 
-type DeleteTopupResponse struct {
+type TopupResponse struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	Success bool `protobuf:"varint,1,opt,name=success,proto3" json:"success,omitempty"`
+	Id          int32  `protobuf:"varint,1,opt,name=id,proto3" json:"id,omitempty"`
+	CardNumber  string `protobuf:"bytes,2,opt,name=card_number,json=cardNumber,proto3" json:"card_number,omitempty"`
+	TopupNo     string `protobuf:"bytes,3,opt,name=topup_no,json=topupNo,proto3" json:"topup_no,omitempty"`
+	TopupAmount int32  `protobuf:"varint,4,opt,name=topup_amount,json=topupAmount,proto3" json:"topup_amount,omitempty"`
+	TopupMethod string `protobuf:"bytes,5,opt,name=topup_method,json=topupMethod,proto3" json:"topup_method,omitempty"`
+	TopupTime   string `protobuf:"bytes,6,opt,name=topup_time,json=topupTime,proto3" json:"topup_time,omitempty"`
+	CreatedAt   string `protobuf:"bytes,7,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty"`
+	UpdatedAt   string `protobuf:"bytes,8,opt,name=updated_at,json=updatedAt,proto3" json:"updated_at,omitempty"`
 }
 
-func (x *DeleteTopupResponse) Reset() {
-	*x = DeleteTopupResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_topup_proto_msgTypes[7]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+func (x *TopupResponse) Reset() {
+	*x = TopupResponse{}
+	mi := &file_topup_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
-func (x *DeleteTopupResponse) String() string {
+func (x *TopupResponse) String() string {
 	return protoimpl.X.MessageStringOf(x)
 }
 
-func (*DeleteTopupResponse) ProtoMessage() {}
+func (*TopupResponse) ProtoMessage() {}
 
-func (x *DeleteTopupResponse) ProtoReflect() protoreflect.Message {
-	mi := &file_topup_proto_msgTypes[7]
-	if protoimpl.UnsafeEnabled && x != nil {
+func (x *TopupResponse) ProtoReflect() protoreflect.Message {
+	mi := &file_topup_proto_msgTypes[5]
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -481,107 +350,500 @@ func (x *DeleteTopupResponse) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use DeleteTopupResponse.ProtoReflect.Descriptor instead.
-func (*DeleteTopupResponse) Descriptor() ([]byte, []int) {
+// Deprecated: Use TopupResponse.ProtoReflect.Descriptor instead.
+func (*TopupResponse) Descriptor() ([]byte, []int) {
+	return file_topup_proto_rawDescGZIP(), []int{5}
+}
+
+func (x *TopupResponse) GetId() int32 {
+	if x != nil {
+		return x.Id
+	}
+	return 0
+}
+
+func (x *TopupResponse) GetCardNumber() string {
+	if x != nil {
+		return x.CardNumber
+	}
+	return ""
+}
+
+func (x *TopupResponse) GetTopupNo() string {
+	if x != nil {
+		return x.TopupNo
+	}
+	return ""
+}
+
+func (x *TopupResponse) GetTopupAmount() int32 {
+	if x != nil {
+		return x.TopupAmount
+	}
+	return 0
+}
+
+func (x *TopupResponse) GetTopupMethod() string {
+	if x != nil {
+		return x.TopupMethod
+	}
+	return ""
+}
+
+func (x *TopupResponse) GetTopupTime() string {
+	if x != nil {
+		return x.TopupTime
+	}
+	return ""
+}
+
+func (x *TopupResponse) GetCreatedAt() string {
+	if x != nil {
+		return x.CreatedAt
+	}
+	return ""
+}
+
+func (x *TopupResponse) GetUpdatedAt() string {
+	if x != nil {
+		return x.UpdatedAt
+	}
+	return ""
+}
+
+type ApiResponseTopup struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Status  string         `protobuf:"bytes,1,opt,name=status,proto3" json:"status,omitempty"`
+	Message string         `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
+	Data    *TopupResponse `protobuf:"bytes,3,opt,name=data,proto3" json:"data,omitempty"`
+}
+
+func (x *ApiResponseTopup) Reset() {
+	*x = ApiResponseTopup{}
+	mi := &file_topup_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
+
+func (x *ApiResponseTopup) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*ApiResponseTopup) ProtoMessage() {}
+
+func (x *ApiResponseTopup) ProtoReflect() protoreflect.Message {
+	mi := &file_topup_proto_msgTypes[6]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use ApiResponseTopup.ProtoReflect.Descriptor instead.
+func (*ApiResponseTopup) Descriptor() ([]byte, []int) {
+	return file_topup_proto_rawDescGZIP(), []int{6}
+}
+
+func (x *ApiResponseTopup) GetStatus() string {
+	if x != nil {
+		return x.Status
+	}
+	return ""
+}
+
+func (x *ApiResponseTopup) GetMessage() string {
+	if x != nil {
+		return x.Message
+	}
+	return ""
+}
+
+func (x *ApiResponseTopup) GetData() *TopupResponse {
+	if x != nil {
+		return x.Data
+	}
+	return nil
+}
+
+type ApiResponsesTopup struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Status  string           `protobuf:"bytes,1,opt,name=status,proto3" json:"status,omitempty"`
+	Message string           `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
+	Data    []*TopupResponse `protobuf:"bytes,3,rep,name=data,proto3" json:"data,omitempty"`
+}
+
+func (x *ApiResponsesTopup) Reset() {
+	*x = ApiResponsesTopup{}
+	mi := &file_topup_proto_msgTypes[7]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
+
+func (x *ApiResponsesTopup) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*ApiResponsesTopup) ProtoMessage() {}
+
+func (x *ApiResponsesTopup) ProtoReflect() protoreflect.Message {
+	mi := &file_topup_proto_msgTypes[7]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use ApiResponsesTopup.ProtoReflect.Descriptor instead.
+func (*ApiResponsesTopup) Descriptor() ([]byte, []int) {
 	return file_topup_proto_rawDescGZIP(), []int{7}
 }
 
-func (x *DeleteTopupResponse) GetSuccess() bool {
+func (x *ApiResponsesTopup) GetStatus() string {
 	if x != nil {
-		return x.Success
+		return x.Status
 	}
-	return false
+	return ""
+}
+
+func (x *ApiResponsesTopup) GetMessage() string {
+	if x != nil {
+		return x.Message
+	}
+	return ""
+}
+
+func (x *ApiResponsesTopup) GetData() []*TopupResponse {
+	if x != nil {
+		return x.Data
+	}
+	return nil
+}
+
+type ApiResponsePaginationTopup struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Status     string           `protobuf:"bytes,1,opt,name=status,proto3" json:"status,omitempty"`
+	Message    string           `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
+	Data       []*TopupResponse `protobuf:"bytes,3,rep,name=data,proto3" json:"data,omitempty"`
+	Pagination *PaginationMeta  `protobuf:"bytes,4,opt,name=pagination,proto3" json:"pagination,omitempty"`
+}
+
+func (x *ApiResponsePaginationTopup) Reset() {
+	*x = ApiResponsePaginationTopup{}
+	mi := &file_topup_proto_msgTypes[8]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
+
+func (x *ApiResponsePaginationTopup) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*ApiResponsePaginationTopup) ProtoMessage() {}
+
+func (x *ApiResponsePaginationTopup) ProtoReflect() protoreflect.Message {
+	mi := &file_topup_proto_msgTypes[8]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use ApiResponsePaginationTopup.ProtoReflect.Descriptor instead.
+func (*ApiResponsePaginationTopup) Descriptor() ([]byte, []int) {
+	return file_topup_proto_rawDescGZIP(), []int{8}
+}
+
+func (x *ApiResponsePaginationTopup) GetStatus() string {
+	if x != nil {
+		return x.Status
+	}
+	return ""
+}
+
+func (x *ApiResponsePaginationTopup) GetMessage() string {
+	if x != nil {
+		return x.Message
+	}
+	return ""
+}
+
+func (x *ApiResponsePaginationTopup) GetData() []*TopupResponse {
+	if x != nil {
+		return x.Data
+	}
+	return nil
+}
+
+func (x *ApiResponsePaginationTopup) GetPagination() *PaginationMeta {
+	if x != nil {
+		return x.Pagination
+	}
+	return nil
+}
+
+type ApiResponseTopupDelete struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Status  string `protobuf:"bytes,1,opt,name=status,proto3" json:"status,omitempty"`
+	Message string `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
+}
+
+func (x *ApiResponseTopupDelete) Reset() {
+	*x = ApiResponseTopupDelete{}
+	mi := &file_topup_proto_msgTypes[9]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
+
+func (x *ApiResponseTopupDelete) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*ApiResponseTopupDelete) ProtoMessage() {}
+
+func (x *ApiResponseTopupDelete) ProtoReflect() protoreflect.Message {
+	mi := &file_topup_proto_msgTypes[9]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use ApiResponseTopupDelete.ProtoReflect.Descriptor instead.
+func (*ApiResponseTopupDelete) Descriptor() ([]byte, []int) {
+	return file_topup_proto_rawDescGZIP(), []int{9}
+}
+
+func (x *ApiResponseTopupDelete) GetStatus() string {
+	if x != nil {
+		return x.Status
+	}
+	return ""
+}
+
+func (x *ApiResponseTopupDelete) GetMessage() string {
+	if x != nil {
+		return x.Message
+	}
+	return ""
+}
+
+type ApiResponseTopupRestore struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Status  string `protobuf:"bytes,1,opt,name=status,proto3" json:"status,omitempty"`
+	Message string `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
+}
+
+func (x *ApiResponseTopupRestore) Reset() {
+	*x = ApiResponseTopupRestore{}
+	mi := &file_topup_proto_msgTypes[10]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
+
+func (x *ApiResponseTopupRestore) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*ApiResponseTopupRestore) ProtoMessage() {}
+
+func (x *ApiResponseTopupRestore) ProtoReflect() protoreflect.Message {
+	mi := &file_topup_proto_msgTypes[10]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use ApiResponseTopupRestore.ProtoReflect.Descriptor instead.
+func (*ApiResponseTopupRestore) Descriptor() ([]byte, []int) {
+	return file_topup_proto_rawDescGZIP(), []int{10}
+}
+
+func (x *ApiResponseTopupRestore) GetStatus() string {
+	if x != nil {
+		return x.Status
+	}
+	return ""
+}
+
+func (x *ApiResponseTopupRestore) GetMessage() string {
+	if x != nil {
+		return x.Message
+	}
+	return ""
 }
 
 var File_topup_proto protoreflect.FileDescriptor
 
 var file_topup_proto_rawDesc = []byte{
 	0x0a, 0x0b, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x02, 0x70,
-	0x62, 0x1a, 0x1f, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62,
-	0x75, 0x66, 0x2f, 0x74, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x2e, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x1a, 0x1b, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x62, 0x75, 0x66, 0x2f, 0x65, 0x6d, 0x70, 0x74, 0x79, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22,
-	0xcd, 0x02, 0x0a, 0x05, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x19, 0x0a, 0x08, 0x74, 0x6f, 0x70,
-	0x75, 0x70, 0x5f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52, 0x07, 0x74, 0x6f, 0x70,
-	0x75, 0x70, 0x49, 0x64, 0x12, 0x17, 0x0a, 0x07, 0x75, 0x73, 0x65, 0x72, 0x5f, 0x69, 0x64, 0x18,
-	0x02, 0x20, 0x01, 0x28, 0x05, 0x52, 0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x64, 0x12, 0x19, 0x0a,
-	0x08, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x5f, 0x6e, 0x6f, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52,
-	0x07, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x4e, 0x6f, 0x12, 0x21, 0x0a, 0x0c, 0x74, 0x6f, 0x70, 0x75,
-	0x70, 0x5f, 0x61, 0x6d, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x04, 0x20, 0x01, 0x28, 0x05, 0x52, 0x0b,
-	0x74, 0x6f, 0x70, 0x75, 0x70, 0x41, 0x6d, 0x6f, 0x75, 0x6e, 0x74, 0x12, 0x21, 0x0a, 0x0c, 0x74,
-	0x6f, 0x70, 0x75, 0x70, 0x5f, 0x6d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x18, 0x05, 0x20, 0x01, 0x28,
-	0x09, 0x52, 0x0b, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x4d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x12, 0x39,
-	0x0a, 0x0a, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x5f, 0x74, 0x69, 0x6d, 0x65, 0x18, 0x06, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x09,
-	0x74, 0x6f, 0x70, 0x75, 0x70, 0x54, 0x69, 0x6d, 0x65, 0x12, 0x39, 0x0a, 0x0a, 0x63, 0x72, 0x65,
-	0x61, 0x74, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x07, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e,
-	0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e,
-	0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x09, 0x63, 0x72, 0x65, 0x61, 0x74,
-	0x65, 0x64, 0x41, 0x74, 0x12, 0x39, 0x0a, 0x0a, 0x75, 0x70, 0x64, 0x61, 0x74, 0x65, 0x64, 0x5f,
-	0x61, 0x74, 0x18, 0x08, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c,
-	0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73,
-	0x74, 0x61, 0x6d, 0x70, 0x52, 0x09, 0x75, 0x70, 0x64, 0x61, 0x74, 0x65, 0x64, 0x41, 0x74, 0x22,
-	0x1e, 0x0a, 0x0c, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12,
-	0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52, 0x02, 0x69, 0x64, 0x22,
-	0x0f, 0x0a, 0x0d, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
-	0x22, 0x30, 0x0a, 0x0d, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73,
-	0x65, 0x12, 0x1f, 0x0a, 0x05, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x09, 0x2e, 0x70, 0x62, 0x2e, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x05, 0x74, 0x6f, 0x70,
-	0x75, 0x70, 0x22, 0x33, 0x0a, 0x0e, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x73, 0x52, 0x65, 0x73, 0x70,
-	0x6f, 0x6e, 0x73, 0x65, 0x12, 0x21, 0x0a, 0x06, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x73, 0x18, 0x01,
-	0x20, 0x03, 0x28, 0x0b, 0x32, 0x09, 0x2e, 0x70, 0x62, 0x2e, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52,
-	0x06, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x73, 0x22, 0x8e, 0x01, 0x0a, 0x12, 0x43, 0x72, 0x65, 0x61,
-	0x74, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x17,
-	0x0a, 0x07, 0x75, 0x73, 0x65, 0x72, 0x5f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52,
-	0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x64, 0x12, 0x19, 0x0a, 0x08, 0x74, 0x6f, 0x70, 0x75, 0x70,
-	0x5f, 0x6e, 0x6f, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x74, 0x6f, 0x70, 0x75, 0x70,
-	0x4e, 0x6f, 0x12, 0x21, 0x0a, 0x0c, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x5f, 0x61, 0x6d, 0x6f, 0x75,
-	0x6e, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x05, 0x52, 0x0b, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x41,
-	0x6d, 0x6f, 0x75, 0x6e, 0x74, 0x12, 0x21, 0x0a, 0x0c, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x5f, 0x6d,
-	0x65, 0x74, 0x68, 0x6f, 0x64, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0b, 0x74, 0x6f, 0x70,
-	0x75, 0x70, 0x4d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x22, 0x8e, 0x01, 0x0a, 0x12, 0x55, 0x70, 0x64,
-	0x61, 0x74, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12,
-	0x17, 0x0a, 0x07, 0x75, 0x73, 0x65, 0x72, 0x5f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05,
-	0x52, 0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x64, 0x12, 0x19, 0x0a, 0x08, 0x74, 0x6f, 0x70, 0x75,
-	0x70, 0x5f, 0x69, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x05, 0x52, 0x07, 0x74, 0x6f, 0x70, 0x75,
-	0x70, 0x49, 0x64, 0x12, 0x21, 0x0a, 0x0c, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x5f, 0x61, 0x6d, 0x6f,
-	0x75, 0x6e, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x05, 0x52, 0x0b, 0x74, 0x6f, 0x70, 0x75, 0x70,
-	0x41, 0x6d, 0x6f, 0x75, 0x6e, 0x74, 0x12, 0x21, 0x0a, 0x0c, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x5f,
-	0x6d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0b, 0x74, 0x6f,
-	0x70, 0x75, 0x70, 0x4d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x22, 0x2f, 0x0a, 0x13, 0x44, 0x65, 0x6c,
-	0x65, 0x74, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65,
-	0x12, 0x18, 0x0a, 0x07, 0x73, 0x75, 0x63, 0x63, 0x65, 0x73, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28,
-	0x08, 0x52, 0x07, 0x73, 0x75, 0x63, 0x63, 0x65, 0x73, 0x73, 0x32, 0x98, 0x03, 0x0a, 0x0c, 0x54,
-	0x6f, 0x70, 0x75, 0x70, 0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x12, 0x37, 0x0a, 0x09, 0x47,
-	0x65, 0x74, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x73, 0x12, 0x16, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c,
-	0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79,
-	0x1a, 0x12, 0x2e, 0x70, 0x62, 0x2e, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x73, 0x52, 0x65, 0x73, 0x70,
-	0x6f, 0x6e, 0x73, 0x65, 0x12, 0x2f, 0x0a, 0x08, 0x47, 0x65, 0x74, 0x54, 0x6f, 0x70, 0x75, 0x70,
-	0x12, 0x10, 0x2e, 0x70, 0x62, 0x2e, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65,
-	0x73, 0x74, 0x1a, 0x11, 0x2e, 0x70, 0x62, 0x2e, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x73,
-	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x37, 0x0a, 0x0f, 0x47, 0x65, 0x74, 0x54, 0x6f, 0x70, 0x75,
-	0x70, 0x42, 0x79, 0x55, 0x73, 0x65, 0x72, 0x73, 0x12, 0x10, 0x2e, 0x70, 0x62, 0x2e, 0x54, 0x6f,
-	0x70, 0x75, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x12, 0x2e, 0x70, 0x62, 0x2e,
-	0x54, 0x6f, 0x70, 0x75, 0x70, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x37,
-	0x0a, 0x10, 0x47, 0x65, 0x74, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x42, 0x79, 0x55, 0x73, 0x65, 0x72,
-	0x49, 0x64, 0x12, 0x10, 0x2e, 0x70, 0x62, 0x2e, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x71,
-	0x75, 0x65, 0x73, 0x74, 0x1a, 0x11, 0x2e, 0x70, 0x62, 0x2e, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52,
-	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x38, 0x0a, 0x0b, 0x43, 0x72, 0x65, 0x61, 0x74,
-	0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x16, 0x2e, 0x70, 0x62, 0x2e, 0x43, 0x72, 0x65, 0x61,
-	0x74, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x11,
-	0x2e, 0x70, 0x62, 0x2e, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73,
-	0x65, 0x12, 0x38, 0x0a, 0x0b, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70,
-	0x12, 0x16, 0x2e, 0x70, 0x62, 0x2e, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x54, 0x6f, 0x70, 0x75,
-	0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x11, 0x2e, 0x70, 0x62, 0x2e, 0x54, 0x6f,
-	0x70, 0x75, 0x70, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x38, 0x0a, 0x0b, 0x44,
-	0x65, 0x6c, 0x65, 0x74, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x10, 0x2e, 0x70, 0x62, 0x2e,
-	0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x17, 0x2e, 0x70,
-	0x62, 0x2e, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x73,
-	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x42, 0x2b, 0x5a, 0x29, 0x4d, 0x61, 0x6d, 0x61, 0x6e, 0x67, 0x52,
-	0x75, 0x73, 0x74, 0x2f, 0x70, 0x61, 0x79, 0x6d, 0x65, 0x6e, 0x74, 0x67, 0x61, 0x74, 0x65, 0x77,
-	0x61, 0x79, 0x67, 0x72, 0x70, 0x63, 0x2f, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x2f,
-	0x70, 0x62, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x62, 0x1a, 0x1b, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62,
+	0x75, 0x66, 0x2f, 0x65, 0x6d, 0x70, 0x74, 0x79, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x09,
+	0x61, 0x70, 0x69, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x5e, 0x0a, 0x13, 0x46, 0x69, 0x6e,
+	0x64, 0x41, 0x6c, 0x6c, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
+	0x12, 0x12, 0x0a, 0x04, 0x70, 0x61, 0x67, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52, 0x04,
+	0x70, 0x61, 0x67, 0x65, 0x12, 0x1b, 0x0a, 0x09, 0x70, 0x61, 0x67, 0x65, 0x5f, 0x73, 0x69, 0x7a,
+	0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x05, 0x52, 0x08, 0x70, 0x61, 0x67, 0x65, 0x53, 0x69, 0x7a,
+	0x65, 0x12, 0x16, 0x0a, 0x06, 0x73, 0x65, 0x61, 0x72, 0x63, 0x68, 0x18, 0x03, 0x20, 0x01, 0x28,
+	0x09, 0x52, 0x06, 0x73, 0x65, 0x61, 0x72, 0x63, 0x68, 0x22, 0x31, 0x0a, 0x14, 0x46, 0x69, 0x6e,
+	0x64, 0x42, 0x79, 0x49, 0x64, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
+	0x74, 0x12, 0x19, 0x0a, 0x08, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x5f, 0x69, 0x64, 0x18, 0x01, 0x20,
+	0x01, 0x28, 0x05, 0x52, 0x07, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x49, 0x64, 0x22, 0x3f, 0x0a, 0x1c,
+	0x46, 0x69, 0x6e, 0x64, 0x42, 0x79, 0x43, 0x61, 0x72, 0x64, 0x4e, 0x75, 0x6d, 0x62, 0x65, 0x72,
+	0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1f, 0x0a, 0x0b,
+	0x63, 0x61, 0x72, 0x64, 0x5f, 0x6e, 0x75, 0x6d, 0x62, 0x65, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28,
+	0x09, 0x52, 0x0a, 0x63, 0x61, 0x72, 0x64, 0x4e, 0x75, 0x6d, 0x62, 0x65, 0x72, 0x22, 0x96, 0x01,
+	0x0a, 0x12, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x71,
+	0x75, 0x65, 0x73, 0x74, 0x12, 0x1f, 0x0a, 0x0b, 0x63, 0x61, 0x72, 0x64, 0x5f, 0x6e, 0x75, 0x6d,
+	0x62, 0x65, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x63, 0x61, 0x72, 0x64, 0x4e,
+	0x75, 0x6d, 0x62, 0x65, 0x72, 0x12, 0x19, 0x0a, 0x08, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x5f, 0x6e,
+	0x6f, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x4e, 0x6f,
+	0x12, 0x21, 0x0a, 0x0c, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x5f, 0x61, 0x6d, 0x6f, 0x75, 0x6e, 0x74,
+	0x18, 0x03, 0x20, 0x01, 0x28, 0x05, 0x52, 0x0b, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x41, 0x6d, 0x6f,
+	0x75, 0x6e, 0x74, 0x12, 0x21, 0x0a, 0x0c, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x5f, 0x6d, 0x65, 0x74,
+	0x68, 0x6f, 0x64, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0b, 0x74, 0x6f, 0x70, 0x75, 0x70,
+	0x4d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x22, 0x96, 0x01, 0x0a, 0x12, 0x55, 0x70, 0x64, 0x61, 0x74,
+	0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x19, 0x0a,
+	0x08, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x5f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52,
+	0x07, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x49, 0x64, 0x12, 0x1f, 0x0a, 0x0b, 0x63, 0x61, 0x72, 0x64,
+	0x5f, 0x6e, 0x75, 0x6d, 0x62, 0x65, 0x72, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x63,
+	0x61, 0x72, 0x64, 0x4e, 0x75, 0x6d, 0x62, 0x65, 0x72, 0x12, 0x21, 0x0a, 0x0c, 0x74, 0x6f, 0x70,
+	0x75, 0x70, 0x5f, 0x61, 0x6d, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x05, 0x52,
+	0x0b, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x41, 0x6d, 0x6f, 0x75, 0x6e, 0x74, 0x12, 0x21, 0x0a, 0x0c,
+	0x74, 0x6f, 0x70, 0x75, 0x70, 0x5f, 0x6d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x18, 0x04, 0x20, 0x01,
+	0x28, 0x09, 0x52, 0x0b, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x4d, 0x65, 0x74, 0x68, 0x6f, 0x64, 0x22,
+	0xfe, 0x01, 0x0a, 0x0d, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73,
+	0x65, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52, 0x02, 0x69,
+	0x64, 0x12, 0x1f, 0x0a, 0x0b, 0x63, 0x61, 0x72, 0x64, 0x5f, 0x6e, 0x75, 0x6d, 0x62, 0x65, 0x72,
+	0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x63, 0x61, 0x72, 0x64, 0x4e, 0x75, 0x6d, 0x62,
+	0x65, 0x72, 0x12, 0x19, 0x0a, 0x08, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x5f, 0x6e, 0x6f, 0x18, 0x03,
+	0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x4e, 0x6f, 0x12, 0x21, 0x0a,
+	0x0c, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x5f, 0x61, 0x6d, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x04, 0x20,
+	0x01, 0x28, 0x05, 0x52, 0x0b, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x41, 0x6d, 0x6f, 0x75, 0x6e, 0x74,
+	0x12, 0x21, 0x0a, 0x0c, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x5f, 0x6d, 0x65, 0x74, 0x68, 0x6f, 0x64,
+	0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0b, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x4d, 0x65, 0x74,
+	0x68, 0x6f, 0x64, 0x12, 0x1d, 0x0a, 0x0a, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x5f, 0x74, 0x69, 0x6d,
+	0x65, 0x18, 0x06, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x74, 0x6f, 0x70, 0x75, 0x70, 0x54, 0x69,
+	0x6d, 0x65, 0x12, 0x1d, 0x0a, 0x0a, 0x63, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x5f, 0x61, 0x74,
+	0x18, 0x07, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x63, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x41,
+	0x74, 0x12, 0x1d, 0x0a, 0x0a, 0x75, 0x70, 0x64, 0x61, 0x74, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18,
+	0x08, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x75, 0x70, 0x64, 0x61, 0x74, 0x65, 0x64, 0x41, 0x74,
+	0x22, 0x6b, 0x0a, 0x10, 0x41, 0x70, 0x69, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x54,
+	0x6f, 0x70, 0x75, 0x70, 0x12, 0x16, 0x0a, 0x06, 0x73, 0x74, 0x61, 0x74, 0x75, 0x73, 0x18, 0x01,
+	0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x73, 0x74, 0x61, 0x74, 0x75, 0x73, 0x12, 0x18, 0x0a, 0x07,
+	0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x6d,
+	0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x25, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x18, 0x03,
+	0x20, 0x01, 0x28, 0x0b, 0x32, 0x11, 0x2e, 0x70, 0x62, 0x2e, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52,
+	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x52, 0x04, 0x64, 0x61, 0x74, 0x61, 0x22, 0x6c, 0x0a,
+	0x11, 0x41, 0x70, 0x69, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x73, 0x54, 0x6f, 0x70,
+	0x75, 0x70, 0x12, 0x16, 0x0a, 0x06, 0x73, 0x74, 0x61, 0x74, 0x75, 0x73, 0x18, 0x01, 0x20, 0x01,
+	0x28, 0x09, 0x52, 0x06, 0x73, 0x74, 0x61, 0x74, 0x75, 0x73, 0x12, 0x18, 0x0a, 0x07, 0x6d, 0x65,
+	0x73, 0x73, 0x61, 0x67, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x6d, 0x65, 0x73,
+	0x73, 0x61, 0x67, 0x65, 0x12, 0x25, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x18, 0x03, 0x20, 0x03,
+	0x28, 0x0b, 0x32, 0x11, 0x2e, 0x70, 0x62, 0x2e, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x73,
+	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x52, 0x04, 0x64, 0x61, 0x74, 0x61, 0x22, 0xa9, 0x01, 0x0a, 0x1a,
+	0x41, 0x70, 0x69, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x50, 0x61, 0x67, 0x69, 0x6e,
+	0x61, 0x74, 0x69, 0x6f, 0x6e, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x16, 0x0a, 0x06, 0x73, 0x74,
+	0x61, 0x74, 0x75, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x73, 0x74, 0x61, 0x74,
+	0x75, 0x73, 0x12, 0x18, 0x0a, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x18, 0x02, 0x20,
+	0x01, 0x28, 0x09, 0x52, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x25, 0x0a, 0x04,
+	0x64, 0x61, 0x74, 0x61, 0x18, 0x03, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x11, 0x2e, 0x70, 0x62, 0x2e,
+	0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x52, 0x04, 0x64,
+	0x61, 0x74, 0x61, 0x12, 0x32, 0x0a, 0x0a, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f,
+	0x6e, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x12, 0x2e, 0x70, 0x62, 0x2e, 0x50, 0x61, 0x67,
+	0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x4d, 0x65, 0x74, 0x61, 0x52, 0x0a, 0x70, 0x61, 0x67,
+	0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x22, 0x4a, 0x0a, 0x16, 0x41, 0x70, 0x69, 0x52, 0x65,
+	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x44, 0x65, 0x6c, 0x65, 0x74,
+	0x65, 0x12, 0x16, 0x0a, 0x06, 0x73, 0x74, 0x61, 0x74, 0x75, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28,
+	0x09, 0x52, 0x06, 0x73, 0x74, 0x61, 0x74, 0x75, 0x73, 0x12, 0x18, 0x0a, 0x07, 0x6d, 0x65, 0x73,
+	0x73, 0x61, 0x67, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x6d, 0x65, 0x73, 0x73,
+	0x61, 0x67, 0x65, 0x22, 0x4b, 0x0a, 0x17, 0x41, 0x70, 0x69, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e,
+	0x73, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x73, 0x74, 0x6f, 0x72, 0x65, 0x12, 0x16,
+	0x0a, 0x06, 0x73, 0x74, 0x61, 0x74, 0x75, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06,
+	0x73, 0x74, 0x61, 0x74, 0x75, 0x73, 0x12, 0x18, 0x0a, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67,
+	0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
+	0x32, 0xb0, 0x05, 0x0a, 0x0c, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x53, 0x65, 0x72, 0x76, 0x69, 0x63,
+	0x65, 0x12, 0x47, 0x0a, 0x0c, 0x46, 0x69, 0x6e, 0x64, 0x41, 0x6c, 0x6c, 0x54, 0x6f, 0x70, 0x75,
+	0x70, 0x12, 0x17, 0x2e, 0x70, 0x62, 0x2e, 0x46, 0x69, 0x6e, 0x64, 0x41, 0x6c, 0x6c, 0x54, 0x6f,
+	0x70, 0x75, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1e, 0x2e, 0x70, 0x62, 0x2e,
+	0x41, 0x70, 0x69, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x50, 0x61, 0x67, 0x69, 0x6e,
+	0x61, 0x74, 0x69, 0x6f, 0x6e, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x3f, 0x0a, 0x0d, 0x46, 0x69,
+	0x6e, 0x64, 0x42, 0x79, 0x49, 0x64, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x18, 0x2e, 0x70, 0x62,
+	0x2e, 0x46, 0x69, 0x6e, 0x64, 0x42, 0x79, 0x49, 0x64, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65,
+	0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x14, 0x2e, 0x70, 0x62, 0x2e, 0x41, 0x70, 0x69, 0x52, 0x65,
+	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x4f, 0x0a, 0x15, 0x46,
+	0x69, 0x6e, 0x64, 0x42, 0x79, 0x43, 0x61, 0x72, 0x64, 0x4e, 0x75, 0x6d, 0x62, 0x65, 0x72, 0x54,
+	0x6f, 0x70, 0x75, 0x70, 0x12, 0x20, 0x2e, 0x70, 0x62, 0x2e, 0x46, 0x69, 0x6e, 0x64, 0x42, 0x79,
+	0x43, 0x61, 0x72, 0x64, 0x4e, 0x75, 0x6d, 0x62, 0x65, 0x72, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52,
+	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x14, 0x2e, 0x70, 0x62, 0x2e, 0x41, 0x70, 0x69, 0x52,
+	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x3d, 0x0a, 0x0c,
+	0x46, 0x69, 0x6e, 0x64, 0x42, 0x79, 0x41, 0x63, 0x74, 0x69, 0x76, 0x65, 0x12, 0x16, 0x2e, 0x67,
+	0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x45,
+	0x6d, 0x70, 0x74, 0x79, 0x1a, 0x15, 0x2e, 0x70, 0x62, 0x2e, 0x41, 0x70, 0x69, 0x52, 0x65, 0x73,
+	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x73, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x3e, 0x0a, 0x0d, 0x46,
+	0x69, 0x6e, 0x64, 0x42, 0x79, 0x54, 0x72, 0x61, 0x73, 0x68, 0x65, 0x64, 0x12, 0x16, 0x2e, 0x67,
+	0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x45,
+	0x6d, 0x70, 0x74, 0x79, 0x1a, 0x15, 0x2e, 0x70, 0x62, 0x2e, 0x41, 0x70, 0x69, 0x52, 0x65, 0x73,
+	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x73, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x3b, 0x0a, 0x0b, 0x43,
+	0x72, 0x65, 0x61, 0x74, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x16, 0x2e, 0x70, 0x62, 0x2e,
+	0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65,
+	0x73, 0x74, 0x1a, 0x14, 0x2e, 0x70, 0x62, 0x2e, 0x41, 0x70, 0x69, 0x52, 0x65, 0x73, 0x70, 0x6f,
+	0x6e, 0x73, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x3b, 0x0a, 0x0b, 0x55, 0x70, 0x64, 0x61,
+	0x74, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x16, 0x2e, 0x70, 0x62, 0x2e, 0x55, 0x70, 0x64,
+	0x61, 0x74, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a,
+	0x14, 0x2e, 0x70, 0x62, 0x2e, 0x41, 0x70, 0x69, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65,
+	0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x3e, 0x0a, 0x0c, 0x54, 0x72, 0x61, 0x73, 0x68, 0x65, 0x64,
+	0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x18, 0x2e, 0x70, 0x62, 0x2e, 0x46, 0x69, 0x6e, 0x64, 0x42,
+	0x79, 0x49, 0x64, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a,
+	0x14, 0x2e, 0x70, 0x62, 0x2e, 0x41, 0x70, 0x69, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65,
+	0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x3e, 0x0a, 0x0c, 0x52, 0x65, 0x73, 0x74, 0x6f, 0x72, 0x65,
+	0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x18, 0x2e, 0x70, 0x62, 0x2e, 0x46, 0x69, 0x6e, 0x64, 0x42,
+	0x79, 0x49, 0x64, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a,
+	0x14, 0x2e, 0x70, 0x62, 0x2e, 0x41, 0x70, 0x69, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65,
+	0x54, 0x6f, 0x70, 0x75, 0x70, 0x12, 0x4c, 0x0a, 0x14, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x54,
+	0x6f, 0x70, 0x75, 0x70, 0x50, 0x65, 0x72, 0x6d, 0x61, 0x6e, 0x65, 0x6e, 0x74, 0x12, 0x18, 0x2e,
+	0x70, 0x62, 0x2e, 0x46, 0x69, 0x6e, 0x64, 0x42, 0x79, 0x49, 0x64, 0x54, 0x6f, 0x70, 0x75, 0x70,
+	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1a, 0x2e, 0x70, 0x62, 0x2e, 0x41, 0x70, 0x69,
+	0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x54, 0x6f, 0x70, 0x75, 0x70, 0x44, 0x65, 0x6c,
+	0x65, 0x74, 0x65, 0x42, 0x2b, 0x5a, 0x29, 0x4d, 0x61, 0x6d, 0x61, 0x6e, 0x67, 0x52, 0x75, 0x73,
+	0x74, 0x2f, 0x70, 0x61, 0x79, 0x6d, 0x65, 0x6e, 0x74, 0x67, 0x61, 0x74, 0x65, 0x77, 0x61, 0x79,
+	0x67, 0x72, 0x70, 0x63, 0x2f, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x2f, 0x70, 0x62,
+	0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -596,44 +858,52 @@ func file_topup_proto_rawDescGZIP() []byte {
 	return file_topup_proto_rawDescData
 }
 
-var file_topup_proto_msgTypes = make([]protoimpl.MessageInfo, 8)
-var file_topup_proto_goTypes = []interface{}{
-	(*Topup)(nil),               // 0: pb.Topup
-	(*TopupRequest)(nil),        // 1: pb.TopupRequest
-	(*TopupsRequest)(nil),       // 2: pb.TopupsRequest
-	(*TopupResponse)(nil),       // 3: pb.TopupResponse
-	(*TopupsResponse)(nil),      // 4: pb.TopupsResponse
-	(*CreateTopupRequest)(nil),  // 5: pb.CreateTopupRequest
-	(*UpdateTopupRequest)(nil),  // 6: pb.UpdateTopupRequest
-	(*DeleteTopupResponse)(nil), // 7: pb.DeleteTopupResponse
-	(*timestamp.Timestamp)(nil), // 8: google.protobuf.Timestamp
-	(*empty.Empty)(nil),         // 9: google.protobuf.Empty
+var file_topup_proto_msgTypes = make([]protoimpl.MessageInfo, 11)
+var file_topup_proto_goTypes = []any{
+	(*FindAllTopupRequest)(nil),          // 0: pb.FindAllTopupRequest
+	(*FindByIdTopupRequest)(nil),         // 1: pb.FindByIdTopupRequest
+	(*FindByCardNumberTopupRequest)(nil), // 2: pb.FindByCardNumberTopupRequest
+	(*CreateTopupRequest)(nil),           // 3: pb.CreateTopupRequest
+	(*UpdateTopupRequest)(nil),           // 4: pb.UpdateTopupRequest
+	(*TopupResponse)(nil),                // 5: pb.TopupResponse
+	(*ApiResponseTopup)(nil),             // 6: pb.ApiResponseTopup
+	(*ApiResponsesTopup)(nil),            // 7: pb.ApiResponsesTopup
+	(*ApiResponsePaginationTopup)(nil),   // 8: pb.ApiResponsePaginationTopup
+	(*ApiResponseTopupDelete)(nil),       // 9: pb.ApiResponseTopupDelete
+	(*ApiResponseTopupRestore)(nil),      // 10: pb.ApiResponseTopupRestore
+	(*PaginationMeta)(nil),               // 11: pb.PaginationMeta
+	(*emptypb.Empty)(nil),                // 12: google.protobuf.Empty
 }
 var file_topup_proto_depIdxs = []int32{
-	8,  // 0: pb.Topup.topup_time:type_name -> google.protobuf.Timestamp
-	8,  // 1: pb.Topup.created_at:type_name -> google.protobuf.Timestamp
-	8,  // 2: pb.Topup.updated_at:type_name -> google.protobuf.Timestamp
-	0,  // 3: pb.TopupResponse.topup:type_name -> pb.Topup
-	0,  // 4: pb.TopupsResponse.topups:type_name -> pb.Topup
-	9,  // 5: pb.TopupService.GetTopups:input_type -> google.protobuf.Empty
-	1,  // 6: pb.TopupService.GetTopup:input_type -> pb.TopupRequest
-	1,  // 7: pb.TopupService.GetTopupByUsers:input_type -> pb.TopupRequest
-	1,  // 8: pb.TopupService.GetTopupByUserId:input_type -> pb.TopupRequest
-	5,  // 9: pb.TopupService.CreateTopup:input_type -> pb.CreateTopupRequest
-	6,  // 10: pb.TopupService.UpdateTopup:input_type -> pb.UpdateTopupRequest
-	1,  // 11: pb.TopupService.DeleteTopup:input_type -> pb.TopupRequest
-	4,  // 12: pb.TopupService.GetTopups:output_type -> pb.TopupsResponse
-	3,  // 13: pb.TopupService.GetTopup:output_type -> pb.TopupResponse
-	4,  // 14: pb.TopupService.GetTopupByUsers:output_type -> pb.TopupsResponse
-	3,  // 15: pb.TopupService.GetTopupByUserId:output_type -> pb.TopupResponse
-	3,  // 16: pb.TopupService.CreateTopup:output_type -> pb.TopupResponse
-	3,  // 17: pb.TopupService.UpdateTopup:output_type -> pb.TopupResponse
-	7,  // 18: pb.TopupService.DeleteTopup:output_type -> pb.DeleteTopupResponse
-	12, // [12:19] is the sub-list for method output_type
-	5,  // [5:12] is the sub-list for method input_type
-	5,  // [5:5] is the sub-list for extension type_name
-	5,  // [5:5] is the sub-list for extension extendee
-	0,  // [0:5] is the sub-list for field type_name
+	5,  // 0: pb.ApiResponseTopup.data:type_name -> pb.TopupResponse
+	5,  // 1: pb.ApiResponsesTopup.data:type_name -> pb.TopupResponse
+	5,  // 2: pb.ApiResponsePaginationTopup.data:type_name -> pb.TopupResponse
+	11, // 3: pb.ApiResponsePaginationTopup.pagination:type_name -> pb.PaginationMeta
+	0,  // 4: pb.TopupService.FindAllTopup:input_type -> pb.FindAllTopupRequest
+	1,  // 5: pb.TopupService.FindByIdTopup:input_type -> pb.FindByIdTopupRequest
+	2,  // 6: pb.TopupService.FindByCardNumberTopup:input_type -> pb.FindByCardNumberTopupRequest
+	12, // 7: pb.TopupService.FindByActive:input_type -> google.protobuf.Empty
+	12, // 8: pb.TopupService.FindByTrashed:input_type -> google.protobuf.Empty
+	3,  // 9: pb.TopupService.CreateTopup:input_type -> pb.CreateTopupRequest
+	4,  // 10: pb.TopupService.UpdateTopup:input_type -> pb.UpdateTopupRequest
+	1,  // 11: pb.TopupService.TrashedTopup:input_type -> pb.FindByIdTopupRequest
+	1,  // 12: pb.TopupService.RestoreTopup:input_type -> pb.FindByIdTopupRequest
+	1,  // 13: pb.TopupService.DeleteTopupPermanent:input_type -> pb.FindByIdTopupRequest
+	8,  // 14: pb.TopupService.FindAllTopup:output_type -> pb.ApiResponsePaginationTopup
+	6,  // 15: pb.TopupService.FindByIdTopup:output_type -> pb.ApiResponseTopup
+	6,  // 16: pb.TopupService.FindByCardNumberTopup:output_type -> pb.ApiResponseTopup
+	7,  // 17: pb.TopupService.FindByActive:output_type -> pb.ApiResponsesTopup
+	7,  // 18: pb.TopupService.FindByTrashed:output_type -> pb.ApiResponsesTopup
+	6,  // 19: pb.TopupService.CreateTopup:output_type -> pb.ApiResponseTopup
+	6,  // 20: pb.TopupService.UpdateTopup:output_type -> pb.ApiResponseTopup
+	6,  // 21: pb.TopupService.TrashedTopup:output_type -> pb.ApiResponseTopup
+	6,  // 22: pb.TopupService.RestoreTopup:output_type -> pb.ApiResponseTopup
+	9,  // 23: pb.TopupService.DeleteTopupPermanent:output_type -> pb.ApiResponseTopupDelete
+	14, // [14:24] is the sub-list for method output_type
+	4,  // [4:14] is the sub-list for method input_type
+	4,  // [4:4] is the sub-list for extension type_name
+	4,  // [4:4] is the sub-list for extension extendee
+	0,  // [0:4] is the sub-list for field type_name
 }
 
 func init() { file_topup_proto_init() }
@@ -641,111 +911,14 @@ func file_topup_proto_init() {
 	if File_topup_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_topup_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Topup); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_topup_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TopupRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_topup_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TopupsRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_topup_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TopupResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_topup_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TopupsResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_topup_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CreateTopupRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_topup_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*UpdateTopupRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_topup_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*DeleteTopupResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
+	file_api_proto_init()
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_topup_proto_rawDesc,
 			NumEnums:      0,
-			NumMessages:   8,
+			NumMessages:   11,
 			NumExtensions: 0,
 			NumServices:   1,
 		},
