@@ -2,13 +2,32 @@ package requests
 
 import "github.com/go-playground/validator/v10"
 
-type AuthLoginRequest struct {
-	Email    string `json:"email"`
-	Password string `json:"password"`
+type AuthRequest struct {
+	Email    string `json:"email" validate:"required,email"`
+	Password string `json:"password" validate:"required,min=6"`
 }
 
-func (r *AuthLoginRequest) Validate() error {
+type RegisterRequest struct {
+	FirstName       string `json:"firstname"`
+	LastName        string `json:"lastname"`
+	Email           string `json:"email" validate:"required,email"`
+	Password        string `json:"password" validate:"required,min=6"`
+	ConfirmPassword string `json:"confirm_password" validate:"required,min=6"`
+}
 
+func (r *AuthRequest) Validate() error {
+	validate := validator.New()
+
+	err := validate.Struct(r)
+
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (r *RegisterRequest) Validate() error {
 	validate := validator.New()
 
 	err := validate.Struct(r)
