@@ -21,326 +21,306 @@ func NewHandlerTransfer(client pb.TransferServiceClient, router *echo.Echo) *tra
 	}
 	routerTransfer := router.Group("/api/transfer")
 
-	routerTransfer.GET("/hello", transferHandler.handleHello)
-	routerTransfer.GET("/", transferHandler.handleGetTransfers)
-	routerTransfer.GET("/:id", transferHandler.handleGetTransfer)
-	routerTransfer.GET("/user-all/:id", transferHandler.handleGetTransferByUsers)
-	routerTransfer.GET("/user/:id", transferHandler.GetTransferByUserId)
-	routerTransfer.POST("/create", transferHandler.handleCreateTransfer)
-	routerTransfer.PUT("/update/:id", transferHandler.handleUpdateTransfer)
-	routerTransfer.DELETE("/:id", transferHandler.handleDeleteTransfer)
+	routerTransfer.GET("", transferHandler.FindAll)
+	routerTransfer.GET("/:id", transferHandler.FindById)
+	routerTransfer.GET("/transfer_from/:transfer_from", transferHandler.FindByTransferByTransferFrom)
+	routerTransfer.GET("/transfer_to/:transfer_to", transferHandler.FindByTransferByTransferTo)
+
+	routerTransfer.GET("/active", transferHandler.FindByActiveTransfer)
+	routerTransfer.GET("/trashed", transferHandler.FindByTrashedTransfer)
+
+	routerTransfer.POST("/create", transferHandler.CreateTransfer)
+	routerTransfer.POST("/update/:id", transferHandler.UpdateTransfer)
+	routerTransfer.POST("/trashed/:id", transferHandler.TrashTransfer)
+	routerTransfer.POST("/restore/:id", transferHandler.RestoreTransfer)
+	routerTransfer.DELETE("/:id", transferHandler.DeleteTransferPermanent)
 
 	return transferHandler
 
 }
 
-// @Summary Get a greeting message
-// @Description Get a greeting message
-// @Tags Transfer
-// @Accept json
-// @Produce json
-// @Success 200 {string} string "Hello"
-// @Router /transfer/hello [get]
-func (h *transferHandleApi) handleHello(c echo.Context) error {
-	return c.String(200, "Hello")
-}
+func (h *transferHandleApi) FindAll(c echo.Context) error {
+	page, err := strconv.Atoi(c.QueryParam("page"))
+	if err != nil || page <= 0 {
+		page = 1
+	}
 
-// @Summary Get all transfers
-// @Description Get all transfers
-// @Tags Transfer
-// @Accept json
-// @Produce json
-// @Success 200 {object} response.ResponseMessage "Success"
-// @Failure 400 {object} response.ResponseMessage "Failed to retrieve transfers: Error message"
-// @Failure 500 {object} response.ResponseMessage "Internal Server Error"
-// @Router /transfer [get]
-func (h *transferHandleApi) handleGetTransfers(c echo.Context) error {
-	res, err := h.client.GetTransfers(c.Request().Context(), &emptypb.Empty{})
+	pageSize, err := strconv.Atoi(c.QueryParam("page_size"))
+	if err != nil || pageSize <= 0 {
+		pageSize = 10
+	}
+
+	search := c.QueryParam("search")
+
+	ctx := c.Request().Context()
+
+	req := &pb.FindAllTransferRequest{
+		Page:     int32(page),
+		PageSize: int32(pageSize),
+		Search:   search,
+	}
+
+	res, err := h.client.FindAllTransfer(ctx, req)
 
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Failed to retrieve transfers: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to retrieve transfer data: ",
 		})
 	}
 
-	return c.JSON(http.StatusOK, response.ResponseMessage{
-		StatusCode: http.StatusOK,
-		Message:    "Success",
-		Data:       res,
-	})
+	return c.JSON(http.StatusOK, res)
 }
 
-// @Summary Get a transfer by ID
-// @Description Get a transfer by ID
-// @Tags Transfer
-// @Accept json
-// @Produce json
-// @Param id path int true "Transfer ID"
-// @Success 200 {object} response.ResponseMessage "Success"
-// @Failure 400 {object} response.ResponseMessage "Bad Request: Error message"
-// @Failure 500 {object} response.ResponseMessage "Internal Server Error"
-// @Router /transfer/{id} [get]
-func (h *transferHandleApi) handleGetTransfer(c echo.Context) error {
+func (h *transferHandleApi) FindById(c echo.Context) error {
 	id := c.Param("id")
 
 	idInt, err := strconv.Atoi(id)
 
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Bad Request: Invalid ID: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusBadRequest, response.ErrorResponse{
+			Status:  "error",
+			Message: "Bad Request: Invalid ID",
 		})
+
 	}
 
-	res, err := h.client.GetTransfer(c.Request().Context(), &pb.TransferRequest{
-		Id: int32(idInt),
+	ctx := c.Request().Context()
+
+	res, err := h.client.FindByIdTransfer(ctx, &pb.FindByIdTransferRequest{
+		TransferId: int32(idInt),
 	})
 
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Failed to retrieve transfer: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to retrieve transfer data: ",
 		})
 	}
 
-	return c.JSON(http.StatusOK, response.ResponseMessage{
-		StatusCode: http.StatusOK,
-		Message:    "Success",
-		Data:       res,
-	})
+	return c.JSON(http.StatusOK, res)
 }
 
-// @Summary Get transfers by user ID
-// @Description Get transfers by user ID
-// @Tags Transfer
-// @Accept json
-// @Produce json
-// @Param id path int true "User ID"
-// @Success 200 {object} response.ResponseMessage "Success"
-// @Failure 400 {object} response.ResponseMessage "Bad Request: Error message"
-// @Failure 500 {object} response.ResponseMessage "Internal Server Error"
-// @Router /transfer/user-all/{id} [get]
-func (h *transferHandleApi) handleGetTransferByUsers(c echo.Context) error {
-	id := c.Param("id")
+func (h *transferHandleApi) FindByTransferByTransferFrom(c echo.Context) error {
+	transfer_from := c.Param("transfer_from")
 
-	idInt, err := strconv.Atoi(id)
+	ctx := c.Request().Context()
 
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Bad Request: Invalid ID: " + err.Error(),
-			Data:       nil,
-		})
-	}
-
-	res, err := h.client.GetTransferByUsers(c.Request().Context(), &pb.TransferRequest{
-		Id: int32(idInt),
+	res, err := h.client.FindTransferByTransferFrom(ctx, &pb.FindTransferByTransferFromRequest{
+		TransferFrom: transfer_from,
 	})
 
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Failed to retrieve transfers by user: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to retrieve transfer data: ",
 		})
 	}
 
-	return c.JSON(http.StatusOK, response.ResponseMessage{
-		StatusCode: http.StatusOK,
-		Message:    "Success",
-		Data:       res,
-	})
+	return c.JSON(http.StatusOK, res)
 }
 
-// @Summary Get a transfer by user ID
-// @Description Get a transfer by user ID
-// @Tags Transfer
-// @Accept json
-// @Produce json
-// @Param id path int true "User ID"
-// @Success 200 {object} response.ResponseMessage "Success"
-// @Failure 400 {object} response.ResponseMessage "Bad Request: Error message"
-// @Failure 500 {object} response.ResponseMessage "Internal Server Error"
-// @Router /transfer/user/{id} [get]
-func (h *transferHandleApi) GetTransferByUserId(c echo.Context) error {
-	id := c.Param("id")
+func (h *transferHandleApi) FindByTransferByTransferTo(c echo.Context) error {
+	transfer_to := c.Param("transfer_to")
 
-	idInt, err := strconv.Atoi(id)
+	ctx := c.Request().Context()
 
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Bad Request: Invalid ID: " + err.Error(),
-			Data:       nil,
-		})
-	}
-
-	res, err := h.client.GetTransferByUserId(c.Request().Context(), &pb.TransferRequest{
-		Id: int32(idInt),
+	res, err := h.client.FindTransferByTransferTo(ctx, &pb.FindTransferByTransferToRequest{
+		TransferTo: transfer_to,
 	})
 
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Failed to retrieve transfer by user: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to retrieve transfer data: ",
 		})
 	}
 
-	return c.JSON(http.StatusOK, response.ResponseMessage{
-		StatusCode: http.StatusOK,
-		Message:    "Success",
-		Data:       res,
-	})
+	return c.JSON(http.StatusOK, res)
 }
 
-// @Summary Create a new transfer
-// @Description Create a new transfer
-// @Tags Transfer
-// @Accept json
-// @Produce json
-// @Param body body requests.CreateTransferRequest true "Transfer details"
-// @Success 200 {object} response.ResponseMessage "Success"
-// @Failure 400 {object} response.ResponseMessage "Bad Request: Error message"
-// @Failure 500 {object} response.ResponseMessage "Internal Server Error"
-// @Router /transfer/create [post]
-func (h *transferHandleApi) handleCreateTransfer(c echo.Context) error {
+func (h *transferHandleApi) FindByActiveTransfer(c echo.Context) error {
+	ctx := c.Request().Context()
+
+	res, err := h.client.FindByActiveTransfer(ctx, &emptypb.Empty{})
+
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to retrieve transfer data: ",
+		})
+	}
+
+	return c.JSON(http.StatusOK, res)
+}
+
+func (h *transferHandleApi) FindByTrashedTransfer(c echo.Context) error {
+	ctx := c.Request().Context()
+
+	res, err := h.client.FindByTrashedTransfer(ctx, &emptypb.Empty{})
+
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to retrieve transfer data: ",
+		})
+	}
+
+	return c.JSON(http.StatusOK, res)
+}
+
+func (h *transferHandleApi) CreateTransfer(c echo.Context) error {
 	var body requests.CreateTransferRequest
 
 	if err := c.Bind(&body); err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Bad Request: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusBadRequest, response.ErrorResponse{
+			Status:  "error",
+			Message: "Invalid request body",
 		})
 	}
 
 	if err := body.Validate(); err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Bad Request Validate: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusBadRequest, response.ErrorResponse{
+			Status:  "error",
+			Message: "Validation Error: " + err.Error(),
 		})
 	}
 
-	data := &pb.CreateTransferRequest{
-		TransferFrom:   int32(body.TransferFrom),
-		TransferTo:     int32(body.TransferTo),
-		TransferAmount: int32(body.TransferAmount),
-	}
+	ctx := c.Request().Context()
 
-	res, err := h.client.CreateTransfer(c.Request().Context(), data)
+	res, err := h.client.CreateTransfer(ctx, &pb.CreateTransferRequest{
+		TransferFrom:   body.TransferFrom,
+		TransferTo:     body.TransferTo,
+		TransferAmount: int32(body.TransferAmount),
+	})
 
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Failed to create transfer: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to create transfer: ",
 		})
 	}
 
-	return c.JSON(http.StatusOK, response.ResponseMessage{
-		StatusCode: http.StatusOK,
-		Message:    "Success",
-		Data:       res,
-	})
+	return c.JSON(http.StatusOK, res)
 }
 
-// @Summary Update a transfer
-// @Description Update a transfer
-// @Tags Transfer
-// @Accept json
-// @Produce json
-// @Param body body requests.UpdateTransferRequest true "Transfer details"
-// @Success 200 {object} response.ResponseMessage "Success"
-// @Failure 400 {object} response.ResponseMessage "Bad Request: Error message"
-// @Failure 500 {object} response.ResponseMessage "Internal Server Error"
-// @Router /transfer/update/{id} [put]
-func (h *transferHandleApi) handleUpdateTransfer(c echo.Context) error {
+func (h *transferHandleApi) UpdateTransfer(c echo.Context) error {
 	var body requests.UpdateTransferRequest
 
 	if err := c.Bind(&body); err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Bad Request: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusBadRequest, response.ErrorResponse{
+			Status:  "error",
+			Message: "Invalid request body",
 		})
 	}
 
 	if err := body.Validate(); err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Bad Request Validate: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusBadRequest, response.ErrorResponse{
+			Status:  "error",
+			Message: "Validation Error: " + err.Error(),
 		})
 	}
 
-	data := &pb.UpdateTransferRequest{
-		Id:             int32(body.TransferID),
-		TransferFrom:   int32(body.TransferFrom),
-		TransferTo:     int32(body.TransferTo),
-		TransferAmount: int32(body.TransferAmount),
-	}
+	ctx := c.Request().Context()
 
-	res, err := h.client.UpdateTransfer(c.Request().Context(), data)
+	res, err := h.client.UpdateTransfer(ctx, &pb.UpdateTransferRequest{
+		TransferId:     int32(body.TransferID),
+		TransferFrom:   body.TransferFrom,
+		TransferTo:     body.TransferTo,
+		TransferAmount: int32(body.TransferAmount),
+	})
 
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Failed to update transfer: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to update transfer: ",
 		})
 	}
 
-	return c.JSON(http.StatusOK, response.ResponseMessage{
-		StatusCode: http.StatusOK,
-		Message:    "Success",
-		Data:       res,
-	})
+	return c.JSON(http.StatusOK, res)
 }
 
-// @Summary Delete a transfer by ID
-// @Description Delete a transfer by ID
-// @Tags Transfer
-// @Accept json
-// @Produce json
-// @Param id path int true "Transfer ID"
-// @Success 200 {object} response.ResponseMessage "Success"
-// @Failure 400 {object} response.ResponseMessage "Bad Request: Error message"
-// @Failure 500 {object} response.ResponseMessage "Internal Server Error"
-// @Router /transfer/delete/{id} [delete]
-func (h *transferHandleApi) handleDeleteTransfer(c echo.Context) error {
+func (h *transferHandleApi) TrashTransfer(c echo.Context) error {
 	id := c.Param("id")
 
 	idInt, err := strconv.Atoi(id)
 
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Bad Request: Invalid ID: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusBadRequest, response.ErrorResponse{
+			Status:  "error",
+			Message: "Bad Request: Invalid ID",
 		})
+
 	}
 
-	res, err := h.client.DeleteTransfer(c.Request().Context(), &pb.TransferRequest{
-		Id: int32(idInt),
+	ctx := c.Request().Context()
+
+	res, err := h.client.TrashedTransfer(ctx, &pb.FindByIdTransferRequest{
+		TransferId: int32(idInt),
 	})
 
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Failed to delete transfer: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to trash transfer:",
 		})
 	}
 
-	return c.JSON(http.StatusOK, response.ResponseMessage{
-		StatusCode: http.StatusOK,
-		Message:    "Success",
-		Data:       res,
+	return c.JSON(http.StatusOK, res)
+}
+
+func (h *transferHandleApi) RestoreTransfer(c echo.Context) error {
+	id := c.Param("id")
+
+	idInt, err := strconv.Atoi(id)
+
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, response.ErrorResponse{
+			Status:  "error",
+			Message: "Bad Request: Invalid ID",
+		})
+	}
+
+	ctx := c.Request().Context()
+
+	res, err := h.client.RestoreTransfer(ctx, &pb.FindByIdTransferRequest{
+		TransferId: int32(idInt),
 	})
+
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to restore transfer:",
+		})
+	}
+
+	return c.JSON(http.StatusOK, res)
+}
+
+func (h *transferHandleApi) DeleteTransferPermanent(c echo.Context) error {
+	id := c.Param("id")
+
+	idInt, err := strconv.Atoi(id)
+
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, response.ErrorResponse{
+			Status:  "error",
+			Message: "Bad Request: Invalid ID",
+		})
+	}
+
+	ctx := c.Request().Context()
+
+	res, err := h.client.DeleteTransferPermanent(ctx, &pb.FindByIdTransferRequest{
+		TransferId: int32(idInt),
+	})
+
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to delete transfer:",
+		})
+	}
+
+	return c.JSON(http.StatusOK, res)
 }
