@@ -8,7 +8,6 @@ import (
 	"strconv"
 
 	"github.com/labstack/echo/v4"
-	"google.golang.org/protobuf/types/known/emptypb"
 )
 
 type userHandleApi struct {
@@ -21,124 +20,100 @@ func NewHandlerUser(client pb.UserServiceClient, router *echo.Echo) *userHandleA
 	}
 	routerUser := router.Group("/api/user")
 
-	routerUser.GET("/hello", userHandler.handleHello)
-	routerUser.GET("/", userHandler.handleGetUsers)
-	routerUser.GET("/:id", userHandler.handleGetUser)
-	routerUser.POST("/create", userHandler.handleCreateUser)
-	routerUser.PUT("/update/:id", userHandler.handleUpdateUser)
-	routerUser.DELETE("/delete/:id", userHandler.handleDeleteUser)
+	routerUser.GET("/", userHandler.FindAllUser)
+	routerUser.GET("/:id", userHandler.FindById)
+
+	routerUser.POST("/create", userHandler.Create)
+	routerUser.POST("/update/:id", userHandler.Update)
+	
+	routerUser.POST("/trashed/:id", userHandler.TrashedUser)
+	routerUser.POST("/restore/:id", userHandler.RestoreUser)
+	routerUser.DELETE("/:id", userHandler.DeleteUserPermanent)
 
 	return userHandler
 }
 
-// handleHello godoc
-// @Summary Menampilkan pesan hello
-// @Description Menampilkan pesan hello
-// @Tags User
-// @Accept  json
-// @Produce  json
-// @Success 200 {string} string "Hello"
-// @Router /user/hello [get]
-func (h *userHandleApi) handleHello(c echo.Context) error {
-	return c.String(200, "Hello")
-}
+func (h *userHandleApi) FindAllUser(c echo.Context) error {
+	page, err := strconv.Atoi(c.QueryParam("page"))
+	if err != nil || page <= 0 {
+		page = 1
+	}
 
-// @Summary Get all users
-// @Description Get all users
-// @Tags User
-// @Accept json
-// @Produce json
-// @Success 200 {object} response.ResponseMessage "Success"
-// @Failure 400 {object} response.ResponseMessage "Bad Request: Error message"
-// @Router /user/ [get]
-func (h *userHandleApi) handleGetUsers(c echo.Context) error {
-	res, err := h.client.GetUsers(c.Request().Context(), &emptypb.Empty{})
+	pageSize, err := strconv.Atoi(c.QueryParam("page_size"))
+	if err != nil || pageSize <= 0 {
+		pageSize = 10
+	}
+
+	search := c.QueryParam("search")
+
+	ctx := c.Request().Context()
+
+	req := &pb.FindAllUserRequest{
+		Page:     int32(page),
+		PageSize: int32(pageSize),
+		Search:   search,
+	}
+
+	res, err := h.client.FindAll(ctx, req)
 
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Failed to get users: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to retrieve user data: ",
 		})
 	}
 
-	return c.JSON(http.StatusOK, response.ResponseMessage{
-		StatusCode: http.StatusOK,
-		Message:    "Success",
-		Data:       res,
-	})
+	return c.JSON(http.StatusOK, res)
 }
 
-// @Summary Get a user by ID
-// @Description Get a user by ID
-// @Tags User
-// @Accept json
-// @Produce json
-// @Param id path int true "User ID"
-// @Success 200 {object} response.ResponseMessage "Success"
-// @Failure 400 {object} response.ResponseMessage "Bad Request: Error message"
-// @Router /user/{id} [get]
-func (h *userHandleApi) handleGetUser(c echo.Context) error {
-	id := c.Param("id")
-
-	idInt, err := strconv.Atoi(id)
+func (h *userHandleApi) FindById(c echo.Context) error {
+	id, err := strconv.Atoi(c.Param("id"))
 
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Bad Request: Invalid ID",
-			Data:       nil,
+		return c.JSON(http.StatusBadRequest, response.ErrorResponse{
+			Status:  "error",
+			Message: "Invalid user ID",
 		})
 	}
 
-	res, err := h.client.GetUser(c.Request().Context(), &pb.UserRequest{
-		Id: int32(idInt),
-	})
+	ctx := c.Request().Context()
+
+	req := &pb.FindByIdUserRequest{
+		Id: int32(id),
+	}
+
+	user, err := h.client.FindById(ctx, req)
 
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Failed to get user: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to retrieve user data: ",
 		})
 	}
 
-	return c.JSON(http.StatusOK, response.ResponseMessage{
-		StatusCode: http.StatusOK,
-		Message:    "Success",
-		Data:       res,
-	})
+	return c.JSON(http.StatusOK, user)
 }
 
-// @Summary Create a new user
-// @Description Create a new user
-// @Tags User
-// @Accept json
-// @Produce json
-// @Param body body requests.CreateUserRequest true "User details"
-// @Success 200 {object} response.ResponseMessage "Success"
-// @Failure 400 {object} response.ResponseMessage "Bad Request: Error message"
-// @Router /user/create [post]
-func (h *userHandleApi) handleCreateUser(c echo.Context) error {
+func (h *userHandleApi) Create(c echo.Context) error {
 	var body requests.CreateUserRequest
 
 	if err := c.Bind(&body); err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Bad Request: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusBadRequest, response.ErrorResponse{
+			Status:  "error",
+			Message: "Invalid request body",
 		})
 	}
 
 	if err := body.Validate(); err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Bad Request Validate: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusBadRequest, response.ErrorResponse{
+			Status:  "error",
+			Message: "Validation Error: " + err.Error(),
 		})
 	}
 
-	data := &pb.CreateUserRequest{
+	ctx := c.Request().Context()
+
+	req := &pb.CreateUserRequest{
 		Firstname:       body.FirstName,
 		Lastname:        body.LastName,
 		Email:           body.Email,
@@ -146,53 +121,39 @@ func (h *userHandleApi) handleCreateUser(c echo.Context) error {
 		ConfirmPassword: body.ConfirmPassword,
 	}
 
-	res, err := h.client.CreateUser(c.Request().Context(), data)
+	res, err := h.client.Create(ctx, req)
 
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Failed to create user: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to create user: ",
 		})
 	}
 
-	return c.JSON(http.StatusOK, response.ResponseMessage{
-		StatusCode: http.StatusOK,
-		Message:    "Success",
-		Data:       res,
-	})
+	return c.JSON(http.StatusOK, res)
 }
 
-// @Summary Update a user
-// @Description Update a user
-// @Tags User
-// @Accept json
-// @Produce json
-// @Param body body requests.UpdateUserRequest true "User details"
-// @Success 200 {object} response.ResponseMessage "Success"
-// @Failure 400 {object} response.ResponseMessage "Bad Request: Error message"
-// @Router /user/update/{id} [put]
-func (h *userHandleApi) handleUpdateUser(c echo.Context) error {
+func (h *userHandleApi) Update(c echo.Context) error {
 	var body requests.UpdateUserRequest
 
 	if err := c.Bind(&body); err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Bad Request: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusBadRequest, response.ErrorResponse{
+			Status:  "error",
+			Message: "Invalid request body",
 		})
 	}
 
 	if err := body.Validate(); err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Bad Request Validate: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusBadRequest, response.ErrorResponse{
+			Status:  "error",
+			Message: "Validation Error: " + err.Error(),
 		})
 	}
 
-	data := &pb.UpdateUserRequest{
-		Id:              int32(body.ID),
+	ctx := c.Request().Context()
+
+	req := &pb.UpdateUserRequest{
+		Id:              int32(body.UserID),
 		Firstname:       body.FirstName,
 		Lastname:        body.LastName,
 		Email:           body.Email,
@@ -200,60 +161,98 @@ func (h *userHandleApi) handleUpdateUser(c echo.Context) error {
 		ConfirmPassword: body.ConfirmPassword,
 	}
 
-	res, err := h.client.UpdateUser(c.Request().Context(), data)
+	res, err := h.client.Update(ctx, req)
 
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Failed to update user: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to update user: ",
 		})
 	}
 
-	return c.JSON(http.StatusOK, response.ResponseMessage{
-		StatusCode: http.StatusOK,
-		Message:    "Success",
-		Data:       res,
-	})
+	return c.JSON(http.StatusOK, res)
 }
 
-// @Summary Delete a user by ID
-// @Description Delete a user by ID
-// @Tags User
-// @Accept json
-// @Produce json
-// @Param id path int true "User ID"
-// @Success 200 {object} response.ResponseMessage "Success"
-// @Failure 400 {object} response.ResponseMessage "Bad Request: Error message"
-// @Router /user/delete/{id} [delete]
-func (h *userHandleApi) handleDeleteUser(c echo.Context) error {
-	id := c.Param("id")
-
-	idInt, err := strconv.Atoi(id)
+func (h *userHandleApi) TrashedUser(c echo.Context) error {
+	id, err := strconv.Atoi(c.Param("id"))
 
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Bad Request: Invalid ID: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusBadRequest, response.ErrorResponse{
+			Status:  "error",
+			Message: "Invalid user ID",
 		})
 	}
 
-	res, err := h.client.DeleteUser(c.Request().Context(), &pb.UserRequest{
-		Id: int32(idInt),
-	})
+	ctx := c.Request().Context()
+
+	req := &pb.FindByIdUserRequest{
+		Id: int32(id),
+	}
+
+	user, err := h.client.TrashedUser(ctx, req)
 
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, response.ResponseMessage{
-			StatusCode: http.StatusBadRequest,
-			Message:    "Failed to delete user: " + err.Error(),
-			Data:       nil,
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to trashed user: ",
 		})
 	}
 
-	return c.JSON(http.StatusOK, response.ResponseMessage{
-		StatusCode: http.StatusOK,
-		Message:    "Success",
-		Data:       res,
-	})
+	return c.JSON(http.StatusOK, user)
+}
+
+func (h *userHandleApi) RestoreUser(c echo.Context) error {
+	id, err := strconv.Atoi(c.Param("id"))
+
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, response.ErrorResponse{
+			Status:  "error",
+			Message: "Invalid user ID",
+		})
+	}
+
+	ctx := c.Request().Context()
+
+	req := &pb.FindByIdUserRequest{
+		Id: int32(id),
+	}
+
+	user, err := h.client.RestoreUser(ctx, req)
+
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to restore user: ",
+		})
+	}
+
+	return c.JSON(http.StatusOK, user)
+}
+
+func (h *userHandleApi) DeleteUserPermanent(c echo.Context) error {
+	id, err := strconv.Atoi(c.Param("id"))
+
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, response.ErrorResponse{
+			Status:  "error",
+			Message: "Invalid user ID",
+		})
+	}
+
+	ctx := c.Request().Context()
+
+	req := &pb.FindByIdUserRequest{
+		Id: int32(id),
+	}
+
+	user, err := h.client.DeleteUserPermanent(ctx, req)
+
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, response.ErrorResponse{
+			Status:  "error",
+			Message: "Failed to delete user: ",
+		})
+	}
+
+	return c.JSON(http.StatusOK, user)
 }
