@@ -1,45 +1,51 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v3.12.4
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: saldo.proto
 
 package pb
 
 import (
 	context "context"
-	empty "github.com/golang/protobuf/ptypes/empty"
 	grpc "google.golang.org/grpc"
 	codes "google.golang.org/grpc/codes"
 	status "google.golang.org/grpc/status"
+	emptypb "google.golang.org/protobuf/types/known/emptypb"
 )
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
-	SaldoService_GetSaldos_FullMethodName        = "/pb.SaldoService/GetSaldos"
-	SaldoService_GetSaldo_FullMethodName         = "/pb.SaldoService/GetSaldo"
-	SaldoService_GetSaldoByUsers_FullMethodName  = "/pb.SaldoService/GetSaldoByUsers"
-	SaldoService_GetSaldoByUserId_FullMethodName = "/pb.SaldoService/GetSaldoByUserId"
-	SaldoService_CreateSaldo_FullMethodName      = "/pb.SaldoService/CreateSaldo"
-	SaldoService_UpdateSaldo_FullMethodName      = "/pb.SaldoService/UpdateSaldo"
-	SaldoService_DeleteSaldo_FullMethodName      = "/pb.SaldoService/DeleteSaldo"
+	SaldoService_FindAllSaldo_FullMethodName         = "/pb.SaldoService/FindAllSaldo"
+	SaldoService_FindByIdSaldo_FullMethodName        = "/pb.SaldoService/FindByIdSaldo"
+	SaldoService_FindByCardNumber_FullMethodName     = "/pb.SaldoService/FindByCardNumber"
+	SaldoService_FindByActive_FullMethodName         = "/pb.SaldoService/FindByActive"
+	SaldoService_FindByTrashed_FullMethodName        = "/pb.SaldoService/FindByTrashed"
+	SaldoService_CreateSaldo_FullMethodName          = "/pb.SaldoService/CreateSaldo"
+	SaldoService_UpdateSaldo_FullMethodName          = "/pb.SaldoService/UpdateSaldo"
+	SaldoService_TrashSaldo_FullMethodName           = "/pb.SaldoService/TrashSaldo"
+	SaldoService_RestoreSaldo_FullMethodName         = "/pb.SaldoService/RestoreSaldo"
+	SaldoService_DeleteSaldoPermanent_FullMethodName = "/pb.SaldoService/DeleteSaldoPermanent"
 )
 
 // SaldoServiceClient is the client API for SaldoService service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
 type SaldoServiceClient interface {
-	GetSaldos(ctx context.Context, in *empty.Empty, opts ...grpc.CallOption) (*SaldoResponses, error)
-	GetSaldo(ctx context.Context, in *SaldoRequest, opts ...grpc.CallOption) (*SaldoResponse, error)
-	GetSaldoByUsers(ctx context.Context, in *SaldoRequest, opts ...grpc.CallOption) (*SaldoResponses, error)
-	GetSaldoByUserId(ctx context.Context, in *SaldoRequest, opts ...grpc.CallOption) (*SaldoResponse, error)
-	CreateSaldo(ctx context.Context, in *CreateSaldoRequest, opts ...grpc.CallOption) (*SaldoResponse, error)
-	UpdateSaldo(ctx context.Context, in *UpdateSaldoRequest, opts ...grpc.CallOption) (*SaldoResponse, error)
-	DeleteSaldo(ctx context.Context, in *SaldoRequest, opts ...grpc.CallOption) (*DeleteSaldoResponse, error)
+	FindAllSaldo(ctx context.Context, in *FindAllSaldoRequest, opts ...grpc.CallOption) (*ApiResponsePaginationSaldo, error)
+	FindByIdSaldo(ctx context.Context, in *FindByIdSaldoRequest, opts ...grpc.CallOption) (*ApiResponseSaldo, error)
+	FindByCardNumber(ctx context.Context, in *FindByCardNumberRequest, opts ...grpc.CallOption) (*ApiResponseSaldo, error)
+	FindByActive(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*ApiResponsesSaldo, error)
+	FindByTrashed(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*ApiResponsesSaldo, error)
+	CreateSaldo(ctx context.Context, in *CreateSaldoRequest, opts ...grpc.CallOption) (*ApiResponseSaldo, error)
+	UpdateSaldo(ctx context.Context, in *UpdateSaldoRequest, opts ...grpc.CallOption) (*ApiResponseSaldo, error)
+	TrashSaldo(ctx context.Context, in *FindByIdSaldoRequest, opts ...grpc.CallOption) (*ApiResponseSaldo, error)
+	RestoreSaldo(ctx context.Context, in *FindByIdSaldoRequest, opts ...grpc.CallOption) (*ApiResponseSaldo, error)
+	DeleteSaldoPermanent(ctx context.Context, in *FindByIdSaldoRequest, opts ...grpc.CallOption) (*ApiResponseSaldoDelete, error)
 }
 
 type saldoServiceClient struct {
@@ -50,63 +56,100 @@ func NewSaldoServiceClient(cc grpc.ClientConnInterface) SaldoServiceClient {
 	return &saldoServiceClient{cc}
 }
 
-func (c *saldoServiceClient) GetSaldos(ctx context.Context, in *empty.Empty, opts ...grpc.CallOption) (*SaldoResponses, error) {
-	out := new(SaldoResponses)
-	err := c.cc.Invoke(ctx, SaldoService_GetSaldos_FullMethodName, in, out, opts...)
+func (c *saldoServiceClient) FindAllSaldo(ctx context.Context, in *FindAllSaldoRequest, opts ...grpc.CallOption) (*ApiResponsePaginationSaldo, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponsePaginationSaldo)
+	err := c.cc.Invoke(ctx, SaldoService_FindAllSaldo_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *saldoServiceClient) GetSaldo(ctx context.Context, in *SaldoRequest, opts ...grpc.CallOption) (*SaldoResponse, error) {
-	out := new(SaldoResponse)
-	err := c.cc.Invoke(ctx, SaldoService_GetSaldo_FullMethodName, in, out, opts...)
+func (c *saldoServiceClient) FindByIdSaldo(ctx context.Context, in *FindByIdSaldoRequest, opts ...grpc.CallOption) (*ApiResponseSaldo, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseSaldo)
+	err := c.cc.Invoke(ctx, SaldoService_FindByIdSaldo_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *saldoServiceClient) GetSaldoByUsers(ctx context.Context, in *SaldoRequest, opts ...grpc.CallOption) (*SaldoResponses, error) {
-	out := new(SaldoResponses)
-	err := c.cc.Invoke(ctx, SaldoService_GetSaldoByUsers_FullMethodName, in, out, opts...)
+func (c *saldoServiceClient) FindByCardNumber(ctx context.Context, in *FindByCardNumberRequest, opts ...grpc.CallOption) (*ApiResponseSaldo, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseSaldo)
+	err := c.cc.Invoke(ctx, SaldoService_FindByCardNumber_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *saldoServiceClient) GetSaldoByUserId(ctx context.Context, in *SaldoRequest, opts ...grpc.CallOption) (*SaldoResponse, error) {
-	out := new(SaldoResponse)
-	err := c.cc.Invoke(ctx, SaldoService_GetSaldoByUserId_FullMethodName, in, out, opts...)
+func (c *saldoServiceClient) FindByActive(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*ApiResponsesSaldo, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponsesSaldo)
+	err := c.cc.Invoke(ctx, SaldoService_FindByActive_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *saldoServiceClient) CreateSaldo(ctx context.Context, in *CreateSaldoRequest, opts ...grpc.CallOption) (*SaldoResponse, error) {
-	out := new(SaldoResponse)
-	err := c.cc.Invoke(ctx, SaldoService_CreateSaldo_FullMethodName, in, out, opts...)
+func (c *saldoServiceClient) FindByTrashed(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*ApiResponsesSaldo, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponsesSaldo)
+	err := c.cc.Invoke(ctx, SaldoService_FindByTrashed_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *saldoServiceClient) UpdateSaldo(ctx context.Context, in *UpdateSaldoRequest, opts ...grpc.CallOption) (*SaldoResponse, error) {
-	out := new(SaldoResponse)
-	err := c.cc.Invoke(ctx, SaldoService_UpdateSaldo_FullMethodName, in, out, opts...)
+func (c *saldoServiceClient) CreateSaldo(ctx context.Context, in *CreateSaldoRequest, opts ...grpc.CallOption) (*ApiResponseSaldo, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseSaldo)
+	err := c.cc.Invoke(ctx, SaldoService_CreateSaldo_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *saldoServiceClient) DeleteSaldo(ctx context.Context, in *SaldoRequest, opts ...grpc.CallOption) (*DeleteSaldoResponse, error) {
-	out := new(DeleteSaldoResponse)
-	err := c.cc.Invoke(ctx, SaldoService_DeleteSaldo_FullMethodName, in, out, opts...)
+func (c *saldoServiceClient) UpdateSaldo(ctx context.Context, in *UpdateSaldoRequest, opts ...grpc.CallOption) (*ApiResponseSaldo, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseSaldo)
+	err := c.cc.Invoke(ctx, SaldoService_UpdateSaldo_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *saldoServiceClient) TrashSaldo(ctx context.Context, in *FindByIdSaldoRequest, opts ...grpc.CallOption) (*ApiResponseSaldo, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseSaldo)
+	err := c.cc.Invoke(ctx, SaldoService_TrashSaldo_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *saldoServiceClient) RestoreSaldo(ctx context.Context, in *FindByIdSaldoRequest, opts ...grpc.CallOption) (*ApiResponseSaldo, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseSaldo)
+	err := c.cc.Invoke(ctx, SaldoService_RestoreSaldo_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *saldoServiceClient) DeleteSaldoPermanent(ctx context.Context, in *FindByIdSaldoRequest, opts ...grpc.CallOption) (*ApiResponseSaldoDelete, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseSaldoDelete)
+	err := c.cc.Invoke(ctx, SaldoService_DeleteSaldoPermanent_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -115,44 +158,60 @@ func (c *saldoServiceClient) DeleteSaldo(ctx context.Context, in *SaldoRequest, 
 
 // SaldoServiceServer is the server API for SaldoService service.
 // All implementations must embed UnimplementedSaldoServiceServer
-// for forward compatibility
+// for forward compatibility.
 type SaldoServiceServer interface {
-	GetSaldos(context.Context, *empty.Empty) (*SaldoResponses, error)
-	GetSaldo(context.Context, *SaldoRequest) (*SaldoResponse, error)
-	GetSaldoByUsers(context.Context, *SaldoRequest) (*SaldoResponses, error)
-	GetSaldoByUserId(context.Context, *SaldoRequest) (*SaldoResponse, error)
-	CreateSaldo(context.Context, *CreateSaldoRequest) (*SaldoResponse, error)
-	UpdateSaldo(context.Context, *UpdateSaldoRequest) (*SaldoResponse, error)
-	DeleteSaldo(context.Context, *SaldoRequest) (*DeleteSaldoResponse, error)
+	FindAllSaldo(context.Context, *FindAllSaldoRequest) (*ApiResponsePaginationSaldo, error)
+	FindByIdSaldo(context.Context, *FindByIdSaldoRequest) (*ApiResponseSaldo, error)
+	FindByCardNumber(context.Context, *FindByCardNumberRequest) (*ApiResponseSaldo, error)
+	FindByActive(context.Context, *emptypb.Empty) (*ApiResponsesSaldo, error)
+	FindByTrashed(context.Context, *emptypb.Empty) (*ApiResponsesSaldo, error)
+	CreateSaldo(context.Context, *CreateSaldoRequest) (*ApiResponseSaldo, error)
+	UpdateSaldo(context.Context, *UpdateSaldoRequest) (*ApiResponseSaldo, error)
+	TrashSaldo(context.Context, *FindByIdSaldoRequest) (*ApiResponseSaldo, error)
+	RestoreSaldo(context.Context, *FindByIdSaldoRequest) (*ApiResponseSaldo, error)
+	DeleteSaldoPermanent(context.Context, *FindByIdSaldoRequest) (*ApiResponseSaldoDelete, error)
 	mustEmbedUnimplementedSaldoServiceServer()
 }
 
-// UnimplementedSaldoServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedSaldoServiceServer struct {
-}
+// UnimplementedSaldoServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedSaldoServiceServer struct{}
 
-func (UnimplementedSaldoServiceServer) GetSaldos(context.Context, *empty.Empty) (*SaldoResponses, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method GetSaldos not implemented")
+func (UnimplementedSaldoServiceServer) FindAllSaldo(context.Context, *FindAllSaldoRequest) (*ApiResponsePaginationSaldo, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method FindAllSaldo not implemented")
 }
-func (UnimplementedSaldoServiceServer) GetSaldo(context.Context, *SaldoRequest) (*SaldoResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method GetSaldo not implemented")
+func (UnimplementedSaldoServiceServer) FindByIdSaldo(context.Context, *FindByIdSaldoRequest) (*ApiResponseSaldo, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method FindByIdSaldo not implemented")
 }
-func (UnimplementedSaldoServiceServer) GetSaldoByUsers(context.Context, *SaldoRequest) (*SaldoResponses, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method GetSaldoByUsers not implemented")
+func (UnimplementedSaldoServiceServer) FindByCardNumber(context.Context, *FindByCardNumberRequest) (*ApiResponseSaldo, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method FindByCardNumber not implemented")
 }
-func (UnimplementedSaldoServiceServer) GetSaldoByUserId(context.Context, *SaldoRequest) (*SaldoResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method GetSaldoByUserId not implemented")
+func (UnimplementedSaldoServiceServer) FindByActive(context.Context, *emptypb.Empty) (*ApiResponsesSaldo, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method FindByActive not implemented")
 }
-func (UnimplementedSaldoServiceServer) CreateSaldo(context.Context, *CreateSaldoRequest) (*SaldoResponse, error) {
+func (UnimplementedSaldoServiceServer) FindByTrashed(context.Context, *emptypb.Empty) (*ApiResponsesSaldo, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method FindByTrashed not implemented")
+}
+func (UnimplementedSaldoServiceServer) CreateSaldo(context.Context, *CreateSaldoRequest) (*ApiResponseSaldo, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method CreateSaldo not implemented")
 }
-func (UnimplementedSaldoServiceServer) UpdateSaldo(context.Context, *UpdateSaldoRequest) (*SaldoResponse, error) {
+func (UnimplementedSaldoServiceServer) UpdateSaldo(context.Context, *UpdateSaldoRequest) (*ApiResponseSaldo, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateSaldo not implemented")
 }
-func (UnimplementedSaldoServiceServer) DeleteSaldo(context.Context, *SaldoRequest) (*DeleteSaldoResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method DeleteSaldo not implemented")
+func (UnimplementedSaldoServiceServer) TrashSaldo(context.Context, *FindByIdSaldoRequest) (*ApiResponseSaldo, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method TrashSaldo not implemented")
+}
+func (UnimplementedSaldoServiceServer) RestoreSaldo(context.Context, *FindByIdSaldoRequest) (*ApiResponseSaldo, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method RestoreSaldo not implemented")
+}
+func (UnimplementedSaldoServiceServer) DeleteSaldoPermanent(context.Context, *FindByIdSaldoRequest) (*ApiResponseSaldoDelete, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method DeleteSaldoPermanent not implemented")
 }
 func (UnimplementedSaldoServiceServer) mustEmbedUnimplementedSaldoServiceServer() {}
+func (UnimplementedSaldoServiceServer) testEmbeddedByValue()                      {}
 
 // UnsafeSaldoServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to SaldoServiceServer will
@@ -162,77 +221,102 @@ type UnsafeSaldoServiceServer interface {
 }
 
 func RegisterSaldoServiceServer(s grpc.ServiceRegistrar, srv SaldoServiceServer) {
+	// If the following call pancis, it indicates UnimplementedSaldoServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&SaldoService_ServiceDesc, srv)
 }
 
-func _SaldoService_GetSaldos_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(empty.Empty)
+func _SaldoService_FindAllSaldo_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindAllSaldoRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(SaldoServiceServer).GetSaldos(ctx, in)
+		return srv.(SaldoServiceServer).FindAllSaldo(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: SaldoService_GetSaldos_FullMethodName,
+		FullMethod: SaldoService_FindAllSaldo_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(SaldoServiceServer).GetSaldos(ctx, req.(*empty.Empty))
+		return srv.(SaldoServiceServer).FindAllSaldo(ctx, req.(*FindAllSaldoRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
-func _SaldoService_GetSaldo_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(SaldoRequest)
+func _SaldoService_FindByIdSaldo_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindByIdSaldoRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(SaldoServiceServer).GetSaldo(ctx, in)
+		return srv.(SaldoServiceServer).FindByIdSaldo(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: SaldoService_GetSaldo_FullMethodName,
+		FullMethod: SaldoService_FindByIdSaldo_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(SaldoServiceServer).GetSaldo(ctx, req.(*SaldoRequest))
+		return srv.(SaldoServiceServer).FindByIdSaldo(ctx, req.(*FindByIdSaldoRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
-func _SaldoService_GetSaldoByUsers_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(SaldoRequest)
+func _SaldoService_FindByCardNumber_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindByCardNumberRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(SaldoServiceServer).GetSaldoByUsers(ctx, in)
+		return srv.(SaldoServiceServer).FindByCardNumber(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: SaldoService_GetSaldoByUsers_FullMethodName,
+		FullMethod: SaldoService_FindByCardNumber_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(SaldoServiceServer).GetSaldoByUsers(ctx, req.(*SaldoRequest))
+		return srv.(SaldoServiceServer).FindByCardNumber(ctx, req.(*FindByCardNumberRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
-func _SaldoService_GetSaldoByUserId_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(SaldoRequest)
+func _SaldoService_FindByActive_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(emptypb.Empty)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(SaldoServiceServer).GetSaldoByUserId(ctx, in)
+		return srv.(SaldoServiceServer).FindByActive(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: SaldoService_GetSaldoByUserId_FullMethodName,
+		FullMethod: SaldoService_FindByActive_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(SaldoServiceServer).GetSaldoByUserId(ctx, req.(*SaldoRequest))
+		return srv.(SaldoServiceServer).FindByActive(ctx, req.(*emptypb.Empty))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _SaldoService_FindByTrashed_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(emptypb.Empty)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(SaldoServiceServer).FindByTrashed(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: SaldoService_FindByTrashed_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(SaldoServiceServer).FindByTrashed(ctx, req.(*emptypb.Empty))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -273,20 +357,56 @@ func _SaldoService_UpdateSaldo_Handler(srv interface{}, ctx context.Context, dec
 	return interceptor(ctx, in, info, handler)
 }
 
-func _SaldoService_DeleteSaldo_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(SaldoRequest)
+func _SaldoService_TrashSaldo_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindByIdSaldoRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(SaldoServiceServer).DeleteSaldo(ctx, in)
+		return srv.(SaldoServiceServer).TrashSaldo(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: SaldoService_DeleteSaldo_FullMethodName,
+		FullMethod: SaldoService_TrashSaldo_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(SaldoServiceServer).DeleteSaldo(ctx, req.(*SaldoRequest))
+		return srv.(SaldoServiceServer).TrashSaldo(ctx, req.(*FindByIdSaldoRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _SaldoService_RestoreSaldo_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindByIdSaldoRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(SaldoServiceServer).RestoreSaldo(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: SaldoService_RestoreSaldo_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(SaldoServiceServer).RestoreSaldo(ctx, req.(*FindByIdSaldoRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _SaldoService_DeleteSaldoPermanent_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindByIdSaldoRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(SaldoServiceServer).DeleteSaldoPermanent(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: SaldoService_DeleteSaldoPermanent_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(SaldoServiceServer).DeleteSaldoPermanent(ctx, req.(*FindByIdSaldoRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -299,20 +419,24 @@ var SaldoService_ServiceDesc = grpc.ServiceDesc{
 	HandlerType: (*SaldoServiceServer)(nil),
 	Methods: []grpc.MethodDesc{
 		{
-			MethodName: "GetSaldos",
-			Handler:    _SaldoService_GetSaldos_Handler,
+			MethodName: "FindAllSaldo",
+			Handler:    _SaldoService_FindAllSaldo_Handler,
 		},
 		{
-			MethodName: "GetSaldo",
-			Handler:    _SaldoService_GetSaldo_Handler,
+			MethodName: "FindByIdSaldo",
+			Handler:    _SaldoService_FindByIdSaldo_Handler,
 		},
 		{
-			MethodName: "GetSaldoByUsers",
-			Handler:    _SaldoService_GetSaldoByUsers_Handler,
+			MethodName: "FindByCardNumber",
+			Handler:    _SaldoService_FindByCardNumber_Handler,
 		},
 		{
-			MethodName: "GetSaldoByUserId",
-			Handler:    _SaldoService_GetSaldoByUserId_Handler,
+			MethodName: "FindByActive",
+			Handler:    _SaldoService_FindByActive_Handler,
+		},
+		{
+			MethodName: "FindByTrashed",
+			Handler:    _SaldoService_FindByTrashed_Handler,
 		},
 		{
 			MethodName: "CreateSaldo",
@@ -323,8 +447,16 @@ var SaldoService_ServiceDesc = grpc.ServiceDesc{
 			Handler:    _SaldoService_UpdateSaldo_Handler,
 		},
 		{
-			MethodName: "DeleteSaldo",
-			Handler:    _SaldoService_DeleteSaldo_Handler,
+			MethodName: "TrashSaldo",
+			Handler:    _SaldoService_TrashSaldo_Handler,
+		},
+		{
+			MethodName: "RestoreSaldo",
+			Handler:    _SaldoService_RestoreSaldo_Handler,
+		},
+		{
+			MethodName: "DeleteSaldoPermanent",
+			Handler:    _SaldoService_DeleteSaldoPermanent_Handler,
 		},
 	},
 	Streams:  []grpc.StreamDesc{},
