@@ -1,45 +1,51 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v3.12.4
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: topup.proto
 
 package pb
 
 import (
 	context "context"
-	empty "github.com/golang/protobuf/ptypes/empty"
 	grpc "google.golang.org/grpc"
 	codes "google.golang.org/grpc/codes"
 	status "google.golang.org/grpc/status"
+	emptypb "google.golang.org/protobuf/types/known/emptypb"
 )
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
-	TopupService_GetTopups_FullMethodName        = "/pb.TopupService/GetTopups"
-	TopupService_GetTopup_FullMethodName         = "/pb.TopupService/GetTopup"
-	TopupService_GetTopupByUsers_FullMethodName  = "/pb.TopupService/GetTopupByUsers"
-	TopupService_GetTopupByUserId_FullMethodName = "/pb.TopupService/GetTopupByUserId"
-	TopupService_CreateTopup_FullMethodName      = "/pb.TopupService/CreateTopup"
-	TopupService_UpdateTopup_FullMethodName      = "/pb.TopupService/UpdateTopup"
-	TopupService_DeleteTopup_FullMethodName      = "/pb.TopupService/DeleteTopup"
+	TopupService_FindAllTopup_FullMethodName          = "/pb.TopupService/FindAllTopup"
+	TopupService_FindByIdTopup_FullMethodName         = "/pb.TopupService/FindByIdTopup"
+	TopupService_FindByCardNumberTopup_FullMethodName = "/pb.TopupService/FindByCardNumberTopup"
+	TopupService_FindByActive_FullMethodName          = "/pb.TopupService/FindByActive"
+	TopupService_FindByTrashed_FullMethodName         = "/pb.TopupService/FindByTrashed"
+	TopupService_CreateTopup_FullMethodName           = "/pb.TopupService/CreateTopup"
+	TopupService_UpdateTopup_FullMethodName           = "/pb.TopupService/UpdateTopup"
+	TopupService_TrashedTopup_FullMethodName          = "/pb.TopupService/TrashedTopup"
+	TopupService_RestoreTopup_FullMethodName          = "/pb.TopupService/RestoreTopup"
+	TopupService_DeleteTopupPermanent_FullMethodName  = "/pb.TopupService/DeleteTopupPermanent"
 )
 
 // TopupServiceClient is the client API for TopupService service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
 type TopupServiceClient interface {
-	GetTopups(ctx context.Context, in *empty.Empty, opts ...grpc.CallOption) (*TopupsResponse, error)
-	GetTopup(ctx context.Context, in *TopupRequest, opts ...grpc.CallOption) (*TopupResponse, error)
-	GetTopupByUsers(ctx context.Context, in *TopupRequest, opts ...grpc.CallOption) (*TopupsResponse, error)
-	GetTopupByUserId(ctx context.Context, in *TopupRequest, opts ...grpc.CallOption) (*TopupResponse, error)
-	CreateTopup(ctx context.Context, in *CreateTopupRequest, opts ...grpc.CallOption) (*TopupResponse, error)
-	UpdateTopup(ctx context.Context, in *UpdateTopupRequest, opts ...grpc.CallOption) (*TopupResponse, error)
-	DeleteTopup(ctx context.Context, in *TopupRequest, opts ...grpc.CallOption) (*DeleteTopupResponse, error)
+	FindAllTopup(ctx context.Context, in *FindAllTopupRequest, opts ...grpc.CallOption) (*ApiResponsePaginationTopup, error)
+	FindByIdTopup(ctx context.Context, in *FindByIdTopupRequest, opts ...grpc.CallOption) (*ApiResponseTopup, error)
+	FindByCardNumberTopup(ctx context.Context, in *FindByCardNumberTopupRequest, opts ...grpc.CallOption) (*ApiResponseTopup, error)
+	FindByActive(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*ApiResponsesTopup, error)
+	FindByTrashed(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*ApiResponsesTopup, error)
+	CreateTopup(ctx context.Context, in *CreateTopupRequest, opts ...grpc.CallOption) (*ApiResponseTopup, error)
+	UpdateTopup(ctx context.Context, in *UpdateTopupRequest, opts ...grpc.CallOption) (*ApiResponseTopup, error)
+	TrashedTopup(ctx context.Context, in *FindByIdTopupRequest, opts ...grpc.CallOption) (*ApiResponseTopup, error)
+	RestoreTopup(ctx context.Context, in *FindByIdTopupRequest, opts ...grpc.CallOption) (*ApiResponseTopup, error)
+	DeleteTopupPermanent(ctx context.Context, in *FindByIdTopupRequest, opts ...grpc.CallOption) (*ApiResponseTopupDelete, error)
 }
 
 type topupServiceClient struct {
@@ -50,63 +56,100 @@ func NewTopupServiceClient(cc grpc.ClientConnInterface) TopupServiceClient {
 	return &topupServiceClient{cc}
 }
 
-func (c *topupServiceClient) GetTopups(ctx context.Context, in *empty.Empty, opts ...grpc.CallOption) (*TopupsResponse, error) {
-	out := new(TopupsResponse)
-	err := c.cc.Invoke(ctx, TopupService_GetTopups_FullMethodName, in, out, opts...)
+func (c *topupServiceClient) FindAllTopup(ctx context.Context, in *FindAllTopupRequest, opts ...grpc.CallOption) (*ApiResponsePaginationTopup, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponsePaginationTopup)
+	err := c.cc.Invoke(ctx, TopupService_FindAllTopup_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *topupServiceClient) GetTopup(ctx context.Context, in *TopupRequest, opts ...grpc.CallOption) (*TopupResponse, error) {
-	out := new(TopupResponse)
-	err := c.cc.Invoke(ctx, TopupService_GetTopup_FullMethodName, in, out, opts...)
+func (c *topupServiceClient) FindByIdTopup(ctx context.Context, in *FindByIdTopupRequest, opts ...grpc.CallOption) (*ApiResponseTopup, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTopup)
+	err := c.cc.Invoke(ctx, TopupService_FindByIdTopup_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *topupServiceClient) GetTopupByUsers(ctx context.Context, in *TopupRequest, opts ...grpc.CallOption) (*TopupsResponse, error) {
-	out := new(TopupsResponse)
-	err := c.cc.Invoke(ctx, TopupService_GetTopupByUsers_FullMethodName, in, out, opts...)
+func (c *topupServiceClient) FindByCardNumberTopup(ctx context.Context, in *FindByCardNumberTopupRequest, opts ...grpc.CallOption) (*ApiResponseTopup, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTopup)
+	err := c.cc.Invoke(ctx, TopupService_FindByCardNumberTopup_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *topupServiceClient) GetTopupByUserId(ctx context.Context, in *TopupRequest, opts ...grpc.CallOption) (*TopupResponse, error) {
-	out := new(TopupResponse)
-	err := c.cc.Invoke(ctx, TopupService_GetTopupByUserId_FullMethodName, in, out, opts...)
+func (c *topupServiceClient) FindByActive(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*ApiResponsesTopup, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponsesTopup)
+	err := c.cc.Invoke(ctx, TopupService_FindByActive_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *topupServiceClient) CreateTopup(ctx context.Context, in *CreateTopupRequest, opts ...grpc.CallOption) (*TopupResponse, error) {
-	out := new(TopupResponse)
-	err := c.cc.Invoke(ctx, TopupService_CreateTopup_FullMethodName, in, out, opts...)
+func (c *topupServiceClient) FindByTrashed(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*ApiResponsesTopup, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponsesTopup)
+	err := c.cc.Invoke(ctx, TopupService_FindByTrashed_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *topupServiceClient) UpdateTopup(ctx context.Context, in *UpdateTopupRequest, opts ...grpc.CallOption) (*TopupResponse, error) {
-	out := new(TopupResponse)
-	err := c.cc.Invoke(ctx, TopupService_UpdateTopup_FullMethodName, in, out, opts...)
+func (c *topupServiceClient) CreateTopup(ctx context.Context, in *CreateTopupRequest, opts ...grpc.CallOption) (*ApiResponseTopup, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTopup)
+	err := c.cc.Invoke(ctx, TopupService_CreateTopup_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *topupServiceClient) DeleteTopup(ctx context.Context, in *TopupRequest, opts ...grpc.CallOption) (*DeleteTopupResponse, error) {
-	out := new(DeleteTopupResponse)
-	err := c.cc.Invoke(ctx, TopupService_DeleteTopup_FullMethodName, in, out, opts...)
+func (c *topupServiceClient) UpdateTopup(ctx context.Context, in *UpdateTopupRequest, opts ...grpc.CallOption) (*ApiResponseTopup, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTopup)
+	err := c.cc.Invoke(ctx, TopupService_UpdateTopup_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *topupServiceClient) TrashedTopup(ctx context.Context, in *FindByIdTopupRequest, opts ...grpc.CallOption) (*ApiResponseTopup, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTopup)
+	err := c.cc.Invoke(ctx, TopupService_TrashedTopup_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *topupServiceClient) RestoreTopup(ctx context.Context, in *FindByIdTopupRequest, opts ...grpc.CallOption) (*ApiResponseTopup, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTopup)
+	err := c.cc.Invoke(ctx, TopupService_RestoreTopup_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *topupServiceClient) DeleteTopupPermanent(ctx context.Context, in *FindByIdTopupRequest, opts ...grpc.CallOption) (*ApiResponseTopupDelete, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTopupDelete)
+	err := c.cc.Invoke(ctx, TopupService_DeleteTopupPermanent_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -115,44 +158,60 @@ func (c *topupServiceClient) DeleteTopup(ctx context.Context, in *TopupRequest, 
 
 // TopupServiceServer is the server API for TopupService service.
 // All implementations must embed UnimplementedTopupServiceServer
-// for forward compatibility
+// for forward compatibility.
 type TopupServiceServer interface {
-	GetTopups(context.Context, *empty.Empty) (*TopupsResponse, error)
-	GetTopup(context.Context, *TopupRequest) (*TopupResponse, error)
-	GetTopupByUsers(context.Context, *TopupRequest) (*TopupsResponse, error)
-	GetTopupByUserId(context.Context, *TopupRequest) (*TopupResponse, error)
-	CreateTopup(context.Context, *CreateTopupRequest) (*TopupResponse, error)
-	UpdateTopup(context.Context, *UpdateTopupRequest) (*TopupResponse, error)
-	DeleteTopup(context.Context, *TopupRequest) (*DeleteTopupResponse, error)
+	FindAllTopup(context.Context, *FindAllTopupRequest) (*ApiResponsePaginationTopup, error)
+	FindByIdTopup(context.Context, *FindByIdTopupRequest) (*ApiResponseTopup, error)
+	FindByCardNumberTopup(context.Context, *FindByCardNumberTopupRequest) (*ApiResponseTopup, error)
+	FindByActive(context.Context, *emptypb.Empty) (*ApiResponsesTopup, error)
+	FindByTrashed(context.Context, *emptypb.Empty) (*ApiResponsesTopup, error)
+	CreateTopup(context.Context, *CreateTopupRequest) (*ApiResponseTopup, error)
+	UpdateTopup(context.Context, *UpdateTopupRequest) (*ApiResponseTopup, error)
+	TrashedTopup(context.Context, *FindByIdTopupRequest) (*ApiResponseTopup, error)
+	RestoreTopup(context.Context, *FindByIdTopupRequest) (*ApiResponseTopup, error)
+	DeleteTopupPermanent(context.Context, *FindByIdTopupRequest) (*ApiResponseTopupDelete, error)
 	mustEmbedUnimplementedTopupServiceServer()
 }
 
-// UnimplementedTopupServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedTopupServiceServer struct {
-}
+// UnimplementedTopupServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedTopupServiceServer struct{}
 
-func (UnimplementedTopupServiceServer) GetTopups(context.Context, *empty.Empty) (*TopupsResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method GetTopups not implemented")
+func (UnimplementedTopupServiceServer) FindAllTopup(context.Context, *FindAllTopupRequest) (*ApiResponsePaginationTopup, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method FindAllTopup not implemented")
 }
-func (UnimplementedTopupServiceServer) GetTopup(context.Context, *TopupRequest) (*TopupResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method GetTopup not implemented")
+func (UnimplementedTopupServiceServer) FindByIdTopup(context.Context, *FindByIdTopupRequest) (*ApiResponseTopup, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method FindByIdTopup not implemented")
 }
-func (UnimplementedTopupServiceServer) GetTopupByUsers(context.Context, *TopupRequest) (*TopupsResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method GetTopupByUsers not implemented")
+func (UnimplementedTopupServiceServer) FindByCardNumberTopup(context.Context, *FindByCardNumberTopupRequest) (*ApiResponseTopup, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method FindByCardNumberTopup not implemented")
 }
-func (UnimplementedTopupServiceServer) GetTopupByUserId(context.Context, *TopupRequest) (*TopupResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method GetTopupByUserId not implemented")
+func (UnimplementedTopupServiceServer) FindByActive(context.Context, *emptypb.Empty) (*ApiResponsesTopup, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method FindByActive not implemented")
 }
-func (UnimplementedTopupServiceServer) CreateTopup(context.Context, *CreateTopupRequest) (*TopupResponse, error) {
+func (UnimplementedTopupServiceServer) FindByTrashed(context.Context, *emptypb.Empty) (*ApiResponsesTopup, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method FindByTrashed not implemented")
+}
+func (UnimplementedTopupServiceServer) CreateTopup(context.Context, *CreateTopupRequest) (*ApiResponseTopup, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method CreateTopup not implemented")
 }
-func (UnimplementedTopupServiceServer) UpdateTopup(context.Context, *UpdateTopupRequest) (*TopupResponse, error) {
+func (UnimplementedTopupServiceServer) UpdateTopup(context.Context, *UpdateTopupRequest) (*ApiResponseTopup, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateTopup not implemented")
 }
-func (UnimplementedTopupServiceServer) DeleteTopup(context.Context, *TopupRequest) (*DeleteTopupResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method DeleteTopup not implemented")
+func (UnimplementedTopupServiceServer) TrashedTopup(context.Context, *FindByIdTopupRequest) (*ApiResponseTopup, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method TrashedTopup not implemented")
+}
+func (UnimplementedTopupServiceServer) RestoreTopup(context.Context, *FindByIdTopupRequest) (*ApiResponseTopup, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method RestoreTopup not implemented")
+}
+func (UnimplementedTopupServiceServer) DeleteTopupPermanent(context.Context, *FindByIdTopupRequest) (*ApiResponseTopupDelete, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method DeleteTopupPermanent not implemented")
 }
 func (UnimplementedTopupServiceServer) mustEmbedUnimplementedTopupServiceServer() {}
+func (UnimplementedTopupServiceServer) testEmbeddedByValue()                      {}
 
 // UnsafeTopupServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to TopupServiceServer will
@@ -162,77 +221,102 @@ type UnsafeTopupServiceServer interface {
 }
 
 func RegisterTopupServiceServer(s grpc.ServiceRegistrar, srv TopupServiceServer) {
+	// If the following call pancis, it indicates UnimplementedTopupServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&TopupService_ServiceDesc, srv)
 }
 
-func _TopupService_GetTopups_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(empty.Empty)
+func _TopupService_FindAllTopup_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindAllTopupRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(TopupServiceServer).GetTopups(ctx, in)
+		return srv.(TopupServiceServer).FindAllTopup(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: TopupService_GetTopups_FullMethodName,
+		FullMethod: TopupService_FindAllTopup_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(TopupServiceServer).GetTopups(ctx, req.(*empty.Empty))
+		return srv.(TopupServiceServer).FindAllTopup(ctx, req.(*FindAllTopupRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
-func _TopupService_GetTopup_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(TopupRequest)
+func _TopupService_FindByIdTopup_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindByIdTopupRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(TopupServiceServer).GetTopup(ctx, in)
+		return srv.(TopupServiceServer).FindByIdTopup(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: TopupService_GetTopup_FullMethodName,
+		FullMethod: TopupService_FindByIdTopup_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(TopupServiceServer).GetTopup(ctx, req.(*TopupRequest))
+		return srv.(TopupServiceServer).FindByIdTopup(ctx, req.(*FindByIdTopupRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
-func _TopupService_GetTopupByUsers_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(TopupRequest)
+func _TopupService_FindByCardNumberTopup_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindByCardNumberTopupRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(TopupServiceServer).GetTopupByUsers(ctx, in)
+		return srv.(TopupServiceServer).FindByCardNumberTopup(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: TopupService_GetTopupByUsers_FullMethodName,
+		FullMethod: TopupService_FindByCardNumberTopup_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(TopupServiceServer).GetTopupByUsers(ctx, req.(*TopupRequest))
+		return srv.(TopupServiceServer).FindByCardNumberTopup(ctx, req.(*FindByCardNumberTopupRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
-func _TopupService_GetTopupByUserId_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(TopupRequest)
+func _TopupService_FindByActive_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(emptypb.Empty)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(TopupServiceServer).GetTopupByUserId(ctx, in)
+		return srv.(TopupServiceServer).FindByActive(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: TopupService_GetTopupByUserId_FullMethodName,
+		FullMethod: TopupService_FindByActive_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(TopupServiceServer).GetTopupByUserId(ctx, req.(*TopupRequest))
+		return srv.(TopupServiceServer).FindByActive(ctx, req.(*emptypb.Empty))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _TopupService_FindByTrashed_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(emptypb.Empty)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(TopupServiceServer).FindByTrashed(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: TopupService_FindByTrashed_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(TopupServiceServer).FindByTrashed(ctx, req.(*emptypb.Empty))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -273,20 +357,56 @@ func _TopupService_UpdateTopup_Handler(srv interface{}, ctx context.Context, dec
 	return interceptor(ctx, in, info, handler)
 }
 
-func _TopupService_DeleteTopup_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(TopupRequest)
+func _TopupService_TrashedTopup_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindByIdTopupRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(TopupServiceServer).DeleteTopup(ctx, in)
+		return srv.(TopupServiceServer).TrashedTopup(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: TopupService_DeleteTopup_FullMethodName,
+		FullMethod: TopupService_TrashedTopup_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(TopupServiceServer).DeleteTopup(ctx, req.(*TopupRequest))
+		return srv.(TopupServiceServer).TrashedTopup(ctx, req.(*FindByIdTopupRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _TopupService_RestoreTopup_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindByIdTopupRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(TopupServiceServer).RestoreTopup(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: TopupService_RestoreTopup_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(TopupServiceServer).RestoreTopup(ctx, req.(*FindByIdTopupRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _TopupService_DeleteTopupPermanent_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindByIdTopupRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(TopupServiceServer).DeleteTopupPermanent(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: TopupService_DeleteTopupPermanent_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(TopupServiceServer).DeleteTopupPermanent(ctx, req.(*FindByIdTopupRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -299,20 +419,24 @@ var TopupService_ServiceDesc = grpc.ServiceDesc{
 	HandlerType: (*TopupServiceServer)(nil),
 	Methods: []grpc.MethodDesc{
 		{
-			MethodName: "GetTopups",
-			Handler:    _TopupService_GetTopups_Handler,
+			MethodName: "FindAllTopup",
+			Handler:    _TopupService_FindAllTopup_Handler,
 		},
 		{
-			MethodName: "GetTopup",
-			Handler:    _TopupService_GetTopup_Handler,
+			MethodName: "FindByIdTopup",
+			Handler:    _TopupService_FindByIdTopup_Handler,
 		},
 		{
-			MethodName: "GetTopupByUsers",
-			Handler:    _TopupService_GetTopupByUsers_Handler,
+			MethodName: "FindByCardNumberTopup",
+			Handler:    _TopupService_FindByCardNumberTopup_Handler,
 		},
 		{
-			MethodName: "GetTopupByUserId",
-			Handler:    _TopupService_GetTopupByUserId_Handler,
+			MethodName: "FindByActive",
+			Handler:    _TopupService_FindByActive_Handler,
+		},
+		{
+			MethodName: "FindByTrashed",
+			Handler:    _TopupService_FindByTrashed_Handler,
 		},
 		{
 			MethodName: "CreateTopup",
@@ -323,8 +447,16 @@ var TopupService_ServiceDesc = grpc.ServiceDesc{
 			Handler:    _TopupService_UpdateTopup_Handler,
 		},
 		{
-			MethodName: "DeleteTopup",
-			Handler:    _TopupService_DeleteTopup_Handler,
+			MethodName: "TrashedTopup",
+			Handler:    _TopupService_TrashedTopup_Handler,
+		},
+		{
+			MethodName: "RestoreTopup",
+			Handler:    _TopupService_RestoreTopup_Handler,
+		},
+		{
+			MethodName: "DeleteTopupPermanent",
+			Handler:    _TopupService_DeleteTopupPermanent_Handler,
 		},
 	},
 	Streams:  []grpc.StreamDesc{},
