@@ -1,45 +1,53 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v3.12.4
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.28.2
 // source: transfer.proto
 
 package pb
 
 import (
 	context "context"
-	empty "github.com/golang/protobuf/ptypes/empty"
 	grpc "google.golang.org/grpc"
 	codes "google.golang.org/grpc/codes"
 	status "google.golang.org/grpc/status"
+	emptypb "google.golang.org/protobuf/types/known/emptypb"
 )
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
-	TransferService_GetTransfers_FullMethodName        = "/pb.TransferService/GetTransfers"
-	TransferService_GetTransfer_FullMethodName         = "/pb.TransferService/GetTransfer"
-	TransferService_GetTransferByUsers_FullMethodName  = "/pb.TransferService/GetTransferByUsers"
-	TransferService_GetTransferByUserId_FullMethodName = "/pb.TransferService/GetTransferByUserId"
-	TransferService_CreateTransfer_FullMethodName      = "/pb.TransferService/CreateTransfer"
-	TransferService_UpdateTransfer_FullMethodName      = "/pb.TransferService/UpdateTransfer"
-	TransferService_DeleteTransfer_FullMethodName      = "/pb.TransferService/DeleteTransfer"
+	TransferService_FindAllTransfer_FullMethodName            = "/pb.TransferService/FindAllTransfer"
+	TransferService_FindByIdTransfer_FullMethodName           = "/pb.TransferService/FindByIdTransfer"
+	TransferService_FindTransferByTransferFrom_FullMethodName = "/pb.TransferService/FindTransferByTransferFrom"
+	TransferService_FindTransferByTransferTo_FullMethodName   = "/pb.TransferService/FindTransferByTransferTo"
+	TransferService_FindByActiveTransfer_FullMethodName       = "/pb.TransferService/FindByActiveTransfer"
+	TransferService_FindByTrashedTransfer_FullMethodName      = "/pb.TransferService/FindByTrashedTransfer"
+	TransferService_CreateTransfer_FullMethodName             = "/pb.TransferService/CreateTransfer"
+	TransferService_UpdateTransfer_FullMethodName             = "/pb.TransferService/UpdateTransfer"
+	TransferService_TrashedTransfer_FullMethodName            = "/pb.TransferService/TrashedTransfer"
+	TransferService_RestoreTransfer_FullMethodName            = "/pb.TransferService/RestoreTransfer"
+	TransferService_DeleteTransferPermanent_FullMethodName    = "/pb.TransferService/DeleteTransferPermanent"
 )
 
 // TransferServiceClient is the client API for TransferService service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
 type TransferServiceClient interface {
-	GetTransfers(ctx context.Context, in *empty.Empty, opts ...grpc.CallOption) (*TransfersResponse, error)
-	GetTransfer(ctx context.Context, in *TransferRequest, opts ...grpc.CallOption) (*TransferResponse, error)
-	GetTransferByUsers(ctx context.Context, in *TransferRequest, opts ...grpc.CallOption) (*TransfersResponse, error)
-	GetTransferByUserId(ctx context.Context, in *TransferRequest, opts ...grpc.CallOption) (*TransferResponse, error)
-	CreateTransfer(ctx context.Context, in *CreateTransferRequest, opts ...grpc.CallOption) (*TransferResponse, error)
-	UpdateTransfer(ctx context.Context, in *UpdateTransferRequest, opts ...grpc.CallOption) (*TransferResponse, error)
-	DeleteTransfer(ctx context.Context, in *TransferRequest, opts ...grpc.CallOption) (*DeleteTransferResponse, error)
+	FindAllTransfer(ctx context.Context, in *FindAllTransferRequest, opts ...grpc.CallOption) (*ApiResponsePaginationTransfer, error)
+	FindByIdTransfer(ctx context.Context, in *FindByIdTransferRequest, opts ...grpc.CallOption) (*ApiResponseTransfer, error)
+	FindTransferByTransferFrom(ctx context.Context, in *FindTransferByTransferFromRequest, opts ...grpc.CallOption) (*ApiResponseTransfers, error)
+	FindTransferByTransferTo(ctx context.Context, in *FindTransferByTransferToRequest, opts ...grpc.CallOption) (*ApiResponseTransfers, error)
+	FindByActiveTransfer(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*ApiResponseTransfers, error)
+	FindByTrashedTransfer(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*ApiResponseTransfers, error)
+	CreateTransfer(ctx context.Context, in *CreateTransferRequest, opts ...grpc.CallOption) (*ApiResponseTransfer, error)
+	UpdateTransfer(ctx context.Context, in *UpdateTransferRequest, opts ...grpc.CallOption) (*ApiResponseTransfer, error)
+	TrashedTransfer(ctx context.Context, in *FindByIdTransferRequest, opts ...grpc.CallOption) (*ApiResponseTransfer, error)
+	RestoreTransfer(ctx context.Context, in *FindByIdTransferRequest, opts ...grpc.CallOption) (*ApiResponseTransfer, error)
+	DeleteTransferPermanent(ctx context.Context, in *FindByIdTransferRequest, opts ...grpc.CallOption) (*ApiResponseTransferDelete, error)
 }
 
 type transferServiceClient struct {
@@ -50,63 +58,110 @@ func NewTransferServiceClient(cc grpc.ClientConnInterface) TransferServiceClient
 	return &transferServiceClient{cc}
 }
 
-func (c *transferServiceClient) GetTransfers(ctx context.Context, in *empty.Empty, opts ...grpc.CallOption) (*TransfersResponse, error) {
-	out := new(TransfersResponse)
-	err := c.cc.Invoke(ctx, TransferService_GetTransfers_FullMethodName, in, out, opts...)
+func (c *transferServiceClient) FindAllTransfer(ctx context.Context, in *FindAllTransferRequest, opts ...grpc.CallOption) (*ApiResponsePaginationTransfer, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponsePaginationTransfer)
+	err := c.cc.Invoke(ctx, TransferService_FindAllTransfer_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *transferServiceClient) GetTransfer(ctx context.Context, in *TransferRequest, opts ...grpc.CallOption) (*TransferResponse, error) {
-	out := new(TransferResponse)
-	err := c.cc.Invoke(ctx, TransferService_GetTransfer_FullMethodName, in, out, opts...)
+func (c *transferServiceClient) FindByIdTransfer(ctx context.Context, in *FindByIdTransferRequest, opts ...grpc.CallOption) (*ApiResponseTransfer, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTransfer)
+	err := c.cc.Invoke(ctx, TransferService_FindByIdTransfer_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *transferServiceClient) GetTransferByUsers(ctx context.Context, in *TransferRequest, opts ...grpc.CallOption) (*TransfersResponse, error) {
-	out := new(TransfersResponse)
-	err := c.cc.Invoke(ctx, TransferService_GetTransferByUsers_FullMethodName, in, out, opts...)
+func (c *transferServiceClient) FindTransferByTransferFrom(ctx context.Context, in *FindTransferByTransferFromRequest, opts ...grpc.CallOption) (*ApiResponseTransfers, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTransfers)
+	err := c.cc.Invoke(ctx, TransferService_FindTransferByTransferFrom_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *transferServiceClient) GetTransferByUserId(ctx context.Context, in *TransferRequest, opts ...grpc.CallOption) (*TransferResponse, error) {
-	out := new(TransferResponse)
-	err := c.cc.Invoke(ctx, TransferService_GetTransferByUserId_FullMethodName, in, out, opts...)
+func (c *transferServiceClient) FindTransferByTransferTo(ctx context.Context, in *FindTransferByTransferToRequest, opts ...grpc.CallOption) (*ApiResponseTransfers, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTransfers)
+	err := c.cc.Invoke(ctx, TransferService_FindTransferByTransferTo_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *transferServiceClient) CreateTransfer(ctx context.Context, in *CreateTransferRequest, opts ...grpc.CallOption) (*TransferResponse, error) {
-	out := new(TransferResponse)
-	err := c.cc.Invoke(ctx, TransferService_CreateTransfer_FullMethodName, in, out, opts...)
+func (c *transferServiceClient) FindByActiveTransfer(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*ApiResponseTransfers, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTransfers)
+	err := c.cc.Invoke(ctx, TransferService_FindByActiveTransfer_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *transferServiceClient) UpdateTransfer(ctx context.Context, in *UpdateTransferRequest, opts ...grpc.CallOption) (*TransferResponse, error) {
-	out := new(TransferResponse)
-	err := c.cc.Invoke(ctx, TransferService_UpdateTransfer_FullMethodName, in, out, opts...)
+func (c *transferServiceClient) FindByTrashedTransfer(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*ApiResponseTransfers, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTransfers)
+	err := c.cc.Invoke(ctx, TransferService_FindByTrashedTransfer_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *transferServiceClient) DeleteTransfer(ctx context.Context, in *TransferRequest, opts ...grpc.CallOption) (*DeleteTransferResponse, error) {
-	out := new(DeleteTransferResponse)
-	err := c.cc.Invoke(ctx, TransferService_DeleteTransfer_FullMethodName, in, out, opts...)
+func (c *transferServiceClient) CreateTransfer(ctx context.Context, in *CreateTransferRequest, opts ...grpc.CallOption) (*ApiResponseTransfer, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTransfer)
+	err := c.cc.Invoke(ctx, TransferService_CreateTransfer_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *transferServiceClient) UpdateTransfer(ctx context.Context, in *UpdateTransferRequest, opts ...grpc.CallOption) (*ApiResponseTransfer, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTransfer)
+	err := c.cc.Invoke(ctx, TransferService_UpdateTransfer_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *transferServiceClient) TrashedTransfer(ctx context.Context, in *FindByIdTransferRequest, opts ...grpc.CallOption) (*ApiResponseTransfer, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTransfer)
+	err := c.cc.Invoke(ctx, TransferService_TrashedTransfer_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *transferServiceClient) RestoreTransfer(ctx context.Context, in *FindByIdTransferRequest, opts ...grpc.CallOption) (*ApiResponseTransfer, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTransfer)
+	err := c.cc.Invoke(ctx, TransferService_RestoreTransfer_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *transferServiceClient) DeleteTransferPermanent(ctx context.Context, in *FindByIdTransferRequest, opts ...grpc.CallOption) (*ApiResponseTransferDelete, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ApiResponseTransferDelete)
+	err := c.cc.Invoke(ctx, TransferService_DeleteTransferPermanent_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -115,44 +170,64 @@ func (c *transferServiceClient) DeleteTransfer(ctx context.Context, in *Transfer
 
 // TransferServiceServer is the server API for TransferService service.
 // All implementations must embed UnimplementedTransferServiceServer
-// for forward compatibility
+// for forward compatibility.
 type TransferServiceServer interface {
-	GetTransfers(context.Context, *empty.Empty) (*TransfersResponse, error)
-	GetTransfer(context.Context, *TransferRequest) (*TransferResponse, error)
-	GetTransferByUsers(context.Context, *TransferRequest) (*TransfersResponse, error)
-	GetTransferByUserId(context.Context, *TransferRequest) (*TransferResponse, error)
-	CreateTransfer(context.Context, *CreateTransferRequest) (*TransferResponse, error)
-	UpdateTransfer(context.Context, *UpdateTransferRequest) (*TransferResponse, error)
-	DeleteTransfer(context.Context, *TransferRequest) (*DeleteTransferResponse, error)
+	FindAllTransfer(context.Context, *FindAllTransferRequest) (*ApiResponsePaginationTransfer, error)
+	FindByIdTransfer(context.Context, *FindByIdTransferRequest) (*ApiResponseTransfer, error)
+	FindTransferByTransferFrom(context.Context, *FindTransferByTransferFromRequest) (*ApiResponseTransfers, error)
+	FindTransferByTransferTo(context.Context, *FindTransferByTransferToRequest) (*ApiResponseTransfers, error)
+	FindByActiveTransfer(context.Context, *emptypb.Empty) (*ApiResponseTransfers, error)
+	FindByTrashedTransfer(context.Context, *emptypb.Empty) (*ApiResponseTransfers, error)
+	CreateTransfer(context.Context, *CreateTransferRequest) (*ApiResponseTransfer, error)
+	UpdateTransfer(context.Context, *UpdateTransferRequest) (*ApiResponseTransfer, error)
+	TrashedTransfer(context.Context, *FindByIdTransferRequest) (*ApiResponseTransfer, error)
+	RestoreTransfer(context.Context, *FindByIdTransferRequest) (*ApiResponseTransfer, error)
+	DeleteTransferPermanent(context.Context, *FindByIdTransferRequest) (*ApiResponseTransferDelete, error)
 	mustEmbedUnimplementedTransferServiceServer()
 }
 
-// UnimplementedTransferServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedTransferServiceServer struct {
-}
+// UnimplementedTransferServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedTransferServiceServer struct{}
 
-func (UnimplementedTransferServiceServer) GetTransfers(context.Context, *empty.Empty) (*TransfersResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method GetTransfers not implemented")
+func (UnimplementedTransferServiceServer) FindAllTransfer(context.Context, *FindAllTransferRequest) (*ApiResponsePaginationTransfer, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method FindAllTransfer not implemented")
 }
-func (UnimplementedTransferServiceServer) GetTransfer(context.Context, *TransferRequest) (*TransferResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method GetTransfer not implemented")
+func (UnimplementedTransferServiceServer) FindByIdTransfer(context.Context, *FindByIdTransferRequest) (*ApiResponseTransfer, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method FindByIdTransfer not implemented")
 }
-func (UnimplementedTransferServiceServer) GetTransferByUsers(context.Context, *TransferRequest) (*TransfersResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method GetTransferByUsers not implemented")
+func (UnimplementedTransferServiceServer) FindTransferByTransferFrom(context.Context, *FindTransferByTransferFromRequest) (*ApiResponseTransfers, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method FindTransferByTransferFrom not implemented")
 }
-func (UnimplementedTransferServiceServer) GetTransferByUserId(context.Context, *TransferRequest) (*TransferResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method GetTransferByUserId not implemented")
+func (UnimplementedTransferServiceServer) FindTransferByTransferTo(context.Context, *FindTransferByTransferToRequest) (*ApiResponseTransfers, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method FindTransferByTransferTo not implemented")
 }
-func (UnimplementedTransferServiceServer) CreateTransfer(context.Context, *CreateTransferRequest) (*TransferResponse, error) {
+func (UnimplementedTransferServiceServer) FindByActiveTransfer(context.Context, *emptypb.Empty) (*ApiResponseTransfers, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method FindByActiveTransfer not implemented")
+}
+func (UnimplementedTransferServiceServer) FindByTrashedTransfer(context.Context, *emptypb.Empty) (*ApiResponseTransfers, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method FindByTrashedTransfer not implemented")
+}
+func (UnimplementedTransferServiceServer) CreateTransfer(context.Context, *CreateTransferRequest) (*ApiResponseTransfer, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method CreateTransfer not implemented")
 }
-func (UnimplementedTransferServiceServer) UpdateTransfer(context.Context, *UpdateTransferRequest) (*TransferResponse, error) {
+func (UnimplementedTransferServiceServer) UpdateTransfer(context.Context, *UpdateTransferRequest) (*ApiResponseTransfer, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateTransfer not implemented")
 }
-func (UnimplementedTransferServiceServer) DeleteTransfer(context.Context, *TransferRequest) (*DeleteTransferResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method DeleteTransfer not implemented")
+func (UnimplementedTransferServiceServer) TrashedTransfer(context.Context, *FindByIdTransferRequest) (*ApiResponseTransfer, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method TrashedTransfer not implemented")
+}
+func (UnimplementedTransferServiceServer) RestoreTransfer(context.Context, *FindByIdTransferRequest) (*ApiResponseTransfer, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method RestoreTransfer not implemented")
+}
+func (UnimplementedTransferServiceServer) DeleteTransferPermanent(context.Context, *FindByIdTransferRequest) (*ApiResponseTransferDelete, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method DeleteTransferPermanent not implemented")
 }
 func (UnimplementedTransferServiceServer) mustEmbedUnimplementedTransferServiceServer() {}
+func (UnimplementedTransferServiceServer) testEmbeddedByValue()                         {}
 
 // UnsafeTransferServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to TransferServiceServer will
@@ -162,77 +237,120 @@ type UnsafeTransferServiceServer interface {
 }
 
 func RegisterTransferServiceServer(s grpc.ServiceRegistrar, srv TransferServiceServer) {
+	// If the following call pancis, it indicates UnimplementedTransferServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&TransferService_ServiceDesc, srv)
 }
 
-func _TransferService_GetTransfers_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(empty.Empty)
+func _TransferService_FindAllTransfer_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindAllTransferRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(TransferServiceServer).GetTransfers(ctx, in)
+		return srv.(TransferServiceServer).FindAllTransfer(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: TransferService_GetTransfers_FullMethodName,
+		FullMethod: TransferService_FindAllTransfer_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(TransferServiceServer).GetTransfers(ctx, req.(*empty.Empty))
+		return srv.(TransferServiceServer).FindAllTransfer(ctx, req.(*FindAllTransferRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
-func _TransferService_GetTransfer_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(TransferRequest)
+func _TransferService_FindByIdTransfer_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindByIdTransferRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(TransferServiceServer).GetTransfer(ctx, in)
+		return srv.(TransferServiceServer).FindByIdTransfer(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: TransferService_GetTransfer_FullMethodName,
+		FullMethod: TransferService_FindByIdTransfer_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(TransferServiceServer).GetTransfer(ctx, req.(*TransferRequest))
+		return srv.(TransferServiceServer).FindByIdTransfer(ctx, req.(*FindByIdTransferRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
-func _TransferService_GetTransferByUsers_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(TransferRequest)
+func _TransferService_FindTransferByTransferFrom_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindTransferByTransferFromRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(TransferServiceServer).GetTransferByUsers(ctx, in)
+		return srv.(TransferServiceServer).FindTransferByTransferFrom(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: TransferService_GetTransferByUsers_FullMethodName,
+		FullMethod: TransferService_FindTransferByTransferFrom_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(TransferServiceServer).GetTransferByUsers(ctx, req.(*TransferRequest))
+		return srv.(TransferServiceServer).FindTransferByTransferFrom(ctx, req.(*FindTransferByTransferFromRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
 
-func _TransferService_GetTransferByUserId_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(TransferRequest)
+func _TransferService_FindTransferByTransferTo_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindTransferByTransferToRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(TransferServiceServer).GetTransferByUserId(ctx, in)
+		return srv.(TransferServiceServer).FindTransferByTransferTo(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: TransferService_GetTransferByUserId_FullMethodName,
+		FullMethod: TransferService_FindTransferByTransferTo_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(TransferServiceServer).GetTransferByUserId(ctx, req.(*TransferRequest))
+		return srv.(TransferServiceServer).FindTransferByTransferTo(ctx, req.(*FindTransferByTransferToRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _TransferService_FindByActiveTransfer_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(emptypb.Empty)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(TransferServiceServer).FindByActiveTransfer(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: TransferService_FindByActiveTransfer_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(TransferServiceServer).FindByActiveTransfer(ctx, req.(*emptypb.Empty))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _TransferService_FindByTrashedTransfer_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(emptypb.Empty)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(TransferServiceServer).FindByTrashedTransfer(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: TransferService_FindByTrashedTransfer_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(TransferServiceServer).FindByTrashedTransfer(ctx, req.(*emptypb.Empty))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -273,20 +391,56 @@ func _TransferService_UpdateTransfer_Handler(srv interface{}, ctx context.Contex
 	return interceptor(ctx, in, info, handler)
 }
 
-func _TransferService_DeleteTransfer_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(TransferRequest)
+func _TransferService_TrashedTransfer_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindByIdTransferRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(TransferServiceServer).DeleteTransfer(ctx, in)
+		return srv.(TransferServiceServer).TrashedTransfer(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: TransferService_DeleteTransfer_FullMethodName,
+		FullMethod: TransferService_TrashedTransfer_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(TransferServiceServer).DeleteTransfer(ctx, req.(*TransferRequest))
+		return srv.(TransferServiceServer).TrashedTransfer(ctx, req.(*FindByIdTransferRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _TransferService_RestoreTransfer_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindByIdTransferRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(TransferServiceServer).RestoreTransfer(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: TransferService_RestoreTransfer_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(TransferServiceServer).RestoreTransfer(ctx, req.(*FindByIdTransferRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _TransferService_DeleteTransferPermanent_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(FindByIdTransferRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(TransferServiceServer).DeleteTransferPermanent(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: TransferService_DeleteTransferPermanent_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(TransferServiceServer).DeleteTransferPermanent(ctx, req.(*FindByIdTransferRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -299,20 +453,28 @@ var TransferService_ServiceDesc = grpc.ServiceDesc{
 	HandlerType: (*TransferServiceServer)(nil),
 	Methods: []grpc.MethodDesc{
 		{
-			MethodName: "GetTransfers",
-			Handler:    _TransferService_GetTransfers_Handler,
+			MethodName: "FindAllTransfer",
+			Handler:    _TransferService_FindAllTransfer_Handler,
 		},
 		{
-			MethodName: "GetTransfer",
-			Handler:    _TransferService_GetTransfer_Handler,
+			MethodName: "FindByIdTransfer",
+			Handler:    _TransferService_FindByIdTransfer_Handler,
 		},
 		{
-			MethodName: "GetTransferByUsers",
-			Handler:    _TransferService_GetTransferByUsers_Handler,
+			MethodName: "FindTransferByTransferFrom",
+			Handler:    _TransferService_FindTransferByTransferFrom_Handler,
 		},
 		{
-			MethodName: "GetTransferByUserId",
-			Handler:    _TransferService_GetTransferByUserId_Handler,
+			MethodName: "FindTransferByTransferTo",
+			Handler:    _TransferService_FindTransferByTransferTo_Handler,
+		},
+		{
+			MethodName: "FindByActiveTransfer",
+			Handler:    _TransferService_FindByActiveTransfer_Handler,
+		},
+		{
+			MethodName: "FindByTrashedTransfer",
+			Handler:    _TransferService_FindByTrashedTransfer_Handler,
 		},
 		{
 			MethodName: "CreateTransfer",
@@ -323,8 +485,16 @@ var TransferService_ServiceDesc = grpc.ServiceDesc{
 			Handler:    _TransferService_UpdateTransfer_Handler,
 		},
 		{
-			MethodName: "DeleteTransfer",
-			Handler:    _TransferService_DeleteTransfer_Handler,
+			MethodName: "TrashedTransfer",
+			Handler:    _TransferService_TrashedTransfer_Handler,
+		},
+		{
+			MethodName: "RestoreTransfer",
+			Handler:    _TransferService_RestoreTransfer_Handler,
+		},
+		{
+			MethodName: "DeleteTransferPermanent",
+			Handler:    _TransferService_DeleteTransferPermanent_Handler,
 		},
 	},
 	Streams:  []grpc.StreamDesc{},
