@@ -1,6 +1,6 @@
 // Code generated by sqlc. DO NOT EDIT.
 // versions:
-//   sqlc v1.25.0
+//   sqlc v1.27.0
 
 package db
 
@@ -9,53 +9,93 @@ import (
 	"time"
 )
 
+type Card struct {
+	CardID       int32        `json:"card_id"`
+	UserID       int32        `json:"user_id"`
+	CardNumber   string       `json:"card_number"`
+	CardType     string       `json:"card_type"`
+	ExpireDate   time.Time    `json:"expire_date"`
+	Cvv          string       `json:"cvv"`
+	CardProvider string       `json:"card_provider"`
+	CreatedAt    sql.NullTime `json:"created_at"`
+	UpdatedAt    sql.NullTime `json:"updated_at"`
+	DeletedAt    sql.NullTime `json:"deleted_at"`
+}
+
+type Merchant struct {
+	MerchantID int32        `json:"merchant_id"`
+	Name       string       `json:"name"`
+	ApiKey     string       `json:"api_key"`
+	UserID     int32        `json:"user_id"`
+	Status     string       `json:"status"`
+	CreatedAt  sql.NullTime `json:"created_at"`
+	UpdatedAt  sql.NullTime `json:"updated_at"`
+	DeletedAt  sql.NullTime `json:"deleted_at"`
+}
+
 type Saldo struct {
 	SaldoID        int32         `json:"saldo_id"`
-	UserID         int32         `json:"user_id"`
+	CardNumber     string        `json:"card_number"`
 	TotalBalance   int32         `json:"total_balance"`
 	WithdrawAmount sql.NullInt32 `json:"withdraw_amount"`
 	WithdrawTime   sql.NullTime  `json:"withdraw_time"`
 	CreatedAt      sql.NullTime  `json:"created_at"`
 	UpdatedAt      sql.NullTime  `json:"updated_at"`
+	DeletedAt      sql.NullTime  `json:"deleted_at"`
 }
 
 type Topup struct {
 	TopupID     int32        `json:"topup_id"`
-	UserID      int32        `json:"user_id"`
+	CardNumber  string       `json:"card_number"`
 	TopupNo     string       `json:"topup_no"`
 	TopupAmount int32        `json:"topup_amount"`
 	TopupMethod string       `json:"topup_method"`
 	TopupTime   time.Time    `json:"topup_time"`
 	CreatedAt   sql.NullTime `json:"created_at"`
 	UpdatedAt   sql.NullTime `json:"updated_at"`
+	DeletedAt   sql.NullTime `json:"deleted_at"`
+}
+
+type Transaction struct {
+	TransactionID   int32        `json:"transaction_id"`
+	CardNumber      string       `json:"card_number"`
+	Amount          int32        `json:"amount"`
+	PaymentMethod   string       `json:"payment_method"`
+	MerchantID      int32        `json:"merchant_id"`
+	TransactionTime time.Time    `json:"transaction_time"`
+	CreatedAt       sql.NullTime `json:"created_at"`
+	UpdatedAt       sql.NullTime `json:"updated_at"`
+	DeletedAt       sql.NullTime `json:"deleted_at"`
 }
 
 type Transfer struct {
 	TransferID     int32        `json:"transfer_id"`
-	TransferFrom   int32        `json:"transfer_from"`
-	TransferTo     int32        `json:"transfer_to"`
+	TransferFrom   string       `json:"transfer_from"`
+	TransferTo     string       `json:"transfer_to"`
 	TransferAmount int32        `json:"transfer_amount"`
 	TransferTime   time.Time    `json:"transfer_time"`
 	CreatedAt      sql.NullTime `json:"created_at"`
 	UpdatedAt      sql.NullTime `json:"updated_at"`
+	DeletedAt      sql.NullTime `json:"deleted_at"`
 }
 
 type User struct {
-	UserID      int32        `json:"user_id"`
-	Firstname   string       `json:"firstname"`
-	Lastname    string       `json:"lastname"`
-	Email       string       `json:"email"`
-	Password    string       `json:"password"`
-	NocTransfer string       `json:"noc_transfer"`
-	CreatedAt   sql.NullTime `json:"created_at"`
-	UpdatedAt   sql.NullTime `json:"updated_at"`
+	UserID    int32        `json:"user_id"`
+	Firstname string       `json:"firstname"`
+	Lastname  string       `json:"lastname"`
+	Email     string       `json:"email"`
+	Password  string       `json:"password"`
+	CreatedAt sql.NullTime `json:"created_at"`
+	UpdatedAt sql.NullTime `json:"updated_at"`
+	DeletedAt sql.NullTime `json:"deleted_at"`
 }
 
 type Withdraw struct {
 	WithdrawID     int32        `json:"withdraw_id"`
-	UserID         int32        `json:"user_id"`
+	CardNumber     string       `json:"card_number"`
 	WithdrawAmount int32        `json:"withdraw_amount"`
 	WithdrawTime   time.Time    `json:"withdraw_time"`
 	CreatedAt      sql.NullTime `json:"created_at"`
 	UpdatedAt      sql.NullTime `json:"updated_at"`
+	DeletedAt      sql.NullTime `json:"deleted_at"`
 }
