@@ -1,6 +1,6 @@
 // Code generated by sqlc. DO NOT EDIT.
 // versions:
-//   sqlc v1.25.0
+//   sqlc v1.27.0
 // source: saldo.sql
 
 package db
@@ -11,51 +11,60 @@ import (
 )
 
 const createSaldo = `-- name: CreateSaldo :one
-INSERT INTO saldo (user_id, total_balance, withdraw_amount, withdraw_time)VALUES ($1, $2, $3, $4)RETURNING saldo_id, user_id, total_balance, withdraw_amount, withdraw_time, created_at, updated_at
+INSERT INTO
+    saldos (
+        card_number,
+        total_balance,
+        created_at,
+        updated_at
+    )
+VALUES (
+        $1,
+        $2,
+        current_timestamp,
+        current_timestamp
+    ) RETURNING saldo_id, card_number, total_balance, withdraw_amount, withdraw_time, created_at, updated_at, deleted_at
 `
 
 type CreateSaldoParams struct {
-	UserID         int32         `json:"user_id"`
-	TotalBalance   int32         `json:"total_balance"`
-	WithdrawAmount sql.NullInt32 `json:"withdraw_amount"`
-	WithdrawTime   sql.NullTime  `json:"withdraw_time"`
+	CardNumber   string `json:"card_number"`
+	TotalBalance int32  `json:"total_balance"`
 }
 
+// Create Saldo
 func (q *Queries) CreateSaldo(ctx context.Context, arg CreateSaldoParams) (*Saldo, error) {
-	row := q.db.QueryRowContext(ctx, createSaldo,
-		arg.UserID,
-		arg.TotalBalance,
-		arg.WithdrawAmount,
-		arg.WithdrawTime,
-	)
+	row := q.db.QueryRowContext(ctx, createSaldo, arg.CardNumber, arg.TotalBalance)
 	var i Saldo
 	err := row.Scan(
 		&i.SaldoID,
-		&i.UserID,
+		&i.CardNumber,
 		&i.TotalBalance,
 		&i.WithdrawAmount,
 		&i.WithdrawTime,
 		&i.CreatedAt,
 		&i.UpdatedAt,
+		&i.DeletedAt,
 	)
 	return &i, err
 }
 
-const deleteSaldo = `-- name: DeleteSaldo :exec
-DELETE FROM saldo WHERE user_id = $1 RETURNING saldo_id, user_id, total_balance, withdraw_amount, withdraw_time, created_at, updated_at
+const deleteSaldoPermanently = `-- name: DeleteSaldoPermanently :exec
+DELETE FROM saldos WHERE saldo_id = $1
 `
 
-func (q *Queries) DeleteSaldo(ctx context.Context, userID int32) error {
-	_, err := q.db.ExecContext(ctx, deleteSaldo, userID)
+// Delete Saldo Permanently
+func (q *Queries) DeleteSaldoPermanently(ctx context.Context, saldoID int32) error {
+	_, err := q.db.ExecContext(ctx, deleteSaldoPermanently, saldoID)
 	return err
 }
 
-const getAllSaldo = `-- name: GetAllSaldo :many
-SELECT saldo_id, user_id, total_balance, withdraw_amount, withdraw_time, created_at, updated_at FROM saldo
+const getActiveSaldos = `-- name: GetActiveSaldos :many
+SELECT saldo_id, card_number, total_balance, withdraw_amount, withdraw_time, created_at, updated_at, deleted_at FROM saldos WHERE deleted_at IS NULL ORDER BY saldo_id
 `
 
-func (q *Queries) GetAllSaldo(ctx context.Context) ([]*Saldo, error) {
-	rows, err := q.db.QueryContext(ctx, getAllSaldo)
+// Get All Active Saldos
+func (q *Queries) GetActiveSaldos(ctx context.Context) ([]*Saldo, error) {
+	rows, err := q.db.QueryContext(ctx, getActiveSaldos)
 	if err != nil {
 		return nil, err
 	}
@@ -65,12 +74,13 @@ func (q *Queries) GetAllSaldo(ctx context.Context) ([]*Saldo, error) {
 		var i Saldo
 		if err := rows.Scan(
 			&i.SaldoID,
-			&i.UserID,
+			&i.CardNumber,
 			&i.TotalBalance,
 			&i.WithdrawAmount,
 			&i.WithdrawTime,
 			&i.CreatedAt,
 			&i.UpdatedAt,
+			&i.DeletedAt,
 		); err != nil {
 			return nil, err
 		}
@@ -85,50 +95,66 @@ func (q *Queries) GetAllSaldo(ctx context.Context) ([]*Saldo, error) {
 	return items, nil
 }
 
-const getSaldoById = `-- name: GetSaldoById :one
-SELECT saldo_id, user_id, total_balance, withdraw_amount, withdraw_time, created_at, updated_at FROM saldo WHERE saldo_id = $1 LIMIT 1
+const getSaldoByCardNumber = `-- name: GetSaldoByCardNumber :one
+SELECT saldo_id, card_number, total_balance, withdraw_amount, withdraw_time, created_at, updated_at, deleted_at FROM saldos WHERE card_number = $1 AND deleted_at IS NULL
 `
 
-func (q *Queries) GetSaldoById(ctx context.Context, saldoID int32) (*Saldo, error) {
-	row := q.db.QueryRowContext(ctx, getSaldoById, saldoID)
+// Get Saldo by Card Number
+func (q *Queries) GetSaldoByCardNumber(ctx context.Context, cardNumber string) (*Saldo, error) {
+	row := q.db.QueryRowContext(ctx, getSaldoByCardNumber, cardNumber)
 	var i Saldo
 	err := row.Scan(
 		&i.SaldoID,
-		&i.UserID,
+		&i.CardNumber,
 		&i.TotalBalance,
 		&i.WithdrawAmount,
 		&i.WithdrawTime,
 		&i.CreatedAt,
 		&i.UpdatedAt,
+		&i.DeletedAt,
 	)
 	return &i, err
 }
 
-const getSaldoByUserId = `-- name: GetSaldoByUserId :one
-SELECT saldo_id, user_id, total_balance, withdraw_amount, withdraw_time, created_at, updated_at FROM saldo WHERE user_id = $1 LIMIT 1
+const getSaldoByID = `-- name: GetSaldoByID :one
+SELECT saldo_id, card_number, total_balance, withdraw_amount, withdraw_time, created_at, updated_at, deleted_at FROM saldos WHERE saldo_id = $1 AND deleted_at IS NULL
 `
 
-func (q *Queries) GetSaldoByUserId(ctx context.Context, userID int32) (*Saldo, error) {
-	row := q.db.QueryRowContext(ctx, getSaldoByUserId, userID)
+// Get Saldo by ID
+func (q *Queries) GetSaldoByID(ctx context.Context, saldoID int32) (*Saldo, error) {
+	row := q.db.QueryRowContext(ctx, getSaldoByID, saldoID)
 	var i Saldo
 	err := row.Scan(
 		&i.SaldoID,
-		&i.UserID,
+		&i.CardNumber,
 		&i.TotalBalance,
 		&i.WithdrawAmount,
 		&i.WithdrawTime,
 		&i.CreatedAt,
 		&i.UpdatedAt,
+		&i.DeletedAt,
 	)
 	return &i, err
 }
 
-const getSaldoByUsers = `-- name: GetSaldoByUsers :many
-SELECT saldo_id, user_id, total_balance, withdraw_amount, withdraw_time, created_at, updated_at FROM saldo WHERE user_id = $1
+const getSaldos = `-- name: GetSaldos :many
+SELECT saldo_id, card_number, total_balance, withdraw_amount, withdraw_time, created_at, updated_at, deleted_at
+FROM saldos
+WHERE deleted_at IS NULL
+  AND ($1::TEXT IS NULL OR card_number ILIKE '%' || $1 || '%')
+ORDER BY saldo_id
+LIMIT $2 OFFSET $3
 `
 
-func (q *Queries) GetSaldoByUsers(ctx context.Context, userID int32) ([]*Saldo, error) {
-	rows, err := q.db.QueryContext(ctx, getSaldoByUsers, userID)
+type GetSaldosParams struct {
+	Column1 string `json:"column_1"`
+	Limit   int32  `json:"limit"`
+	Offset  int32  `json:"offset"`
+}
+
+// Search Saldos with Pagination
+func (q *Queries) GetSaldos(ctx context.Context, arg GetSaldosParams) ([]*Saldo, error) {
+	rows, err := q.db.QueryContext(ctx, getSaldos, arg.Column1, arg.Limit, arg.Offset)
 	if err != nil {
 		return nil, err
 	}
@@ -138,12 +164,13 @@ func (q *Queries) GetSaldoByUsers(ctx context.Context, userID int32) ([]*Saldo, 
 		var i Saldo
 		if err := rows.Scan(
 			&i.SaldoID,
-			&i.UserID,
+			&i.CardNumber,
 			&i.TotalBalance,
 			&i.WithdrawAmount,
 			&i.WithdrawTime,
 			&i.CreatedAt,
 			&i.UpdatedAt,
+			&i.DeletedAt,
 		); err != nil {
 			return nil, err
 		}
@@ -158,57 +185,159 @@ func (q *Queries) GetSaldoByUsers(ctx context.Context, userID int32) ([]*Saldo, 
 	return items, nil
 }
 
-const updateSaldo = `-- name: UpdateSaldo :one
-UPDATE saldo SET total_balance = $1, withdraw_amount = $2, withdraw_time = $3 WHERE user_id = $4 RETURNING saldo_id, user_id, total_balance, withdraw_amount, withdraw_time, created_at, updated_at
+const getTrashedSaldoByID = `-- name: GetTrashedSaldoByID :one
+SELECT saldo_id, card_number, total_balance, withdraw_amount, withdraw_time, created_at, updated_at, deleted_at FROM saldos WHERE saldo_id = $1 AND deleted_at IS NOT NULL
+`
+
+// Get Trashed By Saldo ID
+func (q *Queries) GetTrashedSaldoByID(ctx context.Context, saldoID int32) (*Saldo, error) {
+	row := q.db.QueryRowContext(ctx, getTrashedSaldoByID, saldoID)
+	var i Saldo
+	err := row.Scan(
+		&i.SaldoID,
+		&i.CardNumber,
+		&i.TotalBalance,
+		&i.WithdrawAmount,
+		&i.WithdrawTime,
+		&i.CreatedAt,
+		&i.UpdatedAt,
+		&i.DeletedAt,
+	)
+	return &i, err
+}
+
+const getTrashedSaldos = `-- name: GetTrashedSaldos :many
+SELECT saldo_id, card_number, total_balance, withdraw_amount, withdraw_time, created_at, updated_at, deleted_at FROM saldos WHERE deleted_at IS NOT NULL ORDER BY saldo_id
+`
+
+// Get Trashed Saldos
+func (q *Queries) GetTrashedSaldos(ctx context.Context) ([]*Saldo, error) {
+	rows, err := q.db.QueryContext(ctx, getTrashedSaldos)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var items []*Saldo
+	for rows.Next() {
+		var i Saldo
+		if err := rows.Scan(
+			&i.SaldoID,
+			&i.CardNumber,
+			&i.TotalBalance,
+			&i.WithdrawAmount,
+			&i.WithdrawTime,
+			&i.CreatedAt,
+			&i.UpdatedAt,
+			&i.DeletedAt,
+		); err != nil {
+			return nil, err
+		}
+		items = append(items, &i)
+	}
+	if err := rows.Close(); err != nil {
+		return nil, err
+	}
+	if err := rows.Err(); err != nil {
+		return nil, err
+	}
+	return items, nil
+}
+
+const restoreSaldo = `-- name: RestoreSaldo :exec
+UPDATE saldos
+SET
+    deleted_at = NULL
+WHERE
+    saldo_id = $1
+    AND deleted_at IS NOT NULL
+`
+
+// Restore Trashed Saldo
+func (q *Queries) RestoreSaldo(ctx context.Context, saldoID int32) error {
+	_, err := q.db.ExecContext(ctx, restoreSaldo, saldoID)
+	return err
+}
+
+const trashSaldo = `-- name: TrashSaldo :exec
+UPDATE saldos
+SET
+    deleted_at = current_timestamp
+WHERE
+    saldo_id = $1
+    AND deleted_at IS NULL
+`
+
+// Trash Saldo
+func (q *Queries) TrashSaldo(ctx context.Context, saldoID int32) error {
+	_, err := q.db.ExecContext(ctx, trashSaldo, saldoID)
+	return err
+}
+
+const updateSaldo = `-- name: UpdateSaldo :exec
+UPDATE saldos
+SET
+    card_number = $2,
+    total_balance = $3,
+    updated_at = current_timestamp
+WHERE
+    saldo_id = $1
+    AND deleted_at IS NULL
 `
 
 type UpdateSaldoParams struct {
-	TotalBalance   int32         `json:"total_balance"`
-	WithdrawAmount sql.NullInt32 `json:"withdraw_amount"`
-	WithdrawTime   sql.NullTime  `json:"withdraw_time"`
-	UserID         int32         `json:"user_id"`
+	SaldoID      int32  `json:"saldo_id"`
+	CardNumber   string `json:"card_number"`
+	TotalBalance int32  `json:"total_balance"`
 }
 
-func (q *Queries) UpdateSaldo(ctx context.Context, arg UpdateSaldoParams) (*Saldo, error) {
-	row := q.db.QueryRowContext(ctx, updateSaldo,
-		arg.TotalBalance,
-		arg.WithdrawAmount,
-		arg.WithdrawTime,
-		arg.UserID,
-	)
-	var i Saldo
-	err := row.Scan(
-		&i.SaldoID,
-		&i.UserID,
-		&i.TotalBalance,
-		&i.WithdrawAmount,
-		&i.WithdrawTime,
-		&i.CreatedAt,
-		&i.UpdatedAt,
-	)
-	return &i, err
+// Update Saldo
+func (q *Queries) UpdateSaldo(ctx context.Context, arg UpdateSaldoParams) error {
+	_, err := q.db.ExecContext(ctx, updateSaldo, arg.SaldoID, arg.CardNumber, arg.TotalBalance)
+	return err
 }
 
-const updateSaldoBalance = `-- name: UpdateSaldoBalance :one
-UPDATE saldo SET total_balance = $1 WHERE user_id = $2 RETURNING saldo_id, user_id, total_balance, withdraw_amount, withdraw_time, created_at, updated_at
+const updateSaldoBalance = `-- name: UpdateSaldoBalance :exec
+UPDATE saldos
+SET
+    total_balance = $2,
+    updated_at = current_timestamp
+WHERE
+    card_number = $1
+    AND deleted_at IS NULL
 `
 
 type UpdateSaldoBalanceParams struct {
-	TotalBalance int32 `json:"total_balance"`
-	UserID       int32 `json:"user_id"`
+	CardNumber   string `json:"card_number"`
+	TotalBalance int32  `json:"total_balance"`
 }
 
-func (q *Queries) UpdateSaldoBalance(ctx context.Context, arg UpdateSaldoBalanceParams) (*Saldo, error) {
-	row := q.db.QueryRowContext(ctx, updateSaldoBalance, arg.TotalBalance, arg.UserID)
-	var i Saldo
-	err := row.Scan(
-		&i.SaldoID,
-		&i.UserID,
-		&i.TotalBalance,
-		&i.WithdrawAmount,
-		&i.WithdrawTime,
-		&i.CreatedAt,
-		&i.UpdatedAt,
-	)
-	return &i, err
+// Update Saldo Balance
+func (q *Queries) UpdateSaldoBalance(ctx context.Context, arg UpdateSaldoBalanceParams) error {
+	_, err := q.db.ExecContext(ctx, updateSaldoBalance, arg.CardNumber, arg.TotalBalance)
+	return err
+}
+
+const updateSaldoWithdraw = `-- name: UpdateSaldoWithdraw :exec
+UPDATE saldos
+SET
+    withdraw_amount = $2,
+    total_balance = total_balance - $2,
+    withdraw_time = $3,
+    updated_at = current_timestamp
+WHERE
+    card_number = $1
+    AND deleted_at IS NULL
+    AND total_balance >= $2
+`
+
+type UpdateSaldoWithdrawParams struct {
+	CardNumber     string        `json:"card_number"`
+	WithdrawAmount sql.NullInt32 `json:"withdraw_amount"`
+	WithdrawTime   sql.NullTime  `json:"withdraw_time"`
+}
+
+// Update Saldo Withdraw
+func (q *Queries) UpdateSaldoWithdraw(ctx context.Context, arg UpdateSaldoWithdrawParams) error {
+	_, err := q.db.ExecContext(ctx, updateSaldoWithdraw, arg.CardNumber, arg.WithdrawAmount, arg.WithdrawTime)
+	return err
 }
