@@ -1,6 +1,6 @@
 // Code generated by sqlc. DO NOT EDIT.
 // versions:
-//   sqlc v1.25.0
+//   sqlc v1.27.0
 // source: topup.sql
 
 package db
@@ -10,55 +10,106 @@ import (
 	"time"
 )
 
+const countAllTopups = `-- name: CountAllTopups :one
+SELECT COUNT(*) FROM topups WHERE deleted_at IS NULL
+`
+
+// Count All Topups
+func (q *Queries) CountAllTopups(ctx context.Context) (int64, error) {
+	row := q.db.QueryRowContext(ctx, countAllTopups)
+	var count int64
+	err := row.Scan(&count)
+	return count, err
+}
+
+const countTopupsByDate = `-- name: CountTopupsByDate :one
+SELECT COUNT(*)
+FROM topups
+WHERE deleted_at IS NULL
+  AND topup_time::DATE = $1::DATE
+`
+
+// Count Topups by Date
+func (q *Queries) CountTopupsByDate(ctx context.Context, dollar_1 time.Time) (int64, error) {
+	row := q.db.QueryRowContext(ctx, countTopupsByDate, dollar_1)
+	var count int64
+	err := row.Scan(&count)
+	return count, err
+}
+
 const createTopup = `-- name: CreateTopup :one
-INSERT INTO topups (user_id, topup_no, topup_amount, topup_method, topup_time) VALUES ($1, $2, $3, $4, $5) RETURNING topup_id, user_id, topup_no, topup_amount, topup_method, topup_time, created_at, updated_at
+INSERT INTO
+    topups (
+        card_number,
+        topup_no,
+        topup_amount,
+        topup_method,
+        topup_time,
+        created_at,
+        updated_at
+    )
+VALUES (
+        $1,
+        $2,
+        $3,
+        $4,
+        current_timestamp,
+        current_timestamp,
+        current_timestamp
+    ) RETURNING topup_id, card_number, topup_no, topup_amount, topup_method, topup_time, created_at, updated_at, deleted_at
 `
 
 type CreateTopupParams struct {
-	UserID      int32     `json:"user_id"`
-	TopupNo     string    `json:"topup_no"`
-	TopupAmount int32     `json:"topup_amount"`
-	TopupMethod string    `json:"topup_method"`
-	TopupTime   time.Time `json:"topup_time"`
+	CardNumber  string `json:"card_number"`
+	TopupNo     string `json:"topup_no"`
+	TopupAmount int32  `json:"topup_amount"`
+	TopupMethod string `json:"topup_method"`
 }
 
+// Create Topup
 func (q *Queries) CreateTopup(ctx context.Context, arg CreateTopupParams) (*Topup, error) {
 	row := q.db.QueryRowContext(ctx, createTopup,
-		arg.UserID,
+		arg.CardNumber,
 		arg.TopupNo,
 		arg.TopupAmount,
 		arg.TopupMethod,
-		arg.TopupTime,
 	)
 	var i Topup
 	err := row.Scan(
 		&i.TopupID,
-		&i.UserID,
+		&i.CardNumber,
 		&i.TopupNo,
 		&i.TopupAmount,
 		&i.TopupMethod,
 		&i.TopupTime,
 		&i.CreatedAt,
 		&i.UpdatedAt,
+		&i.DeletedAt,
 	)
 	return &i, err
 }
 
-const deleteTopup = `-- name: DeleteTopup :exec
-DELETE FROM topups WHERE user_id = $1 RETURNING topup_id, user_id, topup_no, topup_amount, topup_method, topup_time, created_at, updated_at
+const deleteTopupPermanently = `-- name: DeleteTopupPermanently :exec
+DELETE FROM topups WHERE topup_id = $1
 `
 
-func (q *Queries) DeleteTopup(ctx context.Context, userID int32) error {
-	_, err := q.db.ExecContext(ctx, deleteTopup, userID)
+// Delete Topup Permanently
+func (q *Queries) DeleteTopupPermanently(ctx context.Context, topupID int32) error {
+	_, err := q.db.ExecContext(ctx, deleteTopupPermanently, topupID)
 	return err
 }
 
-const getAllTopups = `-- name: GetAllTopups :many
-SELECT topup_id, user_id, topup_no, topup_amount, topup_method, topup_time, created_at, updated_at FROM topups
+const getActiveTopups = `-- name: GetActiveTopups :many
+SELECT topup_id, card_number, topup_no, topup_amount, topup_method, topup_time, created_at, updated_at, deleted_at
+FROM topups
+WHERE
+    deleted_at IS NULL
+ORDER BY topup_time DESC
 `
 
-func (q *Queries) GetAllTopups(ctx context.Context) ([]*Topup, error) {
-	rows, err := q.db.QueryContext(ctx, getAllTopups)
+// Get All Active Topups
+func (q *Queries) GetActiveTopups(ctx context.Context) ([]*Topup, error) {
+	rows, err := q.db.QueryContext(ctx, getActiveTopups)
 	if err != nil {
 		return nil, err
 	}
@@ -68,13 +119,14 @@ func (q *Queries) GetAllTopups(ctx context.Context) ([]*Topup, error) {
 		var i Topup
 		if err := rows.Scan(
 			&i.TopupID,
-			&i.UserID,
+			&i.CardNumber,
 			&i.TopupNo,
 			&i.TopupAmount,
 			&i.TopupMethod,
 			&i.TopupTime,
 			&i.CreatedAt,
 			&i.UpdatedAt,
+			&i.DeletedAt,
 		); err != nil {
 			return nil, err
 		}
@@ -89,52 +141,46 @@ func (q *Queries) GetAllTopups(ctx context.Context) ([]*Topup, error) {
 	return items, nil
 }
 
-const getTopupById = `-- name: GetTopupById :one
-SELECT topup_id, user_id, topup_no, topup_amount, topup_method, topup_time, created_at, updated_at FROM topups WHERE topup_id = $1
+const getTopupByID = `-- name: GetTopupByID :one
+SELECT topup_id, card_number, topup_no, topup_amount, topup_method, topup_time, created_at, updated_at, deleted_at FROM topups WHERE topup_id = $1 AND deleted_at IS NULL
 `
 
-func (q *Queries) GetTopupById(ctx context.Context, topupID int32) (*Topup, error) {
-	row := q.db.QueryRowContext(ctx, getTopupById, topupID)
+// Get Topup by ID
+func (q *Queries) GetTopupByID(ctx context.Context, topupID int32) (*Topup, error) {
+	row := q.db.QueryRowContext(ctx, getTopupByID, topupID)
 	var i Topup
 	err := row.Scan(
 		&i.TopupID,
-		&i.UserID,
+		&i.CardNumber,
 		&i.TopupNo,
 		&i.TopupAmount,
 		&i.TopupMethod,
 		&i.TopupTime,
 		&i.CreatedAt,
 		&i.UpdatedAt,
+		&i.DeletedAt,
 	)
 	return &i, err
 }
 
-const getTopupByUserId = `-- name: GetTopupByUserId :one
-SELECT topup_id, user_id, topup_no, topup_amount, topup_method, topup_time, created_at, updated_at FROM topups WHERE topup_id = $1 LIMIT 1
+const getTopups = `-- name: GetTopups :many
+SELECT topup_id, card_number, topup_no, topup_amount, topup_method, topup_time, created_at, updated_at, deleted_at
+FROM topups
+WHERE deleted_at IS NULL
+  AND ($1::TEXT IS NULL OR card_number ILIKE '%' || $1 || '%' OR topup_no ILIKE '%' || $1 || '%' OR topup_method ILIKE '%' || $1 || '%')
+ORDER BY topup_time DESC
+LIMIT $2 OFFSET $3
 `
 
-func (q *Queries) GetTopupByUserId(ctx context.Context, topupID int32) (*Topup, error) {
-	row := q.db.QueryRowContext(ctx, getTopupByUserId, topupID)
-	var i Topup
-	err := row.Scan(
-		&i.TopupID,
-		&i.UserID,
-		&i.TopupNo,
-		&i.TopupAmount,
-		&i.TopupMethod,
-		&i.TopupTime,
-		&i.CreatedAt,
-		&i.UpdatedAt,
-	)
-	return &i, err
+type GetTopupsParams struct {
+	Column1 string `json:"column_1"`
+	Limit   int32  `json:"limit"`
+	Offset  int32  `json:"offset"`
 }
 
-const getTopupByUsers = `-- name: GetTopupByUsers :many
-SELECT topup_id, user_id, topup_no, topup_amount, topup_method, topup_time, created_at, updated_at FROM topups WHERE user_id = $1
-`
-
-func (q *Queries) GetTopupByUsers(ctx context.Context, userID int32) ([]*Topup, error) {
-	rows, err := q.db.QueryContext(ctx, getTopupByUsers, userID)
+// Search Topups with Pagination
+func (q *Queries) GetTopups(ctx context.Context, arg GetTopupsParams) ([]*Topup, error) {
+	rows, err := q.db.QueryContext(ctx, getTopups, arg.Column1, arg.Limit, arg.Offset)
 	if err != nil {
 		return nil, err
 	}
@@ -144,13 +190,14 @@ func (q *Queries) GetTopupByUsers(ctx context.Context, userID int32) ([]*Topup, 
 		var i Topup
 		if err := rows.Scan(
 			&i.TopupID,
-			&i.UserID,
+			&i.CardNumber,
 			&i.TopupNo,
 			&i.TopupAmount,
 			&i.TopupMethod,
 			&i.TopupTime,
 			&i.CreatedAt,
 			&i.UpdatedAt,
+			&i.DeletedAt,
 		); err != nil {
 			return nil, err
 		}
@@ -165,34 +212,195 @@ func (q *Queries) GetTopupByUsers(ctx context.Context, userID int32) ([]*Topup, 
 	return items, nil
 }
 
-const updateTopup = `-- name: UpdateTopup :one
-UPDATE topups SET topup_amount = $1, topup_method = $2, topup_time = $3 WHERE topup_id = $4 RETURNING topup_id, user_id, topup_no, topup_amount, topup_method, topup_time, created_at, updated_at
+const getTopupsByCardNumber = `-- name: GetTopupsByCardNumber :many
+SELECT topup_id, card_number, topup_no, topup_amount, topup_method, topup_time, created_at, updated_at, deleted_at
+FROM topups
+WHERE
+    deleted_at IS NULL
+    AND card_number = $1
+ORDER BY topup_time DESC
+`
+
+// Get Topups by Card Number
+func (q *Queries) GetTopupsByCardNumber(ctx context.Context, cardNumber string) ([]*Topup, error) {
+	rows, err := q.db.QueryContext(ctx, getTopupsByCardNumber, cardNumber)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var items []*Topup
+	for rows.Next() {
+		var i Topup
+		if err := rows.Scan(
+			&i.TopupID,
+			&i.CardNumber,
+			&i.TopupNo,
+			&i.TopupAmount,
+			&i.TopupMethod,
+			&i.TopupTime,
+			&i.CreatedAt,
+			&i.UpdatedAt,
+			&i.DeletedAt,
+		); err != nil {
+			return nil, err
+		}
+		items = append(items, &i)
+	}
+	if err := rows.Close(); err != nil {
+		return nil, err
+	}
+	if err := rows.Err(); err != nil {
+		return nil, err
+	}
+	return items, nil
+}
+
+const getTrashedTopupByID = `-- name: GetTrashedTopupByID :one
+SELECT topup_id, card_number, topup_no, topup_amount, topup_method, topup_time, created_at, updated_at, deleted_at
+FROM topups
+WHERE
+    topup_id = $1
+    AND deleted_at IS NOT NULL
+`
+
+// Get Trashed By Topup ID
+func (q *Queries) GetTrashedTopupByID(ctx context.Context, topupID int32) (*Topup, error) {
+	row := q.db.QueryRowContext(ctx, getTrashedTopupByID, topupID)
+	var i Topup
+	err := row.Scan(
+		&i.TopupID,
+		&i.CardNumber,
+		&i.TopupNo,
+		&i.TopupAmount,
+		&i.TopupMethod,
+		&i.TopupTime,
+		&i.CreatedAt,
+		&i.UpdatedAt,
+		&i.DeletedAt,
+	)
+	return &i, err
+}
+
+const getTrashedTopups = `-- name: GetTrashedTopups :many
+SELECT topup_id, card_number, topup_no, topup_amount, topup_method, topup_time, created_at, updated_at, deleted_at
+FROM topups
+WHERE
+    deleted_at IS NOT NULL
+ORDER BY topup_time DESC
+`
+
+// Get Trashed Topups
+func (q *Queries) GetTrashedTopups(ctx context.Context) ([]*Topup, error) {
+	rows, err := q.db.QueryContext(ctx, getTrashedTopups)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var items []*Topup
+	for rows.Next() {
+		var i Topup
+		if err := rows.Scan(
+			&i.TopupID,
+			&i.CardNumber,
+			&i.TopupNo,
+			&i.TopupAmount,
+			&i.TopupMethod,
+			&i.TopupTime,
+			&i.CreatedAt,
+			&i.UpdatedAt,
+			&i.DeletedAt,
+		); err != nil {
+			return nil, err
+		}
+		items = append(items, &i)
+	}
+	if err := rows.Close(); err != nil {
+		return nil, err
+	}
+	if err := rows.Err(); err != nil {
+		return nil, err
+	}
+	return items, nil
+}
+
+const restoreTopup = `-- name: RestoreTopup :exec
+UPDATE topups
+SET
+    deleted_at = NULL
+WHERE
+    topup_id = $1
+    AND deleted_at IS NOT NULL
+`
+
+// Restore Trashed Topup
+func (q *Queries) RestoreTopup(ctx context.Context, topupID int32) error {
+	_, err := q.db.ExecContext(ctx, restoreTopup, topupID)
+	return err
+}
+
+const trashTopup = `-- name: TrashTopup :exec
+UPDATE topups
+SET
+    deleted_at = current_timestamp
+WHERE
+    topup_id = $1
+    AND deleted_at IS NULL
+`
+
+// Trash Topup
+func (q *Queries) TrashTopup(ctx context.Context, topupID int32) error {
+	_, err := q.db.ExecContext(ctx, trashTopup, topupID)
+	return err
+}
+
+const updateTopup = `-- name: UpdateTopup :exec
+UPDATE topups
+SET
+    card_number = $2,
+    topup_amount = $3,
+    topup_method = $4,
+    topup_time = current_timestamp,
+    updated_at = current_timestamp
+WHERE
+    topup_id = $1
+    AND deleted_at IS NULL
 `
 
 type UpdateTopupParams struct {
-	TopupAmount int32     `json:"topup_amount"`
-	TopupMethod string    `json:"topup_method"`
-	TopupTime   time.Time `json:"topup_time"`
-	TopupID     int32     `json:"topup_id"`
+	TopupID     int32  `json:"topup_id"`
+	CardNumber  string `json:"card_number"`
+	TopupAmount int32  `json:"topup_amount"`
+	TopupMethod string `json:"topup_method"`
 }
 
-func (q *Queries) UpdateTopup(ctx context.Context, arg UpdateTopupParams) (*Topup, error) {
-	row := q.db.QueryRowContext(ctx, updateTopup,
+// Update Topup
+func (q *Queries) UpdateTopup(ctx context.Context, arg UpdateTopupParams) error {
+	_, err := q.db.ExecContext(ctx, updateTopup,
+		arg.TopupID,
+		arg.CardNumber,
 		arg.TopupAmount,
 		arg.TopupMethod,
-		arg.TopupTime,
-		arg.TopupID,
 	)
-	var i Topup
-	err := row.Scan(
-		&i.TopupID,
-		&i.UserID,
-		&i.TopupNo,
-		&i.TopupAmount,
-		&i.TopupMethod,
-		&i.TopupTime,
-		&i.CreatedAt,
-		&i.UpdatedAt,
-	)
-	return &i, err
+	return err
+}
+
+const updateTopupAmount = `-- name: UpdateTopupAmount :exec
+UPDATE topups
+SET
+    topup_amount = $2,
+    updated_at = current_timestamp
+WHERE
+    topup_id = $1
+    AND deleted_at IS NULL
+`
+
+type UpdateTopupAmountParams struct {
+	TopupID     int32 `json:"topup_id"`
+	TopupAmount int32 `json:"topup_amount"`
+}
+
+// Update Topup Amount
+func (q *Queries) UpdateTopupAmount(ctx context.Context, arg UpdateTopupAmountParams) error {
+	_, err := q.db.ExecContext(ctx, updateTopupAmount, arg.TopupID, arg.TopupAmount)
+	return err
 }
