@@ -1,33 +1,61 @@
 // Code generated by sqlc. DO NOT EDIT.
 // versions:
-//   sqlc v1.25.0
+//   sqlc v1.27.0
 // source: user.sql
 
 package db
 
 import (
 	"context"
+	"database/sql"
 )
 
+const countActiveUsers = `-- name: CountActiveUsers :one
+SELECT COUNT(*) FROM users WHERE deleted_at IS NULL
+`
+
+// Count All Active Users
+func (q *Queries) CountActiveUsers(ctx context.Context) (int64, error) {
+	row := q.db.QueryRowContext(ctx, countActiveUsers)
+	var count int64
+	err := row.Scan(&count)
+	return count, err
+}
+
 const createUser = `-- name: CreateUser :one
-INSERT INTO users (firstname, lastname,email, password, noc_transfer) VALUES ($1, $2, $3, $4, $5) RETURNING user_id, firstname, lastname, email, password, noc_transfer, created_at, updated_at
+INSERT INTO
+    users (
+        firstname,
+        lastname,
+        email,
+        password,
+        created_at,
+        updated_at
+    )
+VALUES (
+        $1,
+        $2,
+        $3,
+        $4,
+        current_timestamp,
+        current_timestamp
+    ) RETURNING user_id, firstname, lastname, email, password, created_at, updated_at, deleted_at
 `
 
 type CreateUserParams struct {
-	Firstname   string `json:"firstname"`
-	Lastname    string `json:"lastname"`
-	Email       string `json:"email"`
-	Password    string `json:"password"`
-	NocTransfer string `json:"noc_transfer"`
+	Firstname string `json:"firstname"`
+	Lastname  string `json:"lastname"`
+	Email     string `json:"email"`
+	Password  string `json:"password"`
 }
 
+// Create User
 func (q *Queries) CreateUser(ctx context.Context, arg CreateUserParams) (*User, error) {
 	row := q.db.QueryRowContext(ctx, createUser,
 		arg.Firstname,
 		arg.Lastname,
 		arg.Email,
 		arg.Password,
-		arg.NocTransfer,
 	)
 	var i User
 	err := row.Scan(
@@ -36,28 +64,34 @@ func (q *Queries) CreateUser(ctx context.Context, arg CreateUserParams) (*User, 
 		&i.Lastname,
 		&i.Email,
 		&i.Password,
-		&i.NocTransfer,
 		&i.CreatedAt,
 		&i.UpdatedAt,
+		&i.DeletedAt,
 	)
 	return &i, err
 }
 
-const deleteUser = `-- name: DeleteUser :exec
-DELETE FROM users WHERE user_id = $1 RETURNING user_id, firstname, lastname, email, password, noc_transfer, created_at, updated_at
+const deleteUserPermanently = `-- name: DeleteUserPermanently :exec
+DELETE FROM users WHERE user_id = $1
 `
 
-func (q *Queries) DeleteUser(ctx context.Context, userID int32) error {
-	_, err := q.db.ExecContext(ctx, deleteUser, userID)
+// Delete User Permanently
+func (q *Queries) DeleteUserPermanently(ctx context.Context, userID int32) error {
+	_, err := q.db.ExecContext(ctx, deleteUserPermanently, userID)
 	return err
 }
 
-const getAllUsers = `-- name: GetAllUsers :many
-SELECT user_id, firstname, lastname, email, password, noc_transfer, created_at, updated_at FROM users
+const getActiveUsers = `-- name: GetActiveUsers :many
+SELECT user_id, firstname, lastname, email, password, created_at, updated_at, deleted_at
+FROM users
+WHERE
+    deleted_at IS NULL
+ORDER BY created_at DESC
 `
 
-func (q *Queries) GetAllUsers(ctx context.Context) ([]*User, error) {
-	rows, err := q.db.QueryContext(ctx, getAllUsers)
+// Get All Active Users
+func (q *Queries) GetActiveUsers(ctx context.Context) ([]*User, error) {
+	rows, err := q.db.QueryContext(ctx, getActiveUsers)
 	if err != nil {
 		return nil, err
 	}
@@ -71,9 +105,75 @@ func (q *Queries) GetAllUsers(ctx context.Context) ([]*User, error) {
 			&i.Lastname,
 			&i.Email,
 			&i.Password,
-			&i.NocTransfer,
 			&i.CreatedAt,
 			&i.UpdatedAt,
+			&i.DeletedAt,
+		); err != nil {
+			return nil, err
+		}
+		items = append(items, &i)
+	}
+	if err := rows.Close(); err != nil {
+		return nil, err
+	}
+	if err := rows.Err(); err != nil {
+		return nil, err
+	}
+	return items, nil
+}
+
+const getTrashedUserByID = `-- name: GetTrashedUserByID :one
+SELECT user_id, firstname, lastname, email, password, created_at, updated_at, deleted_at
+FROM users
+WHERE
+    user_id = $1
+    AND deleted_at IS NOT NULL
+`
+
+// Get Trashed By User ID
+func (q *Queries) GetTrashedUserByID(ctx context.Context, userID int32) (*User, error) {
+	row := q.db.QueryRowContext(ctx, getTrashedUserByID, userID)
+	var i User
+	err := row.Scan(
+		&i.UserID,
+		&i.Firstname,
+		&i.Lastname,
+		&i.Email,
+		&i.Password,
+		&i.CreatedAt,
+		&i.UpdatedAt,
+		&i.DeletedAt,
+	)
+	return &i, err
+}
+
+const getTrashedUsers = `-- name: GetTrashedUsers :many
+SELECT user_id, firstname, lastname, email, password, created_at, updated_at, deleted_at
+FROM users
+WHERE
+    deleted_at IS NOT NULL
+ORDER BY created_at DESC
+`
+
+// Get Trashed Users
+func (q *Queries) GetTrashedUsers(ctx context.Context) ([]*User, error) {
+	rows, err := q.db.QueryContext(ctx, getTrashedUsers)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var items []*User
+	for rows.Next() {
+		var i User
+		if err := rows.Scan(
+			&i.UserID,
+			&i.Firstname,
+			&i.Lastname,
+			&i.Email,
+			&i.Password,
+			&i.CreatedAt,
+			&i.UpdatedAt,
+			&i.DeletedAt,
 		); err != nil {
 			return nil, err
 		}
@@ -89,9 +189,10 @@ func (q *Queries) GetAllUsers(ctx context.Context) ([]*User, error) {
 }
 
 const getUserByEmail = `-- name: GetUserByEmail :one
-SELECT user_id, firstname, lastname, email, password, noc_transfer, created_at, updated_at FROM users WHERE email = $1
+SELECT user_id, firstname, lastname, email, password, created_at, updated_at, deleted_at FROM users WHERE email = $1 AND deleted_at IS NULL
 `
 
+// Get User by Email
 func (q *Queries) GetUserByEmail(ctx context.Context, email string) (*User, error) {
 	row := q.db.QueryRowContext(ctx, getUserByEmail, email)
 	var i User
@@ -101,19 +202,20 @@ func (q *Queries) GetUserByEmail(ctx context.Context, email string) (*User, erro
 		&i.Lastname,
 		&i.Email,
 		&i.Password,
-		&i.NocTransfer,
 		&i.CreatedAt,
 		&i.UpdatedAt,
+		&i.DeletedAt,
 	)
 	return &i, err
 }
 
-const getUserById = `-- name: GetUserById :one
-SELECT user_id, firstname, lastname, email, password, noc_transfer, created_at, updated_at FROM users WHERE user_id = $1
+const getUserByID = `-- name: GetUserByID :one
+SELECT user_id, firstname, lastname, email, password, created_at, updated_at, deleted_at FROM users WHERE user_id = $1 AND deleted_at IS NULL
 `
 
-func (q *Queries) GetUserById(ctx context.Context, userID int32) (*User, error) {
-	row := q.db.QueryRowContext(ctx, getUserById, userID)
+// Get User by ID
+func (q *Queries) GetUserByID(ctx context.Context, userID int32) (*User, error) {
+	row := q.db.QueryRowContext(ctx, getUserByID, userID)
 	var i User
 	err := row.Scan(
 		&i.UserID,
@@ -121,45 +223,162 @@ func (q *Queries) GetUserById(ctx context.Context, userID int32) (*User, error) 
 		&i.Lastname,
 		&i.Email,
 		&i.Password,
-		&i.NocTransfer,
 		&i.CreatedAt,
 		&i.UpdatedAt,
+		&i.DeletedAt,
 	)
 	return &i, err
 }
 
-const updateUser = `-- name: UpdateUser :one
-UPDATE users SET noc_transfer = $1, firstname = $2, lastname = $3, email = $4, password = $5 WHERE user_id = $6 RETURNING user_id, firstname, lastname, email, password, noc_transfer, created_at, updated_at
+const restoreUser = `-- name: RestoreUser :exec
+UPDATE users
+SET
+    deleted_at = NULL
+WHERE
+    user_id = $1
+    AND deleted_at IS NOT NULL
+`
+
+// Restore Trashed User
+func (q *Queries) RestoreUser(ctx context.Context, userID int32) error {
+	_, err := q.db.ExecContext(ctx, restoreUser, userID)
+	return err
+}
+
+const searchUsers = `-- name: SearchUsers :many
+SELECT user_id, firstname, lastname, email, password, created_at, updated_at, deleted_at
+FROM users
+WHERE deleted_at IS NULL
+  AND ($1::TEXT IS NULL OR firstname ILIKE '%' || $1 || '%' OR lastname ILIKE '%' || $1 || '%' OR email ILIKE '%' || $1 || '%')
+ORDER BY created_at DESC
+LIMIT $2 OFFSET $3
+`
+
+type SearchUsersParams struct {
+	Column1 string `json:"column_1"`
+	Limit   int32  `json:"limit"`
+	Offset  int32  `json:"offset"`
+}
+
+// Search Users with Pagination
+func (q *Queries) SearchUsers(ctx context.Context, arg SearchUsersParams) ([]*User, error) {
+	rows, err := q.db.QueryContext(ctx, searchUsers, arg.Column1, arg.Limit, arg.Offset)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var items []*User
+	for rows.Next() {
+		var i User
+		if err := rows.Scan(
+			&i.UserID,
+			&i.Firstname,
+			&i.Lastname,
+			&i.Email,
+			&i.Password,
+			&i.CreatedAt,
+			&i.UpdatedAt,
+			&i.DeletedAt,
+		); err != nil {
+			return nil, err
+		}
+		items = append(items, &i)
+	}
+	if err := rows.Close(); err != nil {
+		return nil, err
+	}
+	if err := rows.Err(); err != nil {
+		return nil, err
+	}
+	return items, nil
+}
+
+const searchUsersByEmail = `-- name: SearchUsersByEmail :many
+SELECT user_id, firstname, lastname, email, password, created_at, updated_at, deleted_at
+FROM users
+WHERE
+    deleted_at IS NULL
+    AND email ILIKE '%' || $1 || '%'
+ORDER BY created_at DESC
+`
+
+// Search Users by Email
+func (q *Queries) SearchUsersByEmail(ctx context.Context, dollar_1 sql.NullString) ([]*User, error) {
+	rows, err := q.db.QueryContext(ctx, searchUsersByEmail, dollar_1)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var items []*User
+	for rows.Next() {
+		var i User
+		if err := rows.Scan(
+			&i.UserID,
+			&i.Firstname,
+			&i.Lastname,
+			&i.Email,
+			&i.Password,
+			&i.CreatedAt,
+			&i.UpdatedAt,
+			&i.DeletedAt,
+		); err != nil {
+			return nil, err
+		}
+		items = append(items, &i)
+	}
+	if err := rows.Close(); err != nil {
+		return nil, err
+	}
+	if err := rows.Err(); err != nil {
+		return nil, err
+	}
+	return items, nil
+}
+
+const trashUser = `-- name: TrashUser :exec
+UPDATE users
+SET
+    deleted_at = current_timestamp
+WHERE
+    user_id = $1
+    AND deleted_at IS NULL
+`
+
+// Trash User
+func (q *Queries) TrashUser(ctx context.Context, userID int32) error {
+	_, err := q.db.ExecContext(ctx, trashUser, userID)
+	return err
+}
+
+const updateUser = `-- name: UpdateUser :exec
+UPDATE users
+SET
+    firstname = $2,
+    lastname = $3,
+    email = $4,
+    password = $5,
+    updated_at = current_timestamp
+WHERE
+    user_id = $1
+    AND deleted_at IS NULL
 `
 
 type UpdateUserParams struct {
-	NocTransfer string `json:"noc_transfer"`
-	Firstname   string `json:"firstname"`
-	Lastname    string `json:"lastname"`
-	Email       string `json:"email"`
-	Password    string `json:"password"`
-	UserID      int32  `json:"user_id"`
+	UserID    int32  `json:"user_id"`
+	Firstname string `json:"firstname"`
+	Lastname  string `json:"lastname"`
+	Email     string `json:"email"`
+	Password  string `json:"password"`
 }
 
-func (q *Queries) UpdateUser(ctx context.Context, arg UpdateUserParams) (*User, error) {
-	row := q.db.QueryRowContext(ctx, updateUser,
-		arg.NocTransfer,
+// Update User
+func (q *Queries) UpdateUser(ctx context.Context, arg UpdateUserParams) error {
+	_, err := q.db.ExecContext(ctx, updateUser,
+		arg.UserID,
 		arg.Firstname,
 		arg.Lastname,
 		arg.Email,
 		arg.Password,
-		arg.UserID,
 	)
-	var i User
-	err := row.Scan(
-		&i.UserID,
-		&i.Firstname,
-		&i.Lastname,
-		&i.Email,
-		&i.Password,
-		&i.NocTransfer,
-		&i.CreatedAt,
-		&i.UpdatedAt,
-	)
-	return &i, err
+	return err
 }
