@@ -1,54 +1,91 @@
 // Code generated by sqlc. DO NOT EDIT.
 // versions:
-//   sqlc v1.25.0
+//   sqlc v1.27.0
 // source: withdraw.sql
 
 package db
 
 import (
 	"context"
+	"database/sql"
 	"time"
 )
 
+const countActiveWithdrawsByDate = `-- name: CountActiveWithdrawsByDate :one
+SELECT COUNT(*)
+FROM withdraws
+WHERE deleted_at IS NULL AND withdraw_time::DATE = $1
+`
+
+// Count Active Withdraws by Date
+func (q *Queries) CountActiveWithdrawsByDate(ctx context.Context, withdrawTime time.Time) (int64, error) {
+	row := q.db.QueryRowContext(ctx, countActiveWithdrawsByDate, withdrawTime)
+	var count int64
+	err := row.Scan(&count)
+	return count, err
+}
+
 const createWithdraw = `-- name: CreateWithdraw :one
-INSERT INTO withdraws (user_id, withdraw_amount, withdraw_time) VALUES ($1, $2, $3) RETURNING withdraw_id, user_id, withdraw_amount, withdraw_time, created_at, updated_at
+INSERT INTO
+    withdraws (
+        card_number,
+        withdraw_amount,
+        withdraw_time,
+        created_at,
+        updated_at
+    )
+VALUES (
+        $1,
+        $2,
+        $3,
+        current_timestamp,
+        current_timestamp
+    ) RETURNING withdraw_id, card_number, withdraw_amount, withdraw_time, created_at, updated_at, deleted_at
 `
 
 type CreateWithdrawParams struct {
-	UserID         int32     `json:"user_id"`
+	CardNumber     string    `json:"card_number"`
 	WithdrawAmount int32     `json:"withdraw_amount"`
 	WithdrawTime   time.Time `json:"withdraw_time"`
 }
 
+// Create Withdraw
 func (q *Queries) CreateWithdraw(ctx context.Context, arg CreateWithdrawParams) (*Withdraw, error) {
-	row := q.db.QueryRowContext(ctx, createWithdraw, arg.UserID, arg.WithdrawAmount, arg.WithdrawTime)
+	row := q.db.QueryRowContext(ctx, createWithdraw, arg.CardNumber, arg.WithdrawAmount, arg.WithdrawTime)
 	var i Withdraw
 	err := row.Scan(
 		&i.WithdrawID,
-		&i.UserID,
+		&i.CardNumber,
 		&i.WithdrawAmount,
 		&i.WithdrawTime,
 		&i.CreatedAt,
 		&i.UpdatedAt,
+		&i.DeletedAt,
 	)
 	return &i, err
 }
 
-const deleteWithdraw = `-- name: DeleteWithdraw :exec
-DELETE FROM withdraws WHERE user_id = $1 RETURNING withdraw_id, user_id, withdraw_amount, withdraw_time, created_at, updated_at
+const deleteWithdrawPermanently = `-- name: DeleteWithdrawPermanently :exec
+DELETE FROM withdraws WHERE withdraw_id = $1
 `
 
-func (q *Queries) DeleteWithdraw(ctx context.Context, userID int32) error {
-	_, err := q.db.ExecContext(ctx, deleteWithdraw, userID)
+// Delete Withdraw Permanently
+func (q *Queries) DeleteWithdrawPermanently(ctx context.Context, withdrawID int32) error {
+	_, err := q.db.ExecContext(ctx, deleteWithdrawPermanently, withdrawID)
 	return err
 }
 
-const getAllWithdraws = `-- name: GetAllWithdraws :many
-SELECT withdraw_id, user_id, withdraw_amount, withdraw_time, created_at, updated_at FROM withdraws
+const getActiveWithdraws = `-- name: GetActiveWithdraws :many
+SELECT withdraw_id, card_number, withdraw_amount, withdraw_time, created_at, updated_at, deleted_at
+FROM withdraws
+WHERE
+    deleted_at IS NULL
+ORDER BY withdraw_time DESC
 `
 
-func (q *Queries) GetAllWithdraws(ctx context.Context) ([]*Withdraw, error) {
-	rows, err := q.db.QueryContext(ctx, getAllWithdraws)
+// Get All Active Withdraws
+func (q *Queries) GetActiveWithdraws(ctx context.Context) ([]*Withdraw, error) {
+	rows, err := q.db.QueryContext(ctx, getActiveWithdraws)
 	if err != nil {
 		return nil, err
 	}
@@ -58,11 +95,12 @@ func (q *Queries) GetAllWithdraws(ctx context.Context) ([]*Withdraw, error) {
 		var i Withdraw
 		if err := rows.Scan(
 			&i.WithdrawID,
-			&i.UserID,
+			&i.CardNumber,
 			&i.WithdrawAmount,
 			&i.WithdrawTime,
 			&i.CreatedAt,
 			&i.UpdatedAt,
+			&i.DeletedAt,
 		); err != nil {
 			return nil, err
 		}
@@ -77,48 +115,41 @@ func (q *Queries) GetAllWithdraws(ctx context.Context) ([]*Withdraw, error) {
 	return items, nil
 }
 
-const getWithdrawById = `-- name: GetWithdrawById :one
-SELECT withdraw_id, user_id, withdraw_amount, withdraw_time, created_at, updated_at FROM withdraws WHERE withdraw_id = $1
+const getTrashedWithdrawByID = `-- name: GetTrashedWithdrawByID :one
+SELECT withdraw_id, card_number, withdraw_amount, withdraw_time, created_at, updated_at, deleted_at
+FROM withdraws
+WHERE
+    withdraw_id = $1
+    AND deleted_at IS NOT NULL
 `
 
-func (q *Queries) GetWithdrawById(ctx context.Context, withdrawID int32) (*Withdraw, error) {
-	row := q.db.QueryRowContext(ctx, getWithdrawById, withdrawID)
+// Get Trashed By Withdraw ID
+func (q *Queries) GetTrashedWithdrawByID(ctx context.Context, withdrawID int32) (*Withdraw, error) {
+	row := q.db.QueryRowContext(ctx, getTrashedWithdrawByID, withdrawID)
 	var i Withdraw
 	err := row.Scan(
 		&i.WithdrawID,
-		&i.UserID,
+		&i.CardNumber,
 		&i.WithdrawAmount,
 		&i.WithdrawTime,
 		&i.CreatedAt,
 		&i.UpdatedAt,
+		&i.DeletedAt,
 	)
 	return &i, err
 }
 
-const getWithdrawByUserId = `-- name: GetWithdrawByUserId :one
-SELECT withdraw_id, user_id, withdraw_amount, withdraw_time, created_at, updated_at FROM withdraws WHERE user_id = $1
+const getTrashedWithdraws = `-- name: GetTrashedWithdraws :many
+SELECT withdraw_id, card_number, withdraw_amount, withdraw_time, created_at, updated_at, deleted_at
+FROM withdraws
+WHERE
+    deleted_at IS NOT NULL
+ORDER BY withdraw_time DESC
 `
 
-func (q *Queries) GetWithdrawByUserId(ctx context.Context, userID int32) (*Withdraw, error) {
-	row := q.db.QueryRowContext(ctx, getWithdrawByUserId, userID)
-	var i Withdraw
-	err := row.Scan(
-		&i.WithdrawID,
-		&i.UserID,
-		&i.WithdrawAmount,
-		&i.WithdrawTime,
-		&i.CreatedAt,
-		&i.UpdatedAt,
-	)
-	return &i, err
-}
-
-const getWithdrawByUsers = `-- name: GetWithdrawByUsers :many
-SELECT withdraw_id, user_id, withdraw_amount, withdraw_time, created_at, updated_at FROM withdraws WHERE user_id = $1
-`
-
-func (q *Queries) GetWithdrawByUsers(ctx context.Context, userID int32) ([]*Withdraw, error) {
-	rows, err := q.db.QueryContext(ctx, getWithdrawByUsers, userID)
+// Get Trashed Withdraws
+func (q *Queries) GetTrashedWithdraws(ctx context.Context) ([]*Withdraw, error) {
+	rows, err := q.db.QueryContext(ctx, getTrashedWithdraws)
 	if err != nil {
 		return nil, err
 	}
@@ -128,11 +159,12 @@ func (q *Queries) GetWithdrawByUsers(ctx context.Context, userID int32) ([]*With
 		var i Withdraw
 		if err := rows.Scan(
 			&i.WithdrawID,
-			&i.UserID,
+			&i.CardNumber,
 			&i.WithdrawAmount,
 			&i.WithdrawTime,
 			&i.CreatedAt,
 			&i.UpdatedAt,
+			&i.DeletedAt,
 		); err != nil {
 			return nil, err
 		}
@@ -147,26 +179,174 @@ func (q *Queries) GetWithdrawByUsers(ctx context.Context, userID int32) ([]*With
 	return items, nil
 }
 
-const updateWithdraw = `-- name: UpdateWithdraw :one
-UPDATE withdraws SET withdraw_amount = $1, withdraw_time = $2 WHERE withdraw_id = $3 RETURNING withdraw_id, user_id, withdraw_amount, withdraw_time, created_at, updated_at
+const getWithdrawByID = `-- name: GetWithdrawByID :one
+SELECT withdraw_id, card_number, withdraw_amount, withdraw_time, created_at, updated_at, deleted_at
+FROM withdraws
+WHERE
+    withdraw_id = $1
+    AND deleted_at IS NULL
+`
+
+// Get Withdraw by ID
+func (q *Queries) GetWithdrawByID(ctx context.Context, withdrawID int32) (*Withdraw, error) {
+	row := q.db.QueryRowContext(ctx, getWithdrawByID, withdrawID)
+	var i Withdraw
+	err := row.Scan(
+		&i.WithdrawID,
+		&i.CardNumber,
+		&i.WithdrawAmount,
+		&i.WithdrawTime,
+		&i.CreatedAt,
+		&i.UpdatedAt,
+		&i.DeletedAt,
+	)
+	return &i, err
+}
+
+const restoreWithdraw = `-- name: RestoreWithdraw :exec
+UPDATE withdraws
+SET
+    deleted_at = NULL
+WHERE
+    withdraw_id = $1
+    AND deleted_at IS NOT NULL
+`
+
+// Restore Withdraw (Undelete)
+func (q *Queries) RestoreWithdraw(ctx context.Context, withdrawID int32) error {
+	_, err := q.db.ExecContext(ctx, restoreWithdraw, withdrawID)
+	return err
+}
+
+const searchWithdrawByCardNumber = `-- name: SearchWithdrawByCardNumber :many
+SELECT withdraw_id, card_number, withdraw_amount, withdraw_time, created_at, updated_at, deleted_at
+FROM withdraws
+WHERE
+    deleted_at IS NULL
+    AND card_number ILIKE '%' || $1 || '%'
+ORDER BY withdraw_time DESC
+`
+
+// Search Withdraw by Card Number
+func (q *Queries) SearchWithdrawByCardNumber(ctx context.Context, dollar_1 sql.NullString) ([]*Withdraw, error) {
+	rows, err := q.db.QueryContext(ctx, searchWithdrawByCardNumber, dollar_1)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var items []*Withdraw
+	for rows.Next() {
+		var i Withdraw
+		if err := rows.Scan(
+			&i.WithdrawID,
+			&i.CardNumber,
+			&i.WithdrawAmount,
+			&i.WithdrawTime,
+			&i.CreatedAt,
+			&i.UpdatedAt,
+			&i.DeletedAt,
+		); err != nil {
+			return nil, err
+		}
+		items = append(items, &i)
+	}
+	if err := rows.Close(); err != nil {
+		return nil, err
+	}
+	if err := rows.Err(); err != nil {
+		return nil, err
+	}
+	return items, nil
+}
+
+const searchWithdraws = `-- name: SearchWithdraws :many
+SELECT withdraw_id, card_number, withdraw_amount, withdraw_time, created_at, updated_at, deleted_at
+FROM withdraws
+WHERE deleted_at IS NULL
+  AND ($1::TEXT IS NULL OR card_number ILIKE '%' || $1 || '%')
+ORDER BY withdraw_time DESC
+LIMIT $2 OFFSET $3
+`
+
+type SearchWithdrawsParams struct {
+	Column1 string `json:"column_1"`
+	Limit   int32  `json:"limit"`
+	Offset  int32  `json:"offset"`
+}
+
+// Search Withdraws with Pagination
+func (q *Queries) SearchWithdraws(ctx context.Context, arg SearchWithdrawsParams) ([]*Withdraw, error) {
+	rows, err := q.db.QueryContext(ctx, searchWithdraws, arg.Column1, arg.Limit, arg.Offset)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var items []*Withdraw
+	for rows.Next() {
+		var i Withdraw
+		if err := rows.Scan(
+			&i.WithdrawID,
+			&i.CardNumber,
+			&i.WithdrawAmount,
+			&i.WithdrawTime,
+			&i.CreatedAt,
+			&i.UpdatedAt,
+			&i.DeletedAt,
+		); err != nil {
+			return nil, err
+		}
+		items = append(items, &i)
+	}
+	if err := rows.Close(); err != nil {
+		return nil, err
+	}
+	if err := rows.Err(); err != nil {
+		return nil, err
+	}
+	return items, nil
+}
+
+const trashWithdraw = `-- name: TrashWithdraw :exec
+UPDATE withdraws
+SET
+    deleted_at = current_timestamp
+WHERE
+    withdraw_id = $1
+    AND deleted_at IS NULL
+`
+
+// Trash Withdraw (Soft Delete)
+func (q *Queries) TrashWithdraw(ctx context.Context, withdrawID int32) error {
+	_, err := q.db.ExecContext(ctx, trashWithdraw, withdrawID)
+	return err
+}
+
+const updateWithdraw = `-- name: UpdateWithdraw :exec
+UPDATE withdraws
+SET
+    card_number = $2,
+    withdraw_amount = $3,
+    withdraw_time = $4,
+    updated_at = current_timestamp
+WHERE
+    withdraw_id = $1
+    AND deleted_at IS NULL
 `
 
 type UpdateWithdrawParams struct {
+	WithdrawID     int32     `json:"withdraw_id"`
+	CardNumber     string    `json:"card_number"`
 	WithdrawAmount int32     `json:"withdraw_amount"`
 	WithdrawTime   time.Time `json:"withdraw_time"`
-	WithdrawID     int32     `json:"withdraw_id"`
 }
 
-func (q *Queries) UpdateWithdraw(ctx context.Context, arg UpdateWithdrawParams) (*Withdraw, error) {
-	row := q.db.QueryRowContext(ctx, updateWithdraw, arg.WithdrawAmount, arg.WithdrawTime, arg.WithdrawID)
-	var i Withdraw
-	err := row.Scan(
-		&i.WithdrawID,
-		&i.UserID,
-		&i.WithdrawAmount,
-		&i.WithdrawTime,
-		&i.CreatedAt,
-		&i.UpdatedAt,
+// Update Withdraw
+func (q *Queries) UpdateWithdraw(ctx context.Context, arg UpdateWithdrawParams) error {
+	_, err := q.db.ExecContext(ctx, updateWithdraw,
+		arg.WithdrawID,
+		arg.CardNumber,
+		arg.WithdrawAmount,
+		arg.WithdrawTime,
 	)
-	return &i, err
+	return err
 }
